--- GRAPH-LOGISTICS.py
@@ -1,57 +1,0 @@
-<Graph>
-  Objects: {D1, R1, D3, C2, C3, C1, D2}
-Level 0:
-<Level>
-  Current State: {In(C1, R1), In(C2, D1), In(C3, D2), In(R1, D1), Holding(R1), Container(C1), Container(C2), Container(C3), Place(D1), Place(D2), Place(D3), Robot(R1)}
-  Actions: {PIn(C1, R1), PIn(C2, D1), PIn(C3, D2), PIn(R1, D1), PHolding(R1), PContainer(C1), PContainer(C2), PContainer(C3), PPlace(D1), PPlace(D2), PPlace(D3), PRobot(R1), PutDown(R1, C1, D1), Move(R1, D1, D3), Move(R1, D1, D2)}
-  Mutex: {{PutDown(R1, C1, D1), PHolding(R1)}, {PutDown(R1, C1, D1), PIn(C1, R1)}, {Move(R1, D1, D3), PIn(R1, D1)}, {Move(R1, D1, D2), PIn(R1, D1)}, {PutDown(R1, C1, D1), Move(R1, D1, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D2)}, {Move(R1, D1, D2), Move(R1, D1, D3)}}
-  Next state mutexes: {[{NotHolding(R1), In(C1, R1)}, {NotIn(C1, R1), In(C1, R1)}, {In(C1, D1), In(C1, R1)}, {NotIn(R1, D1), In(R1, D1)}, {In(R1, D1), In(R1, D3)}, {In(R1, D1), In(R1, D2)}, {NotHolding(R1), Holding(R1)}, {NotIn(C1, R1), Holding(R1)}, {In(C1, D1), Holding(R1)}, {NotIn(R1, D1), NotHolding(R1)}, {NotHolding(R1), In(R1, D3)}, {NotHolding(R1), In(R1, D2)}, {NotIn(C1, R1), NotIn(R1, D1)}, {NotIn(C1, R1), In(R1, D3)}, {NotIn(C1, R1), In(R1, D2)}, {NotIn(R1, D1), In(C1, D1)}, {In(C1, D1), In(R1, D3)}, {In(C1, D1), In(R1, D2)}, {In(R1, D3), In(R1, D2)}]}
-
-Level 1:
-<Level>
-  Current State: {Container(C3), Place(D1), Place(D3), In(C1, D1), In(R1, D2), NotIn(R1, D1), Holding(R1), In(C3, D2), In(C1, R1), Robot(R1), NotIn(C1, R1), NotHolding(R1), In(R1, D1), In(R1, D3), Place(D2), In(C2, D1), Container(C1), Container(C2)}
-  Actions: {PContainer(C3), PPlace(D1), PPlace(D3), PIn(C1, D1), PIn(R1, D2), PNotIn(R1, D1), PHolding(R1), PIn(C3, D2), PIn(C1, R1), PRobot(R1), PNotIn(C1, R1), PNotHolding(R1), PIn(R1, D1), PIn(R1, D3), PPlace(D2), PIn(C2, D1), PContainer(C1), PContainer(C2), PickUp(R1, C2, D1), PickUp(R1, C1, D1), PutDown(R1, C1, D1), PutDown(R1, C1, D3), PutDown(R1, C1, D2), Move(R1, D1, D3), Move(R1, D1, D2), Move(R1, D3, D1), Move(R1, D3, D2), Move(R1, D2, D1), Move(R1, D2, D3)}
-  Mutex: {{PNotIn(R1, D1), PIn(R1, D1)}, {Move(R1, D1, D3), PIn(R1, D1)}, {Move(R1, D1, D2), PIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D3, D1)}, {Move(R1, D1, D3), Move(R1, D3, D1)}, {Move(R1, D1, D2), Move(R1, D3, D1)}, {PNotIn(R1, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), Move(R1, D2, D1)}, {Move(R1, D1, D2), Move(R1, D2, D1)}, {PNotIn(C1, R1), PIn(C1, R1)}, {PutDown(R1, C1, D1), PIn(C1, R1)}, {PutDown(R1, C1, D3), PIn(C1, R1)}, {PutDown(R1, C1, D2), PIn(C1, R1)}, {PNotIn(C1, R1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D1)}, {PHolding(R1), PNotHolding(R1)}, {PutDown(R1, C1, D1), PHolding(R1)}, {PutDown(R1, C1, D3), PHolding(R1)}, {PutDown(R1, C1, D2), PHolding(R1)}, {PickUp(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C1, D1), PNotHolding(R1)}, {PIn(C2, D1), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), PIn(C3, D2)}, {PIn(C1, D1), PickUp(R1, C1, D1)}, {Move(R1, D3, D1), PIn(R1, D3)}, {Move(R1, D3, D2), PIn(R1, D3)}, {Move(R1, D1, D3), Move(R1, D3, D2)}, {Move(R1, D3, D1), Move(R1, D2, D3)}, {Move(R1, D3, D2), Move(R1, D2, D3)}, {Move(R1, D2, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D2, D3)}, {Move(R1, D2, D1), Move(R1, D3, D2)}, {PIn(C1, D1), PIn(R1, D2)}, {PIn(C1, D1), PNotIn(R1, D1)}, {PIn(C1, D1), PHolding(R1)}, {PIn(C1, D1), PIn(C1, R1)}, {PIn(C1, D1), PIn(R1, D3)}, {PIn(C1, D1), PickUp(R1, C3, D2)}, {PIn(C1, D1), PutDown(R1, C1, D1)}, {PIn(C1, D1), PutDown(R1, C1, D3)}, {PIn(C1, D1), PutDown(R1, C1, D2)}, {PIn(C1, D1), Move(R1, D3, D1)}, {PIn(C1, D1), Move(R1, D3, D2)}, {PIn(C1, D1), Move(R1, D2, D1)}, {PIn(C1, D1), Move(R1, D2, D3)}, {PNotIn(C1, R1), PIn(R1, D2)}, {PIn(R1, D2), PNotHolding(R1)}, {PIn(R1, D2), PIn(R1, D1)}, {PIn(R1, D2), PIn(R1, D3)}, {PickUp(R1, C2, D1), PIn(R1, D2)}, {PickUp(R1, C3, D2), PIn(R1, D2)}, {PickUp(R1, C1, D1), PIn(R1, D2)}, {PutDown(R1, C1, D1), PIn(R1, D2)}, {PutDown(R1, C1, D3), PIn(R1, D2)}, {Move(R1, D1, D3), PIn(R1, D2)}, {Move(R1, D1, D2), PIn(R1, D2)}, {Move(R1, D3, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D3, D2)}, {PNotIn(R1, D1), PNotIn(C1, R1)}, {PNotIn(R1, D1), PNotHolding(R1)}, {PNotIn(R1, D1), PickUp(R1, C2, D1)}, {PNotIn(R1, D1), PickUp(R1, C3, D2)}, {PNotIn(R1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PNotIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D1, D3)}, {PNotIn(R1, D1), Move(R1, D1, D2)}, {PHolding(R1), PNotIn(C1, R1)}, {PHolding(R1), PickUp(R1, C2, D1)}, {PHolding(R1), PickUp(R1, C3, D2)}, {PHolding(R1), PickUp(R1, C1, D1)}, {PIn(C1, R1), PNotHolding(R1)}, {PickUp(R1, C2, D1), PIn(C1, R1)}, {PickUp(R1, C3, D2), PIn(C1, R1)}, {PickUp(R1, C1, D1), PIn(C1, R1)}, {PNotIn(C1, R1), PIn(R1, D3)}, {PickUp(R1, C3, D2), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PNotIn(C1, R1)}, {PutDown(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D2), PNotIn(C1, R1)}, {PNotIn(C1, R1), Move(R1, D3, D1)}, {PNotIn(C1, R1), Move(R1, D3, D2)}, {PNotIn(C1, R1), Move(R1, D2, D1)}, {PNotIn(C1, R1), Move(R1, D2, D3)}, {PNotHolding(R1), PIn(R1, D3)}, {PutDown(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C1, D2), PNotHolding(R1)}, {Move(R1, D3, D1), PNotHolding(R1)}, {PNotHolding(R1), Move(R1, D3, D2)}, {Move(R1, D2, D1), PNotHolding(R1)}, {PNotHolding(R1), Move(R1, D2, D3)}, {PIn(R1, D3), PIn(R1, D1)}, {PickUp(R1, C3, D2), PIn(R1, D1)}, {PutDown(R1, C1, D3), PIn(R1, D1)}, {PutDown(R1, C1, D2), PIn(R1, D1)}, {Move(R1, D3, D1), PIn(R1, D1)}, {Move(R1, D3, D2), PIn(R1, D1)}, {Move(R1, D2, D1), PIn(R1, D1)}, {Move(R1, D2, D3), PIn(R1, D1)}, {PickUp(R1, C2, D1), PIn(R1, D3)}, {PickUp(R1, C3, D2), PIn(R1, D3)}, {PickUp(R1, C1, D1), PIn(R1, D3)}, {PutDown(R1, C1, D1), PIn(R1, D3)}, {PutDown(R1, C1, D2), PIn(R1, D3)}, {Move(R1, D1, D3), PIn(R1, D3)}, {Move(R1, D1, D2), PIn(R1, D3)}, {Move(R1, D2, D1), PIn(R1, D3)}, {Move(R1, D2, D3), PIn(R1, D3)}, {PickUp(R1, C3, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D2)}, {PickUp(R1, C2, D1), Move(R1, D2, D1)}, {PickUp(R1, C2, D1), Move(R1, D2, D3)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C3, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), Move(R1, D3, D1)}, {PickUp(R1, C3, D2), Move(R1, D3, D2)}, {PickUp(R1, C3, D2), Move(R1, D2, D1)}, {PickUp(R1, C3, D2), Move(R1, D2, D3)}, {PickUp(R1, C1, D1), Move(R1, D3, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D2)}, {PickUp(R1, C1, D1), Move(R1, D2, D1)}, {PickUp(R1, C1, D1), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C1, D1), Move(R1, D3, D1)}, {PutDown(R1, C1, D1), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), Move(R1, D2, D1)}, {PutDown(R1, C1, D1), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D2, D1)}, {PutDown(R1, C1, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D2)}, {PutDown(R1, C1, D2), Move(R1, D3, D1)}, {PutDown(R1, C1, D2), Move(R1, D3, D2)}, {Move(R1, D1, D3), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D3, D2)}, {Move(R1, D3, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), PickUp(R1, C2, D1)}, {Move(R1, D1, D2), PickUp(R1, C2, D1)}, {Move(R1, D1, D3), PickUp(R1, C1, D1)}, {Move(R1, D1, D2), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), Move(R1, D1, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D3, D1)}, {PutDown(R1, C1, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), Move(R1, D2, D1)}, {PutDown(R1, C1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D1, D3)}, {Move(R1, D3, D1), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D2, D3)}}
-  Next state mutexes: {[{In(C1, D1), In(C1, R1)}, {In(C1, D1), NotIn(C1, D1)}, {In(C1, D1), In(C1, D3)}, {In(C1, D1), In(C1, D2)}, {In(C1, D1), NotIn(R1, D3)}, {NotIn(R1, D2), In(C1, D1)}, {In(R1, D1), In(R1, D2)}, {In(R1, D3), In(R1, D2)}, {NotIn(C2, D1), In(R1, D2)}, {In(C2, R1), In(R1, D2)}, {NotIn(C1, D1), In(R1, D2)}, {In(C1, D3), In(R1, D2)}, {NotIn(R1, D2), In(R1, D2)}, {NotIn(R1, D1), In(R1, D1)}, {NotIn(R1, D1), NotIn(C2, D1)}, {NotIn(R1, D1), In(C2, R1)}, {NotIn(R1, D1), NotIn(C1, D1)}, {NotHolding(R1), Holding(R1)}, {In(C1, D3), Holding(R1)}, {Holding(R1), In(C1, D2)}, {NotIn(C1, R1), In(C1, R1)}, {NotHolding(R1), In(C1, R1)}, {In(C1, D3), In(C1, R1)}, {In(C1, D2), In(C1, R1)}, {NotIn(C1, R1), NotIn(C1, D1)}, {NotIn(C1, R1), NotIn(R1, D3)}, {NotIn(C1, R1), NotIn(R1, D2)}, {NotHolding(R1), NotIn(C2, D1)}, {NotHolding(R1), In(C2, R1)}, {NotHolding(R1), NotIn(C1, D1)}, {NotHolding(R1), NotIn(R1, D3)}, {NotIn(R1, D2), NotHolding(R1)}, {In(R1, D1), In(R1, D3)}, {In(C1, D3), In(R1, D1)}, {In(R1, D1), In(C1, D2)}, {NotIn(C2, D1), In(R1, D3)}, {In(R1, D3), In(C2, R1)}, {NotIn(C1, D1), In(R1, D3)}, {In(C1, D2), In(R1, D3)}, {NotIn(R1, D3), In(R1, D3)}, {In(C2, D1), NotIn(C2, D1)}, {In(C2, D1), In(C2, R1)}, {In(C1, D3), NotIn(C2, D1)}, {In(C1, D2), NotIn(C2, D1)}, {NotIn(R1, D3), NotIn(C2, D1)}, {NotIn(R1, D2), NotIn(C2, D1)}, {In(C1, D3), In(C2, R1)}, {In(C1, D2), In(C2, R1)}, {NotIn(R1, D3), In(C2, R1)}, {NotIn(R1, D2), In(C2, R1)}, {In(C1, D3), NotIn(C1, D1)}, {In(C1, D2), NotIn(C1, D1)}, {NotIn(R1, D3), NotIn(C1, D1)}, {NotIn(R1, D2), NotIn(C1, D1)}, {In(C1, D3), In(C1, D2)}, {NotIn(R1, D3), In(C1, D3)}, {NotIn(R1, D2), In(C1, D3)}, {NotIn(R1, D3), In(C1, D2)}, {NotIn(R1, D2), In(C1, D2)}, {NotIn(R1, D2), NotIn(R1, D3)}]}
-
-Level 2:
-<Level>
-  Current State: {Place(D3), NotIn(C2, D1), In(C2, R1), NotIn(C1, R1), In(R1, D1), Place(D2), Container(C3), Container(C1), In(C1, D2), In(R1, D2), Holding(R1), In(C3, D2), NotHolding(R1), In(R1, D3), NotIn(C1, D1), Container(C2), In(C1, D1), NotIn(R1, D3), NotIn(R1, D2), Place(D1), In(C1, D3), NotIn(R1, D1), In(C1, R1), Robot(R1), In(C2, D1)}
-  Actions: {PPlace(D3), PNotIn(C2, D1), PIn(C2, R1), PNotIn(C1, R1), PIn(R1, D1), PPlace(D2), PContainer(C3), PContainer(C1), PIn(C1, D2), PIn(R1, D2), PHolding(R1), PIn(C3, D2), PNotHolding(R1), PIn(R1, D3), PNotIn(C1, D1), PContainer(C2), PIn(C1, D1), PNotIn(R1, D3), PNotIn(R1, D2), PPlace(D1), PIn(C1, D3), PNotIn(R1, D1), PIn(C1, R1), PRobot(R1), PIn(C2, D1), PickUp(R1, C2, D1), PickUp(R1, C3, D2), PickUp(R1, C1, D1), PickUp(R1, C1, D3), PickUp(R1, C1, D2), PutDown(R1, C2, D1), PutDown(R1, C1, D1), PutDown(R1, C1, D3), PutDown(R1, C1, D2), Move(R1, D1, D3), Move(R1, D1, D2), Move(R1, D3, D1), Move(R1, D3, D2), Move(R1, D2, D1), Move(R1, D2, D3)}
-  Mutex: {{PIn(C2, D1), PNotIn(C2, D1)}, {PIn(C2, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D1), PNotIn(C2, D1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D1)}, {PNotIn(C1, R1), PIn(C1, R1)}, {PutDown(R1, C1, D1), PIn(C1, R1)}, {PutDown(R1, C1, D3), PIn(C1, R1)}, {PutDown(R1, C1, D2), PIn(C1, R1)}, {PNotIn(C1, R1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D2)}, {PHolding(R1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PHolding(R1)}, {PutDown(R1, C2, D3), PHolding(R1)}, {PHolding(R1), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D1), PHolding(R1)}, {PutDown(R1, C1, D3), PHolding(R1)}, {PutDown(R1, C1, D2), PHolding(R1)}, {PickUp(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), PutDown(R1, C2, D2)}, {PickUp(R1, C1, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), PutDown(R1, C2, D2)}, {PIn(C1, D1), PNotIn(C1, D1)}, {PIn(C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PNotIn(C1, D1)}, {PNotIn(R1, D3), PIn(R1, D3)}, {Move(R1, D3, D1), PIn(R1, D3)}, {Move(R1, D3, D2), PIn(R1, D3)}, {PNotIn(R1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D3), Move(R1, D3, D1)}, {Move(R1, D1, D3), Move(R1, D3, D2)}, {PNotIn(R1, D3), Move(R1, D2, D3)}, {Move(R1, D3, D1), Move(R1, D2, D3)}, {Move(R1, D3, D2), Move(R1, D2, D3)}, {PNotIn(R1, D2), PIn(R1, D2)}, {Move(R1, D2, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), PNotIn(R1, D2)}, {Move(R1, D1, D2), Move(R1, D2, D1)}, {Move(R1, D1, D2), Move(R1, D2, D3)}, {PNotIn(R1, D2), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D3, D2)}, {PNotIn(R1, D1), PIn(R1, D1)}, {Move(R1, D1, D3), PIn(R1, D1)}, {Move(R1, D1, D2), PIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D3, D1)}, {Move(R1, D1, D2), Move(R1, D3, D1)}, {PNotIn(R1, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), Move(R1, D2, D1)}, {PickUp(R1, C3, D2), PIn(C3, D2)}, {PickUp(R1, C1, D3), PIn(C1, D3)}, {PickUp(R1, C1, D2), PIn(C1, D2)}, {PutDown(R1, C2, D1), PIn(C2, R1)}, {PutDown(R1, C2, D3), PIn(C2, R1)}, {PutDown(R1, C2, D2), PIn(C2, R1)}, {PNotIn(C2, D1), PIn(C1, D2)}, {PIn(R1, D2), PNotIn(C2, D1)}, {PNotHolding(R1), PNotIn(C2, D1)}, {PNotIn(C2, D1), PIn(R1, D3)}, {PNotIn(R1, D3), PNotIn(C2, D1)}, {PNotIn(R1, D2), PNotIn(C2, D1)}, {PNotIn(C2, D1), PIn(C1, D3)}, {PNotIn(R1, D1), PNotIn(C2, D1)}, {PickUp(R1, C2, D1), PNotIn(C2, D1)}, {PickUp(R1, C3, D2), PNotIn(C2, D1)}, {PickUp(R1, C1, D1), PNotIn(C2, D1)}, {PickUp(R1, C1, D3), PNotIn(C2, D1)}, {PickUp(R1, C1, D2), PNotIn(C2, D1)}, {PutDown(R1, C2, D3), PNotIn(C2, D1)}, {PutDown(R1, C2, D2), PNotIn(C2, D1)}, {PutDown(R1, C1, D3), PNotIn(C2, D1)}, {PutDown(R1, C1, D2), PNotIn(C2, D1)}, {Move(R1, D3, D1), PNotIn(C2, D1)}, {Move(R1, D3, D2), PNotIn(C2, D1)}, {Move(R1, D2, D1), PNotIn(C2, D1)}, {Move(R1, D2, D3), PNotIn(C2, D1)}, {PIn(C2, R1), PIn(C1, D2)}, {PIn(C2, R1), PIn(R1, D2)}, {PIn(C2, R1), PNotHolding(R1)}, {PIn(C2, R1), PIn(R1, D3)}, {PNotIn(R1, D3), PIn(C2, R1)}, {PNotIn(R1, D2), PIn(C2, R1)}, {PIn(C2, R1), PIn(C1, D3)}, {PNotIn(R1, D1), PIn(C2, R1)}, {PIn(C2, D1), PIn(C2, R1)}, {PickUp(R1, C2, D1), PIn(C2, R1)}, {PickUp(R1, C3, D2), PIn(C2, R1)}, {PickUp(R1, C1, D1), PIn(C2, R1)}, {PickUp(R1, C1, D3), PIn(C2, R1)}, {PickUp(R1, C1, D2), PIn(C2, R1)}, {PutDown(R1, C1, D3), PIn(C2, R1)}, {PutDown(R1, C1, D2), PIn(C2, R1)}, {Move(R1, D3, D1), PIn(C2, R1)}, {PIn(C2, R1), Move(R1, D3, D2)}, {Move(R1, D2, D1), PIn(C2, R1)}, {PIn(C2, R1), Move(R1, D2, D3)}, {PNotIn(C1, R1), PNotIn(C1, D1)}, {PNotIn(R1, D3), PNotIn(C1, R1)}, {PNotIn(C1, R1), PNotIn(R1, D2)}, {PutDown(R1, C1, D1), PNotIn(C1, R1)}, {PutDown(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D2), PNotIn(C1, R1)}, {PIn(C1, D2), PIn(R1, D1)}, {PIn(R1, D2), PIn(R1, D1)}, {PIn(R1, D3), PIn(R1, D1)}, {PIn(C1, D3), PIn(R1, D1)}, {PickUp(R1, C3, D2), PIn(R1, D1)}, {PickUp(R1, C1, D3), PIn(R1, D1)}, {PickUp(R1, C1, D2), PIn(R1, D1)}, {PutDown(R1, C2, D3), PIn(R1, D1)}, {PutDown(R1, C2, D2), PIn(R1, D1)}, {PutDown(R1, C1, D3), PIn(R1, D1)}, {PutDown(R1, C1, D2), PIn(R1, D1)}, {Move(R1, D3, D1), PIn(R1, D1)}, {Move(R1, D3, D2), PIn(R1, D1)}, {Move(R1, D2, D1), PIn(R1, D1)}, {Move(R1, D2, D3), PIn(R1, D1)}, {PHolding(R1), PIn(C1, D2)}, {PIn(R1, D3), PIn(C1, D2)}, {PNotIn(C1, D1), PIn(C1, D2)}, {PIn(C1, D1), PIn(C1, D2)}, {PNotIn(R1, D3), PIn(C1, D2)}, {PNotIn(R1, D2), PIn(C1, D2)}, {PIn(C1, D3), PIn(C1, D2)}, {PIn(C1, R1), PIn(C1, D2)}, {PickUp(R1, C2, D1), PIn(C1, D2)}, {PickUp(R1, C1, D1), PIn(C1, D2)}, {PickUp(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C2, D1), PIn(C1, D2)}, {PutDown(R1, C2, D3), PIn(C1, D2)}, {PutDown(R1, C2, D2), PIn(C1, D2)}, {PutDown(R1, C1, D1), PIn(C1, D2)}, {PutDown(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C1, D2), PIn(C1, D2)}, {Move(R1, D1, D3), PIn(C1, D2)}, {Move(R1, D1, D2), PIn(C1, D2)}, {Move(R1, D3, D1), PIn(C1, D2)}, {Move(R1, D3, D2), PIn(C1, D2)}, {PIn(R1, D2), PIn(R1, D3)}, {PIn(R1, D2), PNotIn(C1, D1)}, {PIn(R1, D2), PIn(C1, D3)}, {PickUp(R1, C2, D1), PIn(R1, D2)}, {PickUp(R1, C1, D1), PIn(R1, D2)}, {PickUp(R1, C1, D3), PIn(R1, D2)}, {PutDown(R1, C2, D1), PIn(R1, D2)}, {PutDown(R1, C2, D3), PIn(R1, D2)}, {PutDown(R1, C2, D2), PIn(R1, D2)}, {PutDown(R1, C1, D1), PIn(R1, D2)}, {PutDown(R1, C1, D3), PIn(R1, D2)}, {Move(R1, D1, D3), PIn(R1, D2)}, {Move(R1, D1, D2), PIn(R1, D2)}, {Move(R1, D3, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D3, D2)}, {PHolding(R1), PIn(C1, D3)}, {PHolding(R1), PickUp(R1, C2, D1)}, {PHolding(R1), PickUp(R1, C3, D2)}, {PHolding(R1), PickUp(R1, C1, D1)}, {PHolding(R1), PickUp(R1, C1, D3)}, {PHolding(R1), PickUp(R1, C1, D2)}, {PNotHolding(R1), PNotIn(C1, D1)}, {PNotIn(R1, D3), PNotHolding(R1)}, {PNotIn(R1, D2), PNotHolding(R1)}, {PIn(C1, R1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D2), PNotHolding(R1)}, {PutDown(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C1, D2), PNotHolding(R1)}, {PNotIn(C1, D1), PIn(R1, D3)}, {PickUp(R1, C2, D1), PIn(R1, D3)}, {PickUp(R1, C3, D2), PIn(R1, D3)}, {PickUp(R1, C1, D1), PIn(R1, D3)}, {PickUp(R1, C1, D2), PIn(R1, D3)}, {PutDown(R1, C2, D1), PIn(R1, D3)}, {PutDown(R1, C2, D3), PIn(R1, D3)}, {PutDown(R1, C2, D2), PIn(R1, D3)}, {PutDown(R1, C1, D1), PIn(R1, D3)}, {PutDown(R1, C1, D2), PIn(R1, D3)}, {Move(R1, D1, D3), PIn(R1, D3)}, {Move(R1, D1, D2), PIn(R1, D3)}, {Move(R1, D2, D1), PIn(R1, D3)}, {Move(R1, D2, D3), PIn(R1, D3)}, {PNotIn(R1, D3), PNotIn(C1, D1)}, {PNotIn(R1, D2), PNotIn(C1, D1)}, {PNotIn(C1, D1), PIn(C1, D3)}, {PNotIn(R1, D1), PNotIn(C1, D1)}, {PickUp(R1, C2, D1), PNotIn(C1, D1)}, {PickUp(R1, C3, D2), PNotIn(C1, D1)}, {PickUp(R1, C1, D1), PNotIn(C1, D1)}, {PickUp(R1, C1, D3), PNotIn(C1, D1)}, {PickUp(R1, C1, D2), PNotIn(C1, D1)}, {PutDown(R1, C2, D3), PNotIn(C1, D1)}, {PutDown(R1, C2, D2), PNotIn(C1, D1)}, {PutDown(R1, C1, D3), PNotIn(C1, D1)}, {PutDown(R1, C1, D2), PNotIn(C1, D1)}, {Move(R1, D3, D1), PNotIn(C1, D1)}, {Move(R1, D3, D2), PNotIn(C1, D1)}, {Move(R1, D2, D1), PNotIn(C1, D1)}, {Move(R1, D2, D3), PNotIn(C1, D1)}, {PIn(C1, D1), PNotIn(R1, D3)}, {PIn(C1, D1), PNotIn(R1, D2)}, {PIn(C1, D1), PIn(C1, D3)}, {PIn(C1, D1), PIn(C1, R1)}, {PIn(C1, D1), PickUp(R1, C1, D3)}, {PIn(C1, D1), PickUp(R1, C1, D2)}, {PIn(C1, D1), PutDown(R1, C1, D1)}, {PIn(C1, D1), PutDown(R1, C1, D3)}, {PIn(C1, D1), PutDown(R1, C1, D2)}, {PNotIn(R1, D3), PNotIn(R1, D2)}, {PNotIn(R1, D3), PIn(C1, D3)}, {PNotIn(R1, D3), PickUp(R1, C2, D1)}, {PNotIn(R1, D3), PickUp(R1, C3, D2)}, {PNotIn(R1, D3), PickUp(R1, C1, D1)}, {PNotIn(R1, D3), PickUp(R1, C1, D3)}, {PNotIn(R1, D3), PickUp(R1, C1, D2)}, {PNotIn(R1, D3), PutDown(R1, C2, D1)}, {PNotIn(R1, D3), PutDown(R1, C2, D3)}, {PNotIn(R1, D3), PutDown(R1, C2, D2)}, {PNotIn(R1, D3), PutDown(R1, C1, D3)}, {PNotIn(R1, D3), Move(R1, D3, D1)}, {PNotIn(R1, D3), Move(R1, D3, D2)}, {PNotIn(R1, D2), PIn(C1, D3)}, {PNotIn(R1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), PNotIn(R1, D2)}, {PickUp(R1, C1, D1), PNotIn(R1, D2)}, {PickUp(R1, C1, D3), PNotIn(R1, D2)}, {PickUp(R1, C1, D2), PNotIn(R1, D2)}, {PutDown(R1, C2, D1), PNotIn(R1, D2)}, {PutDown(R1, C2, D3), PNotIn(R1, D2)}, {PutDown(R1, C2, D2), PNotIn(R1, D2)}, {PutDown(R1, C1, D2), PNotIn(R1, D2)}, {PNotIn(R1, D2), Move(R1, D2, D1)}, {PNotIn(R1, D2), Move(R1, D2, D3)}, {PIn(C1, R1), PIn(C1, D3)}, {PickUp(R1, C2, D1), PIn(C1, D3)}, {PickUp(R1, C3, D2), PIn(C1, D3)}, {PickUp(R1, C1, D1), PIn(C1, D3)}, {PickUp(R1, C1, D2), PIn(C1, D3)}, {PutDown(R1, C2, D1), PIn(C1, D3)}, {PutDown(R1, C2, D3), PIn(C1, D3)}, {PutDown(R1, C2, D2), PIn(C1, D3)}, {PutDown(R1, C1, D1), PIn(C1, D3)}, {PutDown(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D2), PIn(C1, D3)}, {Move(R1, D1, D3), PIn(C1, D3)}, {Move(R1, D1, D2), PIn(C1, D3)}, {Move(R1, D2, D1), PIn(C1, D3)}, {Move(R1, D2, D3), PIn(C1, D3)}, {PNotIn(R1, D1), PickUp(R1, C2, D1)}, {PNotIn(R1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D1), PNotIn(R1, D1)}, {PutDown(R1, C2, D3), PNotIn(R1, D1)}, {PNotIn(R1, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D1), PNotIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D1, D3)}, {PNotIn(R1, D1), Move(R1, D1, D2)}, {PickUp(R1, C2, D1), PIn(C1, R1)}, {PickUp(R1, C3, D2), PIn(C1, R1)}, {PickUp(R1, C1, D1), PIn(C1, R1)}, {PickUp(R1, C1, D3), PIn(C1, R1)}, {PickUp(R1, C1, D2), PIn(C1, R1)}, {PIn(C2, D1), PutDown(R1, C2, D1)}, {PIn(C2, D1), PutDown(R1, C2, D3)}, {PIn(C2, D1), PutDown(R1, C2, D2)}, {PickUp(R1, C3, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D2)}, {PickUp(R1, C2, D1), Move(R1, D2, D1)}, {PickUp(R1, C2, D1), Move(R1, D2, D3)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C3, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), Move(R1, D3, D1)}, {PickUp(R1, C3, D2), Move(R1, D3, D2)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D2)}, {PickUp(R1, C1, D1), Move(R1, D2, D1)}, {PickUp(R1, C1, D1), Move(R1, D2, D3)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), Move(R1, D2, D1)}, {PickUp(R1, C1, D3), Move(R1, D2, D3)}, {PickUp(R1, C1, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), Move(R1, D3, D1)}, {PickUp(R1, C1, D2), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C2, D1), Move(R1, D3, D1)}, {PutDown(R1, C2, D1), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), Move(R1, D2, D1)}, {PutDown(R1, C2, D1), Move(R1, D2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D2)}, {PutDown(R1, C2, D3), Move(R1, D3, D1)}, {PutDown(R1, C2, D3), Move(R1, D3, D2)}, {PutDown(R1, C2, D3), Move(R1, D2, D1)}, {PutDown(R1, C2, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D3, D1)}, {PutDown(R1, C2, D2), Move(R1, D3, D2)}, {PutDown(R1, C2, D2), Move(R1, D2, D1)}, {PutDown(R1, C2, D2), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C1, D1), Move(R1, D3, D1)}, {PutDown(R1, C1, D1), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), Move(R1, D2, D1)}, {PutDown(R1, C1, D1), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D2, D1)}, {PutDown(R1, C1, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D2)}, {PutDown(R1, C1, D2), Move(R1, D3, D1)}, {PutDown(R1, C1, D2), Move(R1, D3, D2)}, {Move(R1, D1, D3), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D3, D2)}, {Move(R1, D3, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), PickUp(R1, C2, D1)}, {Move(R1, D1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), Move(R1, D2, D1)}, {PickUp(R1, C3, D2), Move(R1, D2, D3)}, {Move(R1, D1, D3), PickUp(R1, C1, D1)}, {Move(R1, D1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D2)}, {PickUp(R1, C1, D2), Move(R1, D2, D1)}, {PickUp(R1, C1, D2), Move(R1, D2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C2, D1), Move(R1, D1, D3)}, {PutDown(R1, C2, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D1), Move(R1, D1, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D3, D1)}, {PutDown(R1, C1, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), Move(R1, D2, D1)}, {PutDown(R1, C1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D1, D3)}, {Move(R1, D3, D1), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D2, D3)}}
-  Next state mutexes: {[{In(C1, D2), NotIn(C2, D1)}, {NotIn(R1, D3), NotIn(C2, D1)}, {NotIn(R1, D2), NotIn(C2, D1)}, {In(C1, D3), NotIn(C2, D1)}, {In(C2, D1), NotIn(C2, D1)}, {NotIn(C3, D2), NotIn(C2, D1)}, {In(C3, R1), NotIn(C2, D1)}, {NotIn(C1, D3), NotIn(C2, D1)}, {NotIn(C1, D2), NotIn(C2, D1)}, {NotIn(C2, R1), NotIn(C2, D1)}, {In(C1, D2), In(C2, R1)}, {NotIn(R1, D3), In(C2, R1)}, {NotIn(R1, D2), In(C2, R1)}, {In(C1, D3), In(C2, R1)}, {In(C2, D1), In(C2, R1)}, {NotIn(C3, D2), In(C2, R1)}, {In(C3, R1), In(C2, R1)}, {NotIn(C1, D3), In(C2, R1)}, {NotIn(C1, D2), In(C2, R1)}, {NotIn(C2, R1), In(C2, R1)}, {NotIn(C1, R1), NotIn(C1, D1)}, {NotIn(C1, R1), In(C1, R1)}, {NotIn(C1, R1), NotIn(C1, D3)}, {NotIn(C1, R1), NotIn(C1, D2)}, {In(R1, D1), In(R1, D2)}, {In(R1, D1), In(R1, D3)}, {NotIn(R1, D1), In(R1, D1)}, {NotIn(C3, D2), In(R1, D1)}, {In(R1, D1), In(C3, R1)}, {NotIn(C1, D3), In(R1, D1)}, {NotIn(C1, D2), In(R1, D1)}, {In(C1, D2), NotIn(C1, D1)}, {In(C1, D1), In(C1, D2)}, {In(C1, D3), In(C1, D2)}, {In(C1, D2), In(C1, R1)}, {NotIn(C1, D3), In(C1, D2)}, {NotIn(C1, D2), In(C1, D2)}, {NotIn(C2, R1), In(C1, D2)}, {In(R1, D3), In(R1, D2)}, {NotIn(R1, D2), In(R1, D2)}, {NotIn(C1, D3), In(R1, D2)}, {NotIn(C2, R1), In(R1, D2)}, {NotHolding(R1), Holding(R1)}, {In(C1, D3), Holding(R1)}, {NotIn(C2, R1), Holding(R1)}, {NotIn(C3, D2), In(C3, D2)}, {In(C3, D2), In(C3, R1)}, {NotIn(C3, D2), NotHolding(R1)}, {NotHolding(R1), In(C3, R1)}, {NotIn(C1, D3), NotHolding(R1)}, {NotIn(C1, D2), NotHolding(R1)}, {NotIn(R1, D3), In(R1, D3)}, {NotIn(C3, D2), In(R1, D3)}, {In(C3, R1), In(R1, D3)}, {NotIn(C1, D2), In(R1, D3)}, {NotIn(C2, R1), In(R1, D3)}, {In(C1, D1), NotIn(C1, D1)}, {NotIn(R1, D3), NotIn(C1, D1)}, {NotIn(R1, D2), NotIn(C1, D1)}, {In(C1, D3), NotIn(C1, D1)}, {NotIn(C3, D2), NotIn(C1, D1)}, {In(C3, R1), NotIn(C1, D1)}, {NotIn(C1, D3), NotIn(C1, D1)}, {NotIn(C1, D2), NotIn(C1, D1)}, {In(C1, D1), In(C1, D3)}, {In(C1, D1), In(C1, R1)}, {NotIn(C1, D3), In(C1, D1)}, {NotIn(C1, D2), In(C1, D1)}, {NotIn(C3, D2), NotIn(R1, D3)}, {NotIn(R1, D3), In(C3, R1)}, {NotIn(C1, D3), NotIn(R1, D3)}, {NotIn(C1, D2), NotIn(R1, D3)}, {NotIn(C2, R1), NotIn(R1, D3)}, {NotIn(R1, D2), NotIn(C3, D2)}, {NotIn(R1, D2), In(C3, R1)}, {NotIn(R1, D2), NotIn(C1, D3)}, {NotIn(R1, D2), NotIn(C1, D2)}, {NotIn(R1, D2), NotIn(C2, R1)}, {In(C1, D3), In(C1, R1)}, {NotIn(C3, D2), In(C1, D3)}, {In(C1, D3), In(C3, R1)}, {NotIn(C1, D3), In(C1, D3)}, {NotIn(C1, D2), In(C1, D3)}, {NotIn(C2, R1), In(C1, D3)}, {NotIn(C2, R1), NotIn(R1, D1)}, {NotIn(C3, D2), NotIn(C1, D3)}, {NotIn(C2, R1), NotIn(C3, D2)}, {NotIn(C1, D3), In(C3, R1)}, {NotIn(C2, R1), In(C3, R1)}, {NotIn(C1, D3), NotIn(C1, D2)}, {NotIn(C2, R1), NotIn(C1, D3)}, {NotIn(C2, R1), NotIn(C1, D2)}]}
-
-Level 3:
-<Level>
-  Current State: {Place(D3), NotIn(C2, D1), In(C2, R1), NotIn(C1, R1), In(R1, D1), Place(D2), NotIn(C1, D3), Container(C3), Container(C1), In(C1, D2), In(C3, R1), In(R1, D2), Holding(R1), In(C3, D2), NotHolding(R1), In(R1, D3), NotIn(C1, D1), Container(C2), In(C1, D1), NotIn(R1, D3), NotIn(C3, D2), NotIn(R1, D2), Place(D1), NotIn(C1, D2), In(C1, D3), NotIn(R1, D1), In(C1, R1), Robot(R1), NotIn(C2, R1), In(C2, D1)}
-  Actions: {PPlace(D3), PNotIn(C2, D1), PIn(C2, R1), PNotIn(C1, R1), PIn(R1, D1), PPlace(D2), PNotIn(C1, D3), PContainer(C3), PContainer(C1), PIn(C1, D2), PIn(C3, R1), PIn(R1, D2), PHolding(R1), PIn(C3, D2), PNotHolding(R1), PIn(R1, D3), PNotIn(C1, D1), PContainer(C2), PIn(C1, D1), PNotIn(R1, D3), PNotIn(C3, D2), PNotIn(R1, D2), PPlace(D1), PNotIn(C1, D2), PIn(C1, D3), PNotIn(R1, D1), PIn(C1, R1), PRobot(R1), PNotIn(C2, R1), PIn(C2, D1), PickUp(R1, C2, D1), PickUp(R1, C3, D2), PickUp(R1, C1, D1), PickUp(R1, C1, D3), PickUp(R1, C1, D2), PutDown(R1, C2, D1), PutDown(R1, C2, D3), PutDown(R1, C2, D2), PutDown(R1, C3, D2), PutDown(R1, C1, D1), PutDown(R1, C1, D3), PutDown(R1, C1, D2), Move(R1, D1, D3), Move(R1, D1, D2), Move(R1, D3, D1), Move(R1, D3, D2), Move(R1, D2, D1), Move(R1, D2, D3)}
-  Mutex: {{PIn(C2, D1), PNotIn(C2, D1)}, {PIn(C2, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D1), PNotIn(C2, D1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D1)}, {PNotIn(C1, R1), PIn(C1, R1)}, {PutDown(R1, C1, D1), PIn(C1, R1)}, {PutDown(R1, C1, D3), PIn(C1, R1)}, {PutDown(R1, C1, D2), PIn(C1, R1)}, {PNotIn(C1, R1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D2)}, {PNotIn(C1, D3), PIn(C1, D3)}, {PickUp(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D3), PNotIn(C1, D3)}, {PHolding(R1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PHolding(R1)}, {PutDown(R1, C2, D3), PHolding(R1)}, {PHolding(R1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PHolding(R1)}, {PutDown(R1, C3, D3), PHolding(R1)}, {PutDown(R1, C3, D2), PHolding(R1)}, {PutDown(R1, C1, D1), PHolding(R1)}, {PutDown(R1, C1, D3), PHolding(R1)}, {PutDown(R1, C1, D2), PHolding(R1)}, {PickUp(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D2), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D2)}, {PIn(C1, D1), PNotIn(C1, D1)}, {PIn(C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PNotIn(C1, D1)}, {PNotIn(R1, D3), PIn(R1, D3)}, {Move(R1, D3, D1), PIn(R1, D3)}, {Move(R1, D3, D2), PIn(R1, D3)}, {PNotIn(R1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D3), Move(R1, D3, D1)}, {Move(R1, D1, D3), Move(R1, D3, D2)}, {PNotIn(R1, D3), Move(R1, D2, D3)}, {Move(R1, D3, D1), Move(R1, D2, D3)}, {Move(R1, D3, D2), Move(R1, D2, D3)}, {PNotIn(C3, D2), PIn(C3, D2)}, {PickUp(R1, C3, D2), PIn(C3, D2)}, {PutDown(R1, C3, D2), PNotIn(C3, D2)}, {PNotIn(R1, D2), PIn(R1, D2)}, {Move(R1, D2, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), PNotIn(R1, D2)}, {Move(R1, D1, D2), Move(R1, D2, D1)}, {Move(R1, D1, D2), Move(R1, D2, D3)}, {PNotIn(R1, D2), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D3, D2)}, {PNotIn(C1, D2), PIn(C1, D2)}, {PickUp(R1, C1, D2), PIn(C1, D2)}, {PNotIn(C1, D2), PutDown(R1, C1, D2)}, {PNotIn(R1, D1), PIn(R1, D1)}, {Move(R1, D1, D3), PIn(R1, D1)}, {Move(R1, D1, D2), PIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D3, D1)}, {Move(R1, D1, D2), Move(R1, D3, D1)}, {PNotIn(R1, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), Move(R1, D2, D1)}, {PNotIn(C2, R1), PIn(C2, R1)}, {PutDown(R1, C2, D1), PIn(C2, R1)}, {PutDown(R1, C2, D3), PIn(C2, R1)}, {PutDown(R1, C2, D2), PIn(C2, R1)}, {PickUp(R1, C2, D1), PNotIn(C2, R1)}, {PutDown(R1, C3, D1), PIn(C3, R1)}, {PutDown(R1, C3, D3), PIn(C3, R1)}, {PutDown(R1, C3, D2), PIn(C3, R1)}, {PNotIn(C1, D3), PNotIn(C2, D1)}, {PNotIn(C2, D1), PIn(C1, D2)}, {PNotIn(C2, D1), PIn(C3, R1)}, {PNotIn(R1, D3), PNotIn(C2, D1)}, {PNotIn(C3, D2), PNotIn(C2, D1)}, {PNotIn(R1, D2), PNotIn(C2, D1)}, {PNotIn(C1, D2), PNotIn(C2, D1)}, {PNotIn(C2, D1), PIn(C1, D3)}, {PNotIn(C2, R1), PNotIn(C2, D1)}, {PickUp(R1, C2, D1), PNotIn(C2, D1)}, {PickUp(R1, C1, D3), PNotIn(C2, D1)}, {PickUp(R1, C1, D2), PNotIn(C2, D1)}, {PutDown(R1, C3, D1), PNotIn(C2, D1)}, {PutDown(R1, C3, D3), PNotIn(C2, D1)}, {PutDown(R1, C3, D2), PNotIn(C2, D1)}, {PNotIn(C1, D3), PIn(C2, R1)}, {PIn(C2, R1), PIn(C1, D2)}, {PIn(C2, R1), PIn(C3, R1)}, {PNotIn(R1, D3), PIn(C2, R1)}, {PNotIn(C3, D2), PIn(C2, R1)}, {PNotIn(R1, D2), PIn(C2, R1)}, {PNotIn(C1, D2), PIn(C2, R1)}, {PIn(C2, R1), PIn(C1, D3)}, {PIn(C2, D1), PIn(C2, R1)}, {PickUp(R1, C2, D1), PIn(C2, R1)}, {PickUp(R1, C1, D3), PIn(C2, R1)}, {PickUp(R1, C1, D2), PIn(C2, R1)}, {PutDown(R1, C3, D1), PIn(C2, R1)}, {PutDown(R1, C3, D3), PIn(C2, R1)}, {PutDown(R1, C3, D2), PIn(C2, R1)}, {PNotIn(C1, D3), PNotIn(C1, R1)}, {PNotIn(C1, R1), PNotIn(C1, D1)}, {PNotIn(C1, D2), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PNotIn(C1, R1)}, {PutDown(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D2), PNotIn(C1, R1)}, {PNotIn(C1, D3), PIn(R1, D1)}, {PIn(C3, R1), PIn(R1, D1)}, {PIn(R1, D2), PIn(R1, D1)}, {PIn(R1, D3), PIn(R1, D1)}, {PNotIn(C3, D2), PIn(R1, D1)}, {PNotIn(C1, D2), PIn(R1, D1)}, {PickUp(R1, C3, D2), PIn(R1, D1)}, {PickUp(R1, C1, D3), PIn(R1, D1)}, {PickUp(R1, C1, D2), PIn(R1, D1)}, {PutDown(R1, C2, D3), PIn(R1, D1)}, {PutDown(R1, C2, D2), PIn(R1, D1)}, {PutDown(R1, C3, D1), PIn(R1, D1)}, {PutDown(R1, C3, D3), PIn(R1, D1)}, {PutDown(R1, C3, D2), PIn(R1, D1)}, {PutDown(R1, C1, D3), PIn(R1, D1)}, {PutDown(R1, C1, D2), PIn(R1, D1)}, {Move(R1, D3, D1), PIn(R1, D1)}, {Move(R1, D3, D2), PIn(R1, D1)}, {Move(R1, D2, D1), PIn(R1, D1)}, {Move(R1, D2, D3), PIn(R1, D1)}, {PNotIn(C1, D3), PIn(C1, D2)}, {PNotIn(C1, D3), PIn(C3, R1)}, {PNotIn(C1, D3), PIn(R1, D2)}, {PNotIn(C1, D3), PNotHolding(R1)}, {PNotIn(C1, D3), PNotIn(C1, D1)}, {PIn(C1, D1), PNotIn(C1, D3)}, {PNotIn(R1, D3), PNotIn(C1, D3)}, {PNotIn(C1, D3), PNotIn(C3, D2)}, {PNotIn(C1, D3), PNotIn(R1, D2)}, {PNotIn(C1, D2), PNotIn(C1, D3)}, {PNotIn(C1, D3), PNotIn(C2, R1)}, {PNotIn(C1, D3), PickUp(R1, C2, D1)}, {PNotIn(C1, D3), PickUp(R1, C3, D2)}, {PNotIn(C1, D3), PickUp(R1, C1, D1)}, {PNotIn(C1, D3), PickUp(R1, C1, D3)}, {PNotIn(C1, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D1), PNotIn(C1, D3)}, {PutDown(R1, C2, D3), PNotIn(C1, D3)}, {PNotIn(C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PNotIn(C1, D3)}, {PNotIn(C1, D3), PutDown(R1, C3, D3)}, {PNotIn(C1, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PNotIn(C1, D3)}, {PutDown(R1, C1, D2), PNotIn(C1, D3)}, {PNotIn(C1, D3), Move(R1, D1, D3)}, {PNotIn(C1, D3), Move(R1, D1, D2)}, {PNotIn(C1, D3), Move(R1, D2, D1)}, {PNotIn(C1, D3), Move(R1, D2, D3)}, {PNotIn(C1, D1), PIn(C1, D2)}, {PIn(C1, D1), PIn(C1, D2)}, {PIn(C1, D3), PIn(C1, D2)}, {PIn(C1, R1), PIn(C1, D2)}, {PNotIn(C2, R1), PIn(C1, D2)}, {PickUp(R1, C1, D1), PIn(C1, D2)}, {PickUp(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C2, D1), PIn(C1, D2)}, {PutDown(R1, C2, D3), PIn(C1, D2)}, {PutDown(R1, C2, D2), PIn(C1, D2)}, {PutDown(R1, C1, D1), PIn(C1, D2)}, {PutDown(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C1, D2), PIn(C1, D2)}, {PIn(C3, R1), PIn(C3, D2)}, {PNotHolding(R1), PIn(C3, R1)}, {PIn(C3, R1), PIn(R1, D3)}, {PNotIn(C1, D1), PIn(C3, R1)}, {PNotIn(R1, D3), PIn(C3, R1)}, {PNotIn(R1, D2), PIn(C3, R1)}, {PIn(C3, R1), PIn(C1, D3)}, {PNotIn(C2, R1), PIn(C3, R1)}, {PickUp(R1, C2, D1), PIn(C3, R1)}, {PickUp(R1, C3, D2), PIn(C3, R1)}, {PickUp(R1, C1, D1), PIn(C3, R1)}, {PickUp(R1, C1, D3), PIn(C3, R1)}, {PickUp(R1, C1, D2), PIn(C3, R1)}, {PutDown(R1, C2, D1), PIn(C3, R1)}, {PutDown(R1, C2, D3), PIn(C3, R1)}, {PutDown(R1, C2, D2), PIn(C3, R1)}, {PutDown(R1, C1, D1), PIn(C3, R1)}, {PutDown(R1, C1, D3), PIn(C3, R1)}, {Move(R1, D1, D3), PIn(C3, R1)}, {Move(R1, D1, D2), PIn(C3, R1)}, {Move(R1, D3, D1), PIn(C3, R1)}, {Move(R1, D3, D2), PIn(C3, R1)}, {PIn(R1, D2), PIn(R1, D3)}, {PNotIn(C2, R1), PIn(R1, D2)}, {PickUp(R1, C2, D1), PIn(R1, D2)}, {PickUp(R1, C1, D1), PIn(R1, D2)}, {PickUp(R1, C1, D3), PIn(R1, D2)}, {PutDown(R1, C2, D1), PIn(R1, D2)}, {PutDown(R1, C2, D3), PIn(R1, D2)}, {PutDown(R1, C3, D1), PIn(R1, D2)}, {PutDown(R1, C3, D3), PIn(R1, D2)}, {PutDown(R1, C1, D1), PIn(R1, D2)}, {PutDown(R1, C1, D3), PIn(R1, D2)}, {Move(R1, D1, D3), PIn(R1, D2)}, {Move(R1, D1, D2), PIn(R1, D2)}, {Move(R1, D3, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D3, D2)}, {PHolding(R1), PIn(C1, D3)}, {PHolding(R1), PNotIn(C2, R1)}, {PHolding(R1), PickUp(R1, C2, D1)}, {PHolding(R1), PickUp(R1, C3, D2)}, {PHolding(R1), PickUp(R1, C1, D1)}, {PHolding(R1), PickUp(R1, C1, D3)}, {PHolding(R1), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D1), PIn(C3, D2)}, {PutDown(R1, C3, D3), PIn(C3, D2)}, {PutDown(R1, C3, D2), PIn(C3, D2)}, {PNotIn(C3, D2), PNotHolding(R1)}, {PNotIn(C1, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D2), PNotHolding(R1)}, {PutDown(R1, C3, D1), PNotHolding(R1)}, {PutDown(R1, C3, D3), PNotHolding(R1)}, {PutDown(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C1, D2), PNotHolding(R1)}, {PNotIn(C3, D2), PIn(R1, D3)}, {PNotIn(C1, D2), PIn(R1, D3)}, {PNotIn(C2, R1), PIn(R1, D3)}, {PickUp(R1, C2, D1), PIn(R1, D3)}, {PickUp(R1, C3, D2), PIn(R1, D3)}, {PickUp(R1, C1, D1), PIn(R1, D3)}, {PickUp(R1, C1, D2), PIn(R1, D3)}, {PutDown(R1, C2, D1), PIn(R1, D3)}, {PutDown(R1, C2, D2), PIn(R1, D3)}, {PutDown(R1, C3, D1), PIn(R1, D3)}, {PutDown(R1, C3, D3), PIn(R1, D3)}, {PutDown(R1, C3, D2), PIn(R1, D3)}, {PutDown(R1, C1, D1), PIn(R1, D3)}, {PutDown(R1, C1, D2), PIn(R1, D3)}, {Move(R1, D1, D3), PIn(R1, D3)}, {Move(R1, D1, D2), PIn(R1, D3)}, {Move(R1, D2, D1), PIn(R1, D3)}, {Move(R1, D2, D3), PIn(R1, D3)}, {PNotIn(R1, D3), PNotIn(C1, D1)}, {PNotIn(C3, D2), PNotIn(C1, D1)}, {PNotIn(R1, D2), PNotIn(C1, D1)}, {PNotIn(C1, D2), PNotIn(C1, D1)}, {PNotIn(C1, D1), PIn(C1, D3)}, {PickUp(R1, C1, D1), PNotIn(C1, D1)}, {PickUp(R1, C1, D3), PNotIn(C1, D1)}, {PickUp(R1, C1, D2), PNotIn(C1, D1)}, {PutDown(R1, C3, D1), PNotIn(C1, D1)}, {PutDown(R1, C3, D3), PNotIn(C1, D1)}, {PutDown(R1, C3, D2), PNotIn(C1, D1)}, {PIn(C1, D1), PNotIn(C1, D2)}, {PIn(C1, D1), PIn(C1, D3)}, {PIn(C1, D1), PIn(C1, R1)}, {PIn(C1, D1), PickUp(R1, C1, D3)}, {PIn(C1, D1), PickUp(R1, C1, D2)}, {PIn(C1, D1), PutDown(R1, C1, D1)}, {PIn(C1, D1), PutDown(R1, C1, D3)}, {PIn(C1, D1), PutDown(R1, C1, D2)}, {PNotIn(R1, D3), PNotIn(C3, D2)}, {PNotIn(R1, D3), PNotIn(C1, D2)}, {PNotIn(R1, D3), PNotIn(C2, R1)}, {PNotIn(R1, D3), PickUp(R1, C1, D3)}, {PNotIn(R1, D3), PutDown(R1, C2, D1)}, {PNotIn(R1, D3), PutDown(R1, C2, D3)}, {PNotIn(R1, D3), PutDown(R1, C2, D2)}, {PNotIn(R1, D3), PutDown(R1, C3, D1)}, {PNotIn(R1, D3), PutDown(R1, C3, D3)}, {PNotIn(R1, D3), PutDown(R1, C3, D2)}, {PNotIn(R1, D3), PutDown(R1, C1, D3)}, {PNotIn(R1, D3), Move(R1, D3, D1)}, {PNotIn(R1, D3), Move(R1, D3, D2)}, {PNotIn(C3, D2), PNotIn(R1, D2)}, {PNotIn(C3, D2), PIn(C1, D3)}, {PNotIn(C3, D2), PNotIn(C2, R1)}, {PNotIn(C3, D2), PickUp(R1, C2, D1)}, {PNotIn(C3, D2), PickUp(R1, C3, D2)}, {PNotIn(C3, D2), PickUp(R1, C1, D1)}, {PNotIn(C3, D2), PickUp(R1, C1, D3)}, {PNotIn(C3, D2), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D1), PNotIn(C3, D2)}, {PutDown(R1, C2, D3), PNotIn(C3, D2)}, {PNotIn(C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PNotIn(C3, D2)}, {PNotIn(C3, D2), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D1), PNotIn(C3, D2)}, {PutDown(R1, C1, D3), PNotIn(C3, D2)}, {PNotIn(C3, D2), Move(R1, D1, D3)}, {PNotIn(C3, D2), Move(R1, D1, D2)}, {PNotIn(C3, D2), Move(R1, D3, D1)}, {PNotIn(C3, D2), Move(R1, D3, D2)}, {PNotIn(C1, D2), PNotIn(R1, D2)}, {PNotIn(R1, D2), PNotIn(C2, R1)}, {PickUp(R1, C3, D2), PNotIn(R1, D2)}, {PickUp(R1, C1, D2), PNotIn(R1, D2)}, {PutDown(R1, C2, D1), PNotIn(R1, D2)}, {PutDown(R1, C2, D3), PNotIn(R1, D2)}, {PutDown(R1, C2, D2), PNotIn(R1, D2)}, {PutDown(R1, C3, D1), PNotIn(R1, D2)}, {PutDown(R1, C3, D3), PNotIn(R1, D2)}, {PutDown(R1, C3, D2), PNotIn(R1, D2)}, {PutDown(R1, C1, D2), PNotIn(R1, D2)}, {PNotIn(R1, D2), Move(R1, D2, D1)}, {PNotIn(R1, D2), Move(R1, D2, D3)}, {PNotIn(C1, D2), PIn(C1, D3)}, {PNotIn(C1, D2), PNotIn(C2, R1)}, {PNotIn(C1, D2), PickUp(R1, C2, D1)}, {PNotIn(C1, D2), PickUp(R1, C3, D2)}, {PNotIn(C1, D2), PickUp(R1, C1, D1)}, {PNotIn(C1, D2), PickUp(R1, C1, D3)}, {PNotIn(C1, D2), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D1), PNotIn(C1, D2)}, {PutDown(R1, C2, D3), PNotIn(C1, D2)}, {PNotIn(C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PNotIn(C1, D2)}, {PNotIn(C1, D2), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D1), PNotIn(C1, D2)}, {PutDown(R1, C1, D3), PNotIn(C1, D2)}, {PNotIn(C1, D2), Move(R1, D1, D3)}, {PNotIn(C1, D2), Move(R1, D1, D2)}, {PNotIn(C1, D2), Move(R1, D3, D1)}, {PNotIn(C1, D2), Move(R1, D3, D2)}, {PIn(C1, R1), PIn(C1, D3)}, {PNotIn(C2, R1), PIn(C1, D3)}, {PickUp(R1, C1, D1), PIn(C1, D3)}, {PickUp(R1, C1, D2), PIn(C1, D3)}, {PutDown(R1, C2, D1), PIn(C1, D3)}, {PutDown(R1, C2, D3), PIn(C1, D3)}, {PutDown(R1, C2, D2), PIn(C1, D3)}, {PutDown(R1, C3, D1), PIn(C1, D3)}, {PutDown(R1, C3, D3), PIn(C1, D3)}, {PutDown(R1, C3, D2), PIn(C1, D3)}, {PutDown(R1, C1, D1), PIn(C1, D3)}, {PutDown(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D2), PIn(C1, D3)}, {PNotIn(R1, D1), PNotIn(C2, R1)}, {PNotIn(R1, D1), PickUp(R1, C2, D1)}, {PNotIn(R1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D1), PNotIn(R1, D1)}, {PutDown(R1, C3, D1), PNotIn(R1, D1)}, {PutDown(R1, C1, D1), PNotIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D1, D3)}, {PNotIn(R1, D1), Move(R1, D1, D2)}, {PickUp(R1, C1, D1), PIn(C1, R1)}, {PickUp(R1, C1, D3), PIn(C1, R1)}, {PickUp(R1, C1, D2), PIn(C1, R1)}, {PickUp(R1, C3, D2), PNotIn(C2, R1)}, {PickUp(R1, C1, D3), PNotIn(C2, R1)}, {PickUp(R1, C1, D2), PNotIn(C2, R1)}, {PutDown(R1, C2, D1), PNotIn(C2, R1)}, {PutDown(R1, C2, D3), PNotIn(C2, R1)}, {PutDown(R1, C2, D2), PNotIn(C2, R1)}, {PutDown(R1, C3, D1), PNotIn(C2, R1)}, {PutDown(R1, C3, D3), PNotIn(C2, R1)}, {PutDown(R1, C3, D2), PNotIn(C2, R1)}, {PutDown(R1, C1, D1), PNotIn(C2, R1)}, {PutDown(R1, C1, D3), PNotIn(C2, R1)}, {PutDown(R1, C1, D2), PNotIn(C2, R1)}, {PNotIn(C2, R1), Move(R1, D3, D1)}, {PNotIn(C2, R1), Move(R1, D3, D2)}, {PNotIn(C2, R1), Move(R1, D2, D1)}, {PNotIn(C2, R1), Move(R1, D2, D3)}, {PIn(C2, D1), PutDown(R1, C2, D1)}, {PIn(C2, D1), PutDown(R1, C2, D3)}, {PIn(C2, D1), PutDown(R1, C2, D2)}, {PickUp(R1, C3, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D2)}, {PickUp(R1, C2, D1), Move(R1, D2, D1)}, {PickUp(R1, C2, D1), Move(R1, D2, D3)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C3, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), Move(R1, D3, D1)}, {PickUp(R1, C3, D2), Move(R1, D3, D2)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D2)}, {PickUp(R1, C1, D1), Move(R1, D2, D1)}, {PickUp(R1, C1, D1), Move(R1, D2, D3)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), Move(R1, D2, D1)}, {PickUp(R1, C1, D3), Move(R1, D2, D3)}, {PickUp(R1, C1, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), Move(R1, D3, D1)}, {PickUp(R1, C1, D2), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D1)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C2, D1), Move(R1, D3, D1)}, {PutDown(R1, C2, D1), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), Move(R1, D2, D1)}, {PutDown(R1, C2, D1), Move(R1, D2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C3, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D2)}, {PutDown(R1, C2, D3), Move(R1, D2, D1)}, {PutDown(R1, C2, D3), Move(R1, D2, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D3, D1)}, {PutDown(R1, C2, D2), Move(R1, D3, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C3, D1), Move(R1, D1, D3)}, {PutDown(R1, C3, D1), Move(R1, D1, D2)}, {PutDown(R1, C3, D1), Move(R1, D3, D1)}, {PutDown(R1, C3, D1), Move(R1, D3, D2)}, {PutDown(R1, C3, D1), Move(R1, D2, D1)}, {PutDown(R1, C3, D1), Move(R1, D2, D3)}, {PutDown(R1, C3, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D3), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D3), Move(R1, D1, D3)}, {PutDown(R1, C3, D3), Move(R1, D1, D2)}, {PutDown(R1, C3, D3), Move(R1, D3, D1)}, {PutDown(R1, C3, D3), Move(R1, D3, D2)}, {PutDown(R1, C3, D3), Move(R1, D2, D1)}, {PutDown(R1, C3, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D2), Move(R1, D1, D3)}, {PutDown(R1, C3, D2), Move(R1, D1, D2)}, {PutDown(R1, C3, D2), Move(R1, D3, D1)}, {PutDown(R1, C3, D2), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C1, D1), Move(R1, D3, D1)}, {PutDown(R1, C1, D1), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), Move(R1, D2, D1)}, {PutDown(R1, C1, D1), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D2, D1)}, {PutDown(R1, C1, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D2)}, {PutDown(R1, C1, D2), Move(R1, D3, D1)}, {PutDown(R1, C1, D2), Move(R1, D3, D2)}, {Move(R1, D1, D3), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D3, D2)}, {Move(R1, D3, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), PickUp(R1, C2, D1)}, {Move(R1, D1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), Move(R1, D2, D1)}, {PickUp(R1, C3, D2), Move(R1, D2, D3)}, {Move(R1, D1, D3), PickUp(R1, C1, D1)}, {Move(R1, D1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D2)}, {PickUp(R1, C1, D2), Move(R1, D2, D1)}, {PickUp(R1, C1, D2), Move(R1, D2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C2, D1), Move(R1, D1, D3)}, {PutDown(R1, C2, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D3, D1)}, {PutDown(R1, C2, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D2, D1)}, {PutDown(R1, C2, D2), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D2), Move(R1, D2, D1)}, {PutDown(R1, C3, D2), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D3, D1)}, {PutDown(R1, C1, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), Move(R1, D2, D1)}, {PutDown(R1, C1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D1, D3)}, {Move(R1, D3, D1), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D2, D3)}}
-  Next state mutexes: {[{NotIn(C1, D3), NotIn(C2, D1)}, {NotIn(C1, D2), NotIn(C2, D1)}, {In(C2, D1), NotIn(C2, D1)}, {NotIn(C3, R1), NotIn(C2, D1)}, {NotIn(C1, D3), In(C2, R1)}, {NotIn(C1, D2), In(C2, R1)}, {NotIn(C2, R1), In(C2, R1)}, {In(C2, D1), In(C2, R1)}, {In(C2, D3), In(C2, R1)}, {In(C2, D2), In(C2, R1)}, {NotIn(C3, R1), In(C2, R1)}, {NotIn(C1, R1), NotIn(C1, D3)}, {NotIn(C1, R1), NotIn(C1, D2)}, {NotIn(C1, R1), In(C1, R1)}, {In(R1, D1), In(R1, D2)}, {In(R1, D1), In(R1, D3)}, {NotIn(R1, D1), In(R1, D1)}, {In(R1, D1), In(C2, D3)}, {In(R1, D1), In(C2, D2)}, {NotIn(C3, R1), In(R1, D1)}, {NotIn(C1, D3), In(C1, D2)}, {NotIn(C1, D3), In(C3, R1)}, {NotIn(C1, D3), NotHolding(R1)}, {NotIn(C1, D3), NotIn(C1, D1)}, {NotIn(C1, D3), In(C1, D1)}, {NotIn(C3, D2), NotIn(C1, D3)}, {NotIn(C1, D3), NotIn(C1, D2)}, {NotIn(C1, D3), In(C1, D3)}, {NotIn(C2, R1), NotIn(C1, D3)}, {NotIn(C1, D3), In(C2, D3)}, {NotIn(C1, D3), In(C2, D2)}, {NotIn(C1, D3), NotIn(C3, R1)}, {In(C1, D1), In(C1, D2)}, {NotIn(C1, D2), In(C1, D2)}, {In(C1, D3), In(C1, D2)}, {In(C1, D2), In(C1, R1)}, {NotIn(C2, R1), In(C1, D2)}, {In(C1, D2), In(C2, D3)}, {In(C1, D2), In(C2, D2)}, {In(C3, D2), In(C3, R1)}, {NotIn(C2, R1), In(C3, R1)}, {In(C2, D3), In(C3, R1)}, {In(C2, D2), In(C3, R1)}, {NotIn(C3, R1), In(C3, R1)}, {In(R1, D3), In(R1, D2)}, {NotIn(R1, D2), In(R1, D2)}, {In(C2, D3), In(R1, D2)}, {NotHolding(R1), Holding(R1)}, {Holding(R1), In(C2, D3)}, {Holding(R1), In(C2, D2)}, {NotIn(C3, R1), Holding(R1)}, {NotIn(C3, D2), In(C3, D2)}, {NotIn(R1, D3), In(R1, D3)}, {In(C2, D2), In(R1, D3)}, {NotIn(C3, R1), In(R1, D3)}, {In(C1, D1), NotIn(C1, D1)}, {NotIn(C1, D2), NotIn(C1, D1)}, {NotIn(C3, R1), NotIn(C1, D1)}, {NotIn(C1, D2), In(C1, D1)}, {In(C1, D1), In(C1, D3)}, {In(C1, D1), In(C1, R1)}, {NotIn(C2, R1), NotIn(R1, D3)}, {NotIn(R1, D3), In(C2, D3)}, {NotIn(R1, D3), In(C2, D2)}, {NotIn(R1, D3), NotIn(C3, R1)}, {NotIn(C2, R1), NotIn(C3, D2)}, {NotIn(C3, D2), In(C2, D3)}, {NotIn(C3, D2), In(C2, D2)}, {NotIn(C3, D2), NotIn(C3, R1)}, {NotIn(R1, D2), NotIn(C2, R1)}, {NotIn(R1, D2), In(C2, D3)}, {NotIn(R1, D2), In(C2, D2)}, {NotIn(R1, D2), NotIn(C3, R1)}, {NotIn(C1, D2), In(C1, D3)}, {NotIn(C2, R1), NotIn(C1, D2)}, {NotIn(C1, D2), In(C2, D3)}, {NotIn(C1, D2), In(C2, D2)}, {In(C1, D3), In(C1, R1)}, {NotIn(C2, R1), In(C1, D3)}, {In(C1, D3), In(C2, D3)}, {In(C1, D3), In(C2, D2)}, {NotIn(C3, R1), In(C1, D3)}, {NotIn(C2, R1), NotIn(C3, R1)}, {In(C2, D1), In(C2, D3)}, {In(C2, D1), In(C2, D2)}, {In(C2, D3), In(C2, D2)}, {NotIn(C3, R1), In(C2, D3)}, {NotIn(C3, R1), In(C2, D2)}]}
-
-Level 4:
-<Level>
-  Current State: {Place(D3), NotIn(C2, D1), In(C2, R1), NotIn(C1, R1), In(R1, D1), Place(D2), NotIn(C1, D3), Container(C3), Container(C1), In(C1, D2), In(C2, D3), In(C3, R1), In(R1, D2), Holding(R1), In(C3, D2), NotHolding(R1), In(C2, D2), In(R1, D3), NotIn(C1, D1), Container(C2), In(C1, D1), NotIn(R1, D3), NotIn(C3, D2), NotIn(R1, D2), NotIn(C3, R1), Place(D1), NotIn(C1, D2), In(C1, D3), NotIn(R1, D1), In(C1, R1), Robot(R1), NotIn(C2, R1), In(C2, D1)}
-  Actions: {PPlace(D3), PNotIn(C2, D1), PIn(C2, R1), PNotIn(C1, R1), PIn(R1, D1), PPlace(D2), PNotIn(C1, D3), PContainer(C3), PContainer(C1), PIn(C1, D2), PIn(C2, D3), PIn(C3, R1), PIn(R1, D2), PHolding(R1), PIn(C3, D2), PNotHolding(R1), PIn(C2, D2), PIn(R1, D3), PNotIn(C1, D1), PContainer(C2), PIn(C1, D1), PNotIn(R1, D3), PNotIn(C3, D2), PNotIn(R1, D2), PNotIn(C3, R1), PPlace(D1), PNotIn(C1, D2), PIn(C1, D3), PNotIn(R1, D1), PIn(C1, R1), PRobot(R1), PNotIn(C2, R1), PIn(C2, D1), PickUp(R1, C2, D1), PickUp(R1, C2, D3), PickUp(R1, C2, D2), PickUp(R1, C3, D2), PickUp(R1, C1, D1), PickUp(R1, C1, D3), PickUp(R1, C1, D2), PutDown(R1, C2, D1), PutDown(R1, C2, D3), PutDown(R1, C2, D2), PutDown(R1, C3, D1), PutDown(R1, C3, D3), PutDown(R1, C3, D2), PutDown(R1, C1, D1), PutDown(R1, C1, D3), PutDown(R1, C1, D2), Move(R1, D1, D3), Move(R1, D1, D2), Move(R1, D3, D1), Move(R1, D3, D2), Move(R1, D2, D1), Move(R1, D2, D3)}
-  Mutex: {{PIn(C2, D1), PNotIn(C2, D1)}, {PIn(C2, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D1), PNotIn(C2, D1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D1)}, {PNotIn(C1, R1), PIn(C1, R1)}, {PutDown(R1, C1, D1), PIn(C1, R1)}, {PutDown(R1, C1, D3), PIn(C1, R1)}, {PutDown(R1, C1, D2), PIn(C1, R1)}, {PNotIn(C1, R1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D2)}, {PNotIn(C1, D3), PIn(C1, D3)}, {PickUp(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D3), PNotIn(C1, D3)}, {PHolding(R1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PHolding(R1)}, {PutDown(R1, C2, D3), PHolding(R1)}, {PHolding(R1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PHolding(R1)}, {PutDown(R1, C3, D3), PHolding(R1)}, {PutDown(R1, C3, D2), PHolding(R1)}, {PutDown(R1, C1, D1), PHolding(R1)}, {PutDown(R1, C1, D3), PHolding(R1)}, {PutDown(R1, C1, D2), PHolding(R1)}, {PickUp(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C2, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D2)}, {PickUp(R1, C2, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D2)}, {PickUp(R1, C2, D2), PutDown(R1, C3, D3)}, {PickUp(R1, C2, D2), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D2)}, {PickUp(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D2), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D2)}, {PIn(C1, D1), PNotIn(C1, D1)}, {PIn(C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PNotIn(C1, D1)}, {PNotIn(R1, D3), PIn(R1, D3)}, {Move(R1, D3, D1), PIn(R1, D3)}, {Move(R1, D3, D2), PIn(R1, D3)}, {PNotIn(R1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D3), Move(R1, D3, D1)}, {Move(R1, D1, D3), Move(R1, D3, D2)}, {PNotIn(R1, D3), Move(R1, D2, D3)}, {Move(R1, D3, D1), Move(R1, D2, D3)}, {Move(R1, D3, D2), Move(R1, D2, D3)}, {PNotIn(C3, D2), PIn(C3, D2)}, {PickUp(R1, C3, D2), PIn(C3, D2)}, {PutDown(R1, C3, D2), PNotIn(C3, D2)}, {PNotIn(R1, D2), PIn(R1, D2)}, {Move(R1, D2, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), PNotIn(R1, D2)}, {Move(R1, D1, D2), Move(R1, D2, D1)}, {Move(R1, D1, D2), Move(R1, D2, D3)}, {PNotIn(R1, D2), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D3, D2)}, {PNotIn(C3, R1), PIn(C3, R1)}, {PutDown(R1, C3, D1), PIn(C3, R1)}, {PutDown(R1, C3, D3), PIn(C3, R1)}, {PutDown(R1, C3, D2), PIn(C3, R1)}, {PNotIn(C3, R1), PickUp(R1, C3, D2)}, {PNotIn(C1, D2), PIn(C1, D2)}, {PickUp(R1, C1, D2), PIn(C1, D2)}, {PNotIn(C1, D2), PutDown(R1, C1, D2)}, {PNotIn(R1, D1), PIn(R1, D1)}, {Move(R1, D1, D3), PIn(R1, D1)}, {Move(R1, D1, D2), PIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D3, D1)}, {Move(R1, D1, D2), Move(R1, D3, D1)}, {PNotIn(R1, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), Move(R1, D2, D1)}, {PNotIn(C2, R1), PIn(C2, R1)}, {PutDown(R1, C2, D1), PIn(C2, R1)}, {PutDown(R1, C2, D3), PIn(C2, R1)}, {PutDown(R1, C2, D2), PIn(C2, R1)}, {PickUp(R1, C2, D1), PNotIn(C2, R1)}, {PickUp(R1, C2, D3), PNotIn(C2, R1)}, {PickUp(R1, C2, D2), PNotIn(C2, R1)}, {PickUp(R1, C2, D3), PIn(C2, D3)}, {PickUp(R1, C2, D2), PIn(C2, D2)}, {PNotIn(C1, D3), PNotIn(C2, D1)}, {PNotIn(C3, R1), PNotIn(C2, D1)}, {PNotIn(C1, D2), PNotIn(C2, D1)}, {PickUp(R1, C2, D1), PNotIn(C2, D1)}, {PNotIn(C1, D3), PIn(C2, R1)}, {PIn(C2, R1), PIn(C2, D3)}, {PIn(C2, R1), PIn(C2, D2)}, {PNotIn(C3, R1), PIn(C2, R1)}, {PNotIn(C1, D2), PIn(C2, R1)}, {PIn(C2, D1), PIn(C2, R1)}, {PickUp(R1, C2, D1), PIn(C2, R1)}, {PickUp(R1, C2, D3), PIn(C2, R1)}, {PickUp(R1, C2, D2), PIn(C2, R1)}, {PNotIn(C1, D3), PNotIn(C1, R1)}, {PNotIn(C1, D2), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PNotIn(C1, R1)}, {PutDown(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D2), PNotIn(C1, R1)}, {PIn(C2, D3), PIn(R1, D1)}, {PIn(R1, D2), PIn(R1, D1)}, {PIn(C2, D2), PIn(R1, D1)}, {PIn(R1, D3), PIn(R1, D1)}, {PNotIn(C3, R1), PIn(R1, D1)}, {PickUp(R1, C2, D3), PIn(R1, D1)}, {PickUp(R1, C2, D2), PIn(R1, D1)}, {PickUp(R1, C3, D2), PIn(R1, D1)}, {PickUp(R1, C1, D3), PIn(R1, D1)}, {PickUp(R1, C1, D2), PIn(R1, D1)}, {PutDown(R1, C2, D3), PIn(R1, D1)}, {PutDown(R1, C2, D2), PIn(R1, D1)}, {PutDown(R1, C3, D3), PIn(R1, D1)}, {PutDown(R1, C3, D2), PIn(R1, D1)}, {PutDown(R1, C1, D3), PIn(R1, D1)}, {PutDown(R1, C1, D2), PIn(R1, D1)}, {Move(R1, D3, D1), PIn(R1, D1)}, {Move(R1, D3, D2), PIn(R1, D1)}, {Move(R1, D2, D1), PIn(R1, D1)}, {Move(R1, D2, D3), PIn(R1, D1)}, {PNotIn(C1, D3), PIn(C1, D2)}, {PNotIn(C1, D3), PIn(C2, D3)}, {PNotIn(C1, D3), PIn(C3, R1)}, {PNotIn(C1, D3), PNotHolding(R1)}, {PNotIn(C1, D3), PIn(C2, D2)}, {PNotIn(C1, D3), PNotIn(C1, D1)}, {PIn(C1, D1), PNotIn(C1, D3)}, {PNotIn(C1, D3), PNotIn(C3, D2)}, {PNotIn(C3, R1), PNotIn(C1, D3)}, {PNotIn(C1, D2), PNotIn(C1, D3)}, {PNotIn(C1, D3), PNotIn(C2, R1)}, {PNotIn(C1, D3), PickUp(R1, C2, D1)}, {PNotIn(C1, D3), PickUp(R1, C2, D3)}, {PNotIn(C1, D3), PickUp(R1, C2, D2)}, {PNotIn(C1, D3), PickUp(R1, C3, D2)}, {PNotIn(C1, D3), PickUp(R1, C1, D1)}, {PNotIn(C1, D3), PickUp(R1, C1, D3)}, {PNotIn(C1, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D1), PNotIn(C1, D3)}, {PutDown(R1, C2, D3), PNotIn(C1, D3)}, {PNotIn(C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PNotIn(C1, D3)}, {PNotIn(C1, D3), PutDown(R1, C3, D3)}, {PNotIn(C1, D3), PutDown(R1, C3, D2)}, {PIn(C2, D3), PIn(C1, D2)}, {PIn(C2, D2), PIn(C1, D2)}, {PIn(C1, D1), PIn(C1, D2)}, {PIn(C1, D3), PIn(C1, D2)}, {PIn(C1, R1), PIn(C1, D2)}, {PNotIn(C2, R1), PIn(C1, D2)}, {PickUp(R1, C2, D3), PIn(C1, D2)}, {PickUp(R1, C2, D2), PIn(C1, D2)}, {PickUp(R1, C1, D1), PIn(C1, D2)}, {PickUp(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C1, D1), PIn(C1, D2)}, {PutDown(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C1, D2), PIn(C1, D2)}, {PIn(C3, R1), PIn(C2, D3)}, {PIn(R1, D2), PIn(C2, D3)}, {PHolding(R1), PIn(C2, D3)}, {PIn(C2, D2), PIn(C2, D3)}, {PNotIn(R1, D3), PIn(C2, D3)}, {PNotIn(C3, D2), PIn(C2, D3)}, {PNotIn(R1, D2), PIn(C2, D3)}, {PNotIn(C3, R1), PIn(C2, D3)}, {PNotIn(C1, D2), PIn(C2, D3)}, {PIn(C1, D3), PIn(C2, D3)}, {PIn(C2, D1), PIn(C2, D3)}, {PickUp(R1, C2, D1), PIn(C2, D3)}, {PickUp(R1, C2, D2), PIn(C2, D3)}, {PickUp(R1, C3, D2), PIn(C2, D3)}, {PickUp(R1, C1, D1), PIn(C2, D3)}, {PickUp(R1, C1, D3), PIn(C2, D3)}, {PickUp(R1, C1, D2), PIn(C2, D3)}, {PutDown(R1, C2, D1), PIn(C2, D3)}, {PutDown(R1, C2, D3), PIn(C2, D3)}, {PutDown(R1, C2, D2), PIn(C2, D3)}, {PutDown(R1, C3, D1), PIn(C2, D3)}, {PutDown(R1, C3, D3), PIn(C2, D3)}, {PutDown(R1, C3, D2), PIn(C2, D3)}, {PutDown(R1, C1, D1), PIn(C2, D3)}, {PutDown(R1, C1, D3), PIn(C2, D3)}, {PutDown(R1, C1, D2), PIn(C2, D3)}, {Move(R1, D1, D3), PIn(C2, D3)}, {Move(R1, D1, D2), PIn(C2, D3)}, {Move(R1, D2, D1), PIn(C2, D3)}, {Move(R1, D2, D3), PIn(C2, D3)}, {PIn(C3, R1), PIn(C3, D2)}, {PIn(C3, R1), PIn(C2, D2)}, {PNotIn(C2, R1), PIn(C3, R1)}, {PickUp(R1, C2, D3), PIn(C3, R1)}, {PickUp(R1, C2, D2), PIn(C3, R1)}, {PickUp(R1, C3, D2), PIn(C3, R1)}, {PIn(R1, D2), PIn(R1, D3)}, {PickUp(R1, C2, D1), PIn(R1, D2)}, {PickUp(R1, C2, D3), PIn(R1, D2)}, {PickUp(R1, C1, D1), PIn(R1, D2)}, {PickUp(R1, C1, D3), PIn(R1, D2)}, {PutDown(R1, C2, D1), PIn(R1, D2)}, {PutDown(R1, C2, D3), PIn(R1, D2)}, {PutDown(R1, C3, D1), PIn(R1, D2)}, {PutDown(R1, C3, D3), PIn(R1, D2)}, {PutDown(R1, C1, D1), PIn(R1, D2)}, {PutDown(R1, C1, D3), PIn(R1, D2)}, {Move(R1, D1, D3), PIn(R1, D2)}, {Move(R1, D1, D2), PIn(R1, D2)}, {Move(R1, D3, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D3, D2)}, {PHolding(R1), PIn(C2, D2)}, {PNotIn(C3, R1), PHolding(R1)}, {PHolding(R1), PickUp(R1, C2, D1)}, {PHolding(R1), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D2), PHolding(R1)}, {PHolding(R1), PickUp(R1, C3, D2)}, {PHolding(R1), PickUp(R1, C1, D1)}, {PHolding(R1), PickUp(R1, C1, D3)}, {PHolding(R1), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D1), PIn(C3, D2)}, {PutDown(R1, C3, D3), PIn(C3, D2)}, {PutDown(R1, C3, D2), PIn(C3, D2)}, {PutDown(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D2), PNotHolding(R1)}, {PutDown(R1, C3, D1), PNotHolding(R1)}, {PutDown(R1, C3, D3), PNotHolding(R1)}, {PutDown(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C1, D2), PNotHolding(R1)}, {PIn(R1, D3), PIn(C2, D2)}, {PNotIn(R1, D3), PIn(C2, D2)}, {PNotIn(C3, D2), PIn(C2, D2)}, {PNotIn(R1, D2), PIn(C2, D2)}, {PNotIn(C3, R1), PIn(C2, D2)}, {PNotIn(C1, D2), PIn(C2, D2)}, {PIn(C1, D3), PIn(C2, D2)}, {PIn(C2, D1), PIn(C2, D2)}, {PickUp(R1, C2, D1), PIn(C2, D2)}, {PickUp(R1, C2, D3), PIn(C2, D2)}, {PickUp(R1, C1, D1), PIn(C2, D2)}, {PickUp(R1, C1, D3), PIn(C2, D2)}, {PickUp(R1, C1, D2), PIn(C2, D2)}, {PutDown(R1, C2, D1), PIn(C2, D2)}, {PutDown(R1, C2, D3), PIn(C2, D2)}, {PutDown(R1, C2, D2), PIn(C2, D2)}, {PutDown(R1, C3, D1), PIn(C2, D2)}, {PutDown(R1, C3, D3), PIn(C2, D2)}, {PutDown(R1, C3, D2), PIn(C2, D2)}, {PutDown(R1, C1, D1), PIn(C2, D2)}, {PutDown(R1, C1, D3), PIn(C2, D2)}, {PutDown(R1, C1, D2), PIn(C2, D2)}, {Move(R1, D1, D3), PIn(C2, D2)}, {Move(R1, D1, D2), PIn(C2, D2)}, {Move(R1, D3, D1), PIn(C2, D2)}, {Move(R1, D3, D2), PIn(C2, D2)}, {PNotIn(C3, R1), PIn(R1, D3)}, {PickUp(R1, C2, D1), PIn(R1, D3)}, {PickUp(R1, C2, D2), PIn(R1, D3)}, {PickUp(R1, C3, D2), PIn(R1, D3)}, {PickUp(R1, C1, D1), PIn(R1, D3)}, {PickUp(R1, C1, D2), PIn(R1, D3)}, {PutDown(R1, C2, D1), PIn(R1, D3)}, {PutDown(R1, C2, D2), PIn(R1, D3)}, {PutDown(R1, C3, D1), PIn(R1, D3)}, {PutDown(R1, C3, D2), PIn(R1, D3)}, {PutDown(R1, C1, D1), PIn(R1, D3)}, {PutDown(R1, C1, D2), PIn(R1, D3)}, {Move(R1, D1, D3), PIn(R1, D3)}, {Move(R1, D1, D2), PIn(R1, D3)}, {Move(R1, D2, D1), PIn(R1, D3)}, {Move(R1, D2, D3), PIn(R1, D3)}, {PNotIn(C3, R1), PNotIn(C1, D1)}, {PNotIn(C1, D2), PNotIn(C1, D1)}, {PickUp(R1, C1, D1), PNotIn(C1, D1)}, {PIn(C1, D1), PNotIn(C1, D2)}, {PIn(C1, D1), PIn(C1, D3)}, {PIn(C1, D1), PIn(C1, R1)}, {PIn(C1, D1), PickUp(R1, C1, D3)}, {PIn(C1, D1), PickUp(R1, C1, D2)}, {PIn(C1, D1), PutDown(R1, C1, D1)}, {PIn(C1, D1), PutDown(R1, C1, D3)}, {PIn(C1, D1), PutDown(R1, C1, D2)}, {PNotIn(R1, D3), PNotIn(C3, R1)}, {PNotIn(R1, D3), PNotIn(C2, R1)}, {PNotIn(R1, D3), PickUp(R1, C2, D3)}, {PNotIn(R1, D3), PickUp(R1, C2, D2)}, {PNotIn(R1, D3), PickUp(R1, C1, D3)}, {PNotIn(R1, D3), PutDown(R1, C2, D3)}, {PNotIn(R1, D3), PutDown(R1, C3, D3)}, {PNotIn(R1, D3), PutDown(R1, C1, D3)}, {PNotIn(R1, D3), Move(R1, D3, D1)}, {PNotIn(R1, D3), Move(R1, D3, D2)}, {PNotIn(C3, R1), PNotIn(C3, D2)}, {PNotIn(C3, D2), PNotIn(C2, R1)}, {PNotIn(C3, D2), PickUp(R1, C2, D3)}, {PNotIn(C3, D2), PickUp(R1, C2, D2)}, {PNotIn(C3, D2), PickUp(R1, C3, D2)}, {PNotIn(C3, R1), PNotIn(R1, D2)}, {PNotIn(R1, D2), PNotIn(C2, R1)}, {PickUp(R1, C2, D3), PNotIn(R1, D2)}, {PickUp(R1, C2, D2), PNotIn(R1, D2)}, {PickUp(R1, C3, D2), PNotIn(R1, D2)}, {PickUp(R1, C1, D2), PNotIn(R1, D2)}, {PutDown(R1, C2, D2), PNotIn(R1, D2)}, {PutDown(R1, C3, D2), PNotIn(R1, D2)}, {PutDown(R1, C1, D2), PNotIn(R1, D2)}, {PNotIn(R1, D2), Move(R1, D2, D1)}, {PNotIn(R1, D2), Move(R1, D2, D3)}, {PNotIn(C3, R1), PIn(C1, D3)}, {PNotIn(C3, R1), PNotIn(C2, R1)}, {PNotIn(C3, R1), PickUp(R1, C2, D1)}, {PNotIn(C3, R1), PickUp(R1, C2, D3)}, {PNotIn(C3, R1), PickUp(R1, C2, D2)}, {PNotIn(C3, R1), PickUp(R1, C1, D1)}, {PNotIn(C3, R1), PickUp(R1, C1, D3)}, {PNotIn(C3, R1), PutDown(R1, C2, D1)}, {PNotIn(C3, R1), PutDown(R1, C2, D3)}, {PNotIn(C3, R1), PutDown(R1, C2, D2)}, {PNotIn(C3, R1), PutDown(R1, C3, D1)}, {PNotIn(C3, R1), PutDown(R1, C3, D3)}, {PNotIn(C3, R1), PutDown(R1, C3, D2)}, {PNotIn(C3, R1), PutDown(R1, C1, D1)}, {PNotIn(C3, R1), PutDown(R1, C1, D3)}, {PNotIn(C3, R1), PutDown(R1, C1, D2)}, {PNotIn(C3, R1), Move(R1, D1, D3)}, {PNotIn(C3, R1), Move(R1, D1, D2)}, {PNotIn(C3, R1), Move(R1, D3, D1)}, {PNotIn(C3, R1), Move(R1, D3, D2)}, {PNotIn(C1, D2), PIn(C1, D3)}, {PNotIn(C1, D2), PNotIn(C2, R1)}, {PNotIn(C1, D2), PickUp(R1, C2, D3)}, {PNotIn(C1, D2), PickUp(R1, C2, D2)}, {PNotIn(C1, D2), PickUp(R1, C1, D1)}, {PNotIn(C1, D2), PickUp(R1, C1, D3)}, {PNotIn(C1, D2), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D1), PNotIn(C1, D2)}, {PutDown(R1, C2, D3), PNotIn(C1, D2)}, {PNotIn(C1, D2), PutDown(R1, C2, D2)}, {PIn(C1, R1), PIn(C1, D3)}, {PNotIn(C2, R1), PIn(C1, D3)}, {PickUp(R1, C2, D3), PIn(C1, D3)}, {PickUp(R1, C2, D2), PIn(C1, D3)}, {PickUp(R1, C1, D1), PIn(C1, D3)}, {PickUp(R1, C1, D2), PIn(C1, D3)}, {PutDown(R1, C1, D1), PIn(C1, D3)}, {PutDown(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D2), PIn(C1, D3)}, {PNotIn(R1, D1), PickUp(R1, C2, D1)}, {PNotIn(R1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D1), PNotIn(R1, D1)}, {PutDown(R1, C3, D1), PNotIn(R1, D1)}, {PutDown(R1, C1, D1), PNotIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D1, D3)}, {PNotIn(R1, D1), Move(R1, D1, D2)}, {PickUp(R1, C1, D1), PIn(C1, R1)}, {PickUp(R1, C1, D3), PIn(C1, R1)}, {PickUp(R1, C1, D2), PIn(C1, R1)}, {PickUp(R1, C1, D3), PNotIn(C2, R1)}, {PickUp(R1, C1, D2), PNotIn(C2, R1)}, {PutDown(R1, C2, D1), PNotIn(C2, R1)}, {PutDown(R1, C2, D3), PNotIn(C2, R1)}, {PutDown(R1, C2, D2), PNotIn(C2, R1)}, {PutDown(R1, C3, D1), PNotIn(C2, R1)}, {PutDown(R1, C3, D3), PNotIn(C2, R1)}, {PutDown(R1, C3, D2), PNotIn(C2, R1)}, {PIn(C2, D1), PickUp(R1, C2, D3)}, {PIn(C2, D1), PickUp(R1, C2, D2)}, {PIn(C2, D1), PutDown(R1, C2, D1)}, {PIn(C2, D1), PutDown(R1, C2, D3)}, {PIn(C2, D1), PutDown(R1, C2, D2)}, {PickUp(R1, C2, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D2)}, {PickUp(R1, C2, D1), Move(R1, D2, D1)}, {PickUp(R1, C2, D1), Move(R1, D2, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C2, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C2, D3), PickUp(R1, C1, D3)}, {PickUp(R1, C2, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C2, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), Move(R1, D2, D1)}, {PickUp(R1, C2, D3), Move(R1, D2, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C2, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C1, D2)}, {PickUp(R1, C2, D2), Move(R1, D1, D3)}, {PickUp(R1, C2, D2), Move(R1, D1, D2)}, {PickUp(R1, C2, D2), Move(R1, D3, D1)}, {PickUp(R1, C2, D2), Move(R1, D3, D2)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C3, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), Move(R1, D3, D1)}, {PickUp(R1, C3, D2), Move(R1, D3, D2)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D2)}, {PickUp(R1, C1, D1), Move(R1, D2, D1)}, {PickUp(R1, C1, D1), Move(R1, D2, D3)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), Move(R1, D2, D1)}, {PickUp(R1, C1, D3), Move(R1, D2, D3)}, {PickUp(R1, C1, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), Move(R1, D3, D1)}, {PickUp(R1, C1, D2), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C2, D1), Move(R1, D3, D1)}, {PutDown(R1, C2, D1), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), Move(R1, D2, D1)}, {PutDown(R1, C2, D1), Move(R1, D2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D2)}, {PutDown(R1, C2, D3), Move(R1, D2, D1)}, {PutDown(R1, C2, D3), Move(R1, D2, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D3, D1)}, {PutDown(R1, C2, D2), Move(R1, D3, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C3, D1), Move(R1, D3, D1)}, {PutDown(R1, C3, D1), Move(R1, D3, D2)}, {PutDown(R1, C3, D1), Move(R1, D2, D1)}, {PutDown(R1, C3, D1), Move(R1, D2, D3)}, {PutDown(R1, C3, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D3), Move(R1, D1, D3)}, {PutDown(R1, C3, D3), Move(R1, D1, D2)}, {PutDown(R1, C3, D3), Move(R1, D2, D1)}, {PutDown(R1, C3, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D2), Move(R1, D1, D3)}, {PutDown(R1, C3, D2), Move(R1, D1, D2)}, {PutDown(R1, C3, D2), Move(R1, D3, D1)}, {PutDown(R1, C3, D2), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C1, D1), Move(R1, D3, D1)}, {PutDown(R1, C1, D1), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), Move(R1, D2, D1)}, {PutDown(R1, C1, D1), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D2, D1)}, {PutDown(R1, C1, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D2)}, {PutDown(R1, C1, D2), Move(R1, D3, D1)}, {PutDown(R1, C1, D2), Move(R1, D3, D2)}, {Move(R1, D1, D3), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D3, D2)}, {Move(R1, D3, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), PickUp(R1, C2, D1)}, {Move(R1, D1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D3), Move(R1, D3, D1)}, {PickUp(R1, C2, D3), Move(R1, D3, D2)}, {PickUp(R1, C2, D2), Move(R1, D2, D1)}, {PickUp(R1, C2, D2), Move(R1, D2, D3)}, {PickUp(R1, C3, D2), Move(R1, D2, D1)}, {PickUp(R1, C3, D2), Move(R1, D2, D3)}, {Move(R1, D1, D3), PickUp(R1, C1, D1)}, {Move(R1, D1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D2)}, {PickUp(R1, C1, D2), Move(R1, D2, D1)}, {PickUp(R1, C1, D2), Move(R1, D2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D1)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C2, D1), Move(R1, D1, D3)}, {PutDown(R1, C2, D1), Move(R1, D1, D2)}, {PutDown(R1, C2, D3), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D3, D1)}, {PutDown(R1, C2, D3), Move(R1, D3, D2)}, {PutDown(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D2, D1)}, {PutDown(R1, C2, D2), Move(R1, D2, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C3, D1), Move(R1, D1, D3)}, {PutDown(R1, C3, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D3), Move(R1, D3, D1)}, {PutDown(R1, C3, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D2), Move(R1, D2, D1)}, {PutDown(R1, C3, D2), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D3, D1)}, {PutDown(R1, C1, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), Move(R1, D2, D1)}, {PutDown(R1, C1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D1, D3)}, {Move(R1, D3, D1), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D2, D3)}}
-  Next state mutexes: {[{In(C2, D1), NotIn(C2, D1)}, {In(C2, D3), In(C2, R1)}, {In(C2, D2), In(C2, R1)}, {NotIn(C2, R1), In(C2, R1)}, {In(C2, D1), In(C2, R1)}, {NotIn(C1, R1), In(C1, R1)}, {In(R1, D1), In(R1, D2)}, {In(R1, D1), In(R1, D3)}, {NotIn(R1, D1), In(R1, D1)}, {NotIn(C2, D3), In(R1, D1)}, {NotIn(C2, D2), In(R1, D1)}, {In(C3, D3), In(R1, D1)}, {NotIn(C1, D3), In(C2, D3)}, {NotIn(C1, D3), In(C2, D2)}, {NotIn(C1, D3), NotIn(C3, R1)}, {NotIn(C1, D3), NotIn(C1, D2)}, {NotIn(C1, D3), In(C1, D3)}, {NotIn(C2, R1), NotIn(C1, D3)}, {NotIn(C1, D3), NotIn(C2, D3)}, {NotIn(C2, D2), NotIn(C1, D3)}, {NotIn(C1, D3), In(C3, D1)}, {NotIn(C1, D3), In(C3, D3)}, {In(C1, D1), In(C1, D2)}, {NotIn(C1, D2), In(C1, D2)}, {In(C1, D3), In(C1, D2)}, {In(C1, D2), In(C1, R1)}, {NotIn(C2, D3), In(C1, D2)}, {NotIn(C2, D2), In(C1, D2)}, {Holding(R1), In(C2, D3)}, {In(C2, D3), In(C2, D2)}, {NotIn(C3, R1), In(C2, D3)}, {NotIn(C1, D2), In(C2, D3)}, {In(C2, D1), In(C2, D3)}, {NotIn(C2, D3), In(C2, D3)}, {NotIn(C2, D2), In(C2, D3)}, {In(C3, D1), In(C2, D3)}, {In(C3, D3), In(C2, D3)}, {In(C3, D2), In(C3, R1)}, {NotIn(C3, R1), In(C3, R1)}, {NotIn(C2, D3), In(C3, R1)}, {In(C3, D1), In(C3, R1)}, {In(C3, D3), In(C3, R1)}, {In(R1, D3), In(R1, D2)}, {NotIn(R1, D2), In(R1, D2)}, {NotIn(C2, D3), In(R1, D2)}, {In(C3, D1), In(R1, D2)}, {In(C3, D3), In(R1, D2)}, {NotHolding(R1), Holding(R1)}, {In(C3, D1), Holding(R1)}, {In(C3, D3), Holding(R1)}, {NotIn(C3, D2), In(C3, D2)}, {In(C3, D1), In(C3, D2)}, {In(C3, D3), In(C3, D2)}, {NotIn(C2, D3), NotHolding(R1)}, {NotIn(C2, D2), NotHolding(R1)}, {NotIn(C3, R1), In(C2, D2)}, {NotIn(C1, D2), In(C2, D2)}, {In(C2, D1), In(C2, D2)}, {NotIn(C2, D3), In(C2, D2)}, {NotIn(C2, D2), In(C2, D2)}, {In(C3, D1), In(C2, D2)}, {In(C3, D3), In(C2, D2)}, {NotIn(R1, D3), In(R1, D3)}, {NotIn(C2, D2), In(R1, D3)}, {In(C3, D1), In(R1, D3)}, {In(C1, D1), NotIn(C1, D1)}, {In(C1, D1), In(C1, D3)}, {In(C1, D1), In(C1, R1)}, {NotIn(C2, D3), NotIn(R1, D3)}, {NotIn(C2, D2), NotIn(R1, D3)}, {NotIn(R1, D3), In(C3, D3)}, {NotIn(C3, D2), NotIn(C2, D3)}, {NotIn(R1, D2), NotIn(C2, D3)}, {NotIn(R1, D2), NotIn(C2, D2)}, {NotIn(C2, R1), NotIn(C3, R1)}, {NotIn(C2, D3), NotIn(C3, R1)}, {NotIn(C2, D2), NotIn(C3, R1)}, {NotIn(C2, R1), NotIn(C1, D2)}, {NotIn(C1, D2), NotIn(C2, D3)}, {NotIn(C2, D2), NotIn(C1, D2)}, {In(C1, D3), In(C1, R1)}, {NotIn(C2, D3), In(C1, D3)}, {NotIn(C2, D2), In(C1, D3)}, {NotIn(R1, D1), In(C3, D1)}, {NotIn(C2, R1), NotIn(C2, D3)}, {NotIn(C2, R1), NotIn(C2, D2)}, {NotIn(C2, R1), In(C3, D1)}, {NotIn(C2, R1), In(C3, D3)}, {NotIn(C2, D3), In(C2, D1)}, {NotIn(C2, D2), In(C2, D1)}, {NotIn(C2, D2), NotIn(C2, D3)}, {NotIn(C2, D3), In(C3, D1)}, {NotIn(C2, D3), In(C3, D3)}, {NotIn(C2, D2), In(C3, D1)}, {NotIn(C2, D2), In(C3, D3)}, {In(C3, D1), In(C3, D3)}]}
-
-Level 5:
-<Level>
-  Current State: {Place(D3), NotIn(C2, D1), In(C2, R1), NotIn(C1, R1), In(R1, D1), Place(D2), NotIn(C1, D3), Container(C3), In(C2, D3), Container(C1), In(C1, D2), In(C3, D1), In(C3, R1), In(R1, D2), Holding(R1), In(C3, D2), NotHolding(R1), In(C2, D2), In(R1, D3), In(C3, D3), NotIn(C1, D1), Container(C2), In(C1, D1), NotIn(C2, D2), NotIn(R1, D3), NotIn(C3, D2), NotIn(R1, D2), NotIn(C3, R1), Place(D1), NotIn(C1, D2), NotIn(C2, D3), In(C1, D3), NotIn(R1, D1), In(C1, R1), Robot(R1), NotIn(C2, R1), In(C2, D1)}
-  Actions: {PPlace(D3), PNotIn(C2, D1), PIn(C2, R1), PNotIn(C1, R1), PIn(R1, D1), PPlace(D2), PNotIn(C1, D3), PContainer(C3), PIn(C2, D3), PContainer(C1), PIn(C1, D2), PIn(C3, D1), PIn(C3, R1), PIn(R1, D2), PHolding(R1), PIn(C3, D2), PNotHolding(R1), PIn(C2, D2), PIn(R1, D3), PIn(C3, D3), PNotIn(C1, D1), PContainer(C2), PIn(C1, D1), PNotIn(C2, D2), PNotIn(R1, D3), PNotIn(C3, D2), PNotIn(R1, D2), PNotIn(C3, R1), PPlace(D1), PNotIn(C1, D2), PNotIn(C2, D3), PIn(C1, D3), PNotIn(R1, D1), PIn(C1, R1), PRobot(R1), PNotIn(C2, R1), PIn(C2, D1), PickUp(R1, C2, D1), PickUp(R1, C2, D3), PickUp(R1, C2, D2), PickUp(R1, C3, D1), PickUp(R1, C3, D3), PickUp(R1, C3, D2), PickUp(R1, C1, D1), PickUp(R1, C1, D3), PickUp(R1, C1, D2), PutDown(R1, C2, D1), PutDown(R1, C2, D3), PutDown(R1, C2, D2), PutDown(R1, C3, D1), PutDown(R1, C3, D3), PutDown(R1, C3, D2), PutDown(R1, C1, D1), PutDown(R1, C1, D3), PutDown(R1, C1, D2), Move(R1, D1, D3), Move(R1, D1, D2), Move(R1, D3, D1), Move(R1, D3, D2), Move(R1, D2, D1), Move(R1, D2, D3)}
-  Mutex: {{PIn(C2, D1), PNotIn(C2, D1)}, {PIn(C2, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D1), PNotIn(C2, D1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D1)}, {PNotIn(C1, R1), PIn(C1, R1)}, {PutDown(R1, C1, D1), PIn(C1, R1)}, {PutDown(R1, C1, D3), PIn(C1, R1)}, {PutDown(R1, C1, D2), PIn(C1, R1)}, {PNotIn(C1, R1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D2)}, {PNotIn(C1, D3), PIn(C1, D3)}, {PickUp(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D3), PNotIn(C1, D3)}, {PHolding(R1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PHolding(R1)}, {PutDown(R1, C2, D3), PHolding(R1)}, {PHolding(R1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PHolding(R1)}, {PutDown(R1, C3, D3), PHolding(R1)}, {PutDown(R1, C3, D2), PHolding(R1)}, {PutDown(R1, C1, D1), PHolding(R1)}, {PutDown(R1, C1, D3), PHolding(R1)}, {PutDown(R1, C1, D2), PHolding(R1)}, {PickUp(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C2, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D2)}, {PickUp(R1, C2, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D2)}, {PickUp(R1, C2, D2), PutDown(R1, C3, D3)}, {PickUp(R1, C2, D2), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D2)}, {PickUp(R1, C3, D1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D1)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C3, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C3, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C3, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C3, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D3)}, {PickUp(R1, C3, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C3, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C3, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C3, D3)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D3)}, {PickUp(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D2), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D2)}, {PIn(C1, D1), PNotIn(C1, D1)}, {PIn(C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PNotIn(C1, D1)}, {PNotIn(C2, D2), PIn(C2, D2)}, {PickUp(R1, C2, D2), PIn(C2, D2)}, {PNotIn(C2, D2), PutDown(R1, C2, D2)}, {PNotIn(R1, D3), PIn(R1, D3)}, {Move(R1, D3, D1), PIn(R1, D3)}, {Move(R1, D3, D2), PIn(R1, D3)}, {PNotIn(R1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D3), Move(R1, D3, D1)}, {Move(R1, D1, D3), Move(R1, D3, D2)}, {PNotIn(R1, D3), Move(R1, D2, D3)}, {Move(R1, D3, D1), Move(R1, D2, D3)}, {Move(R1, D3, D2), Move(R1, D2, D3)}, {PNotIn(C3, D2), PIn(C3, D2)}, {PickUp(R1, C3, D2), PIn(C3, D2)}, {PutDown(R1, C3, D2), PNotIn(C3, D2)}, {PNotIn(R1, D2), PIn(R1, D2)}, {Move(R1, D2, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), PNotIn(R1, D2)}, {Move(R1, D1, D2), Move(R1, D2, D1)}, {Move(R1, D1, D2), Move(R1, D2, D3)}, {PNotIn(R1, D2), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D3, D2)}, {PNotIn(C3, R1), PIn(C3, R1)}, {PutDown(R1, C3, D1), PIn(C3, R1)}, {PutDown(R1, C3, D3), PIn(C3, R1)}, {PutDown(R1, C3, D2), PIn(C3, R1)}, {PNotIn(C3, R1), PickUp(R1, C3, D1)}, {PNotIn(C3, R1), PickUp(R1, C3, D3)}, {PNotIn(C3, R1), PickUp(R1, C3, D2)}, {PNotIn(C1, D2), PIn(C1, D2)}, {PickUp(R1, C1, D2), PIn(C1, D2)}, {PNotIn(C1, D2), PutDown(R1, C1, D2)}, {PNotIn(C2, D3), PIn(C2, D3)}, {PickUp(R1, C2, D3), PIn(C2, D3)}, {PNotIn(C2, D3), PutDown(R1, C2, D3)}, {PNotIn(R1, D1), PIn(R1, D1)}, {Move(R1, D1, D3), PIn(R1, D1)}, {Move(R1, D1, D2), PIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D3, D1)}, {Move(R1, D1, D2), Move(R1, D3, D1)}, {PNotIn(R1, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), Move(R1, D2, D1)}, {PNotIn(C2, R1), PIn(C2, R1)}, {PutDown(R1, C2, D1), PIn(C2, R1)}, {PutDown(R1, C2, D3), PIn(C2, R1)}, {PutDown(R1, C2, D2), PIn(C2, R1)}, {PickUp(R1, C2, D1), PNotIn(C2, R1)}, {PickUp(R1, C2, D3), PNotIn(C2, R1)}, {PickUp(R1, C2, D2), PNotIn(C2, R1)}, {PIn(C3, D1), PickUp(R1, C3, D1)}, {PIn(C3, D3), PickUp(R1, C3, D3)}, {PickUp(R1, C2, D1), PNotIn(C2, D1)}, {PIn(C2, R1), PIn(C2, D3)}, {PIn(C2, R1), PIn(C2, D2)}, {PIn(C2, D1), PIn(C2, R1)}, {PickUp(R1, C2, D1), PIn(C2, R1)}, {PickUp(R1, C2, D3), PIn(C2, R1)}, {PickUp(R1, C2, D2), PIn(C2, R1)}, {PutDown(R1, C1, D1), PNotIn(C1, R1)}, {PutDown(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D2), PNotIn(C1, R1)}, {PIn(R1, D2), PIn(R1, D1)}, {PIn(R1, D3), PIn(R1, D1)}, {PIn(C3, D3), PIn(R1, D1)}, {PNotIn(C2, D2), PIn(R1, D1)}, {PNotIn(C2, D3), PIn(R1, D1)}, {PickUp(R1, C2, D3), PIn(R1, D1)}, {PickUp(R1, C2, D2), PIn(R1, D1)}, {PickUp(R1, C3, D3), PIn(R1, D1)}, {PickUp(R1, C3, D2), PIn(R1, D1)}, {PickUp(R1, C1, D3), PIn(R1, D1)}, {PickUp(R1, C1, D2), PIn(R1, D1)}, {PutDown(R1, C2, D3), PIn(R1, D1)}, {PutDown(R1, C2, D2), PIn(R1, D1)}, {PutDown(R1, C3, D3), PIn(R1, D1)}, {PutDown(R1, C3, D2), PIn(R1, D1)}, {PutDown(R1, C1, D3), PIn(R1, D1)}, {PutDown(R1, C1, D2), PIn(R1, D1)}, {Move(R1, D3, D1), PIn(R1, D1)}, {Move(R1, D3, D2), PIn(R1, D1)}, {Move(R1, D2, D1), PIn(R1, D1)}, {Move(R1, D2, D3), PIn(R1, D1)}, {PNotIn(C1, D3), PIn(C2, D3)}, {PNotIn(C1, D3), PIn(C3, D1)}, {PNotIn(C1, D3), PIn(C2, D2)}, {PIn(C3, D3), PNotIn(C1, D3)}, {PNotIn(C2, D2), PNotIn(C1, D3)}, {PNotIn(C3, R1), PNotIn(C1, D3)}, {PNotIn(C1, D2), PNotIn(C1, D3)}, {PNotIn(C2, D3), PNotIn(C1, D3)}, {PNotIn(C1, D3), PNotIn(C2, R1)}, {PNotIn(C1, D3), PickUp(R1, C2, D3)}, {PNotIn(C1, D3), PickUp(R1, C2, D2)}, {PNotIn(C1, D3), PickUp(R1, C3, D1)}, {PNotIn(C1, D3), PickUp(R1, C3, D3)}, {PNotIn(C1, D3), PickUp(R1, C1, D3)}, {PIn(C3, D1), PIn(C2, D3)}, {PHolding(R1), PIn(C2, D3)}, {PIn(C2, D2), PIn(C2, D3)}, {PIn(C3, D3), PIn(C2, D3)}, {PNotIn(C2, D2), PIn(C2, D3)}, {PNotIn(C3, R1), PIn(C2, D3)}, {PNotIn(C1, D2), PIn(C2, D3)}, {PIn(C2, D1), PIn(C2, D3)}, {PickUp(R1, C2, D1), PIn(C2, D3)}, {PickUp(R1, C2, D2), PIn(C2, D3)}, {PickUp(R1, C3, D1), PIn(C2, D3)}, {PickUp(R1, C3, D3), PIn(C2, D3)}, {PutDown(R1, C2, D1), PIn(C2, D3)}, {PutDown(R1, C2, D3), PIn(C2, D3)}, {PutDown(R1, C2, D2), PIn(C2, D3)}, {PutDown(R1, C3, D1), PIn(C2, D3)}, {PutDown(R1, C3, D3), PIn(C2, D3)}, {PutDown(R1, C3, D2), PIn(C2, D3)}, {PutDown(R1, C1, D1), PIn(C2, D3)}, {PutDown(R1, C1, D3), PIn(C2, D3)}, {PutDown(R1, C1, D2), PIn(C2, D3)}, {PIn(C1, D1), PIn(C1, D2)}, {PNotIn(C2, D2), PIn(C1, D2)}, {PNotIn(C2, D3), PIn(C1, D2)}, {PIn(C1, D3), PIn(C1, D2)}, {PIn(C1, R1), PIn(C1, D2)}, {PickUp(R1, C1, D1), PIn(C1, D2)}, {PickUp(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C1, D1), PIn(C1, D2)}, {PutDown(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C1, D2), PIn(C1, D2)}, {PIn(C3, D1), PIn(C3, R1)}, {PIn(C3, D1), PIn(R1, D2)}, {PIn(C3, D1), PHolding(R1)}, {PIn(C3, D1), PIn(C3, D2)}, {PIn(C3, D1), PIn(C2, D2)}, {PIn(C3, D1), PIn(R1, D3)}, {PIn(C3, D3), PIn(C3, D1)}, {PNotIn(C2, D2), PIn(C3, D1)}, {PNotIn(C2, D3), PIn(C3, D1)}, {PNotIn(R1, D1), PIn(C3, D1)}, {PIn(C3, D1), PNotIn(C2, R1)}, {PIn(C3, D1), PickUp(R1, C2, D3)}, {PIn(C3, D1), PickUp(R1, C2, D2)}, {PIn(C3, D1), PickUp(R1, C3, D3)}, {PIn(C3, D1), PickUp(R1, C3, D2)}, {PIn(C3, D1), PickUp(R1, C1, D3)}, {PIn(C3, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D1), PIn(C3, D1)}, {PutDown(R1, C2, D3), PIn(C3, D1)}, {PIn(C3, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PIn(C3, D1)}, {PIn(C3, D1), PutDown(R1, C3, D3)}, {PIn(C3, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PIn(C3, D1)}, {PutDown(R1, C1, D3), PIn(C3, D1)}, {PutDown(R1, C1, D2), PIn(C3, D1)}, {PIn(C3, D1), Move(R1, D3, D1)}, {PIn(C3, D1), Move(R1, D3, D2)}, {PIn(C3, D1), Move(R1, D2, D1)}, {PIn(C3, D1), Move(R1, D2, D3)}, {PIn(C3, R1), PIn(C3, D2)}, {PIn(C3, D3), PIn(C3, R1)}, {PNotIn(C2, D3), PIn(C3, R1)}, {PickUp(R1, C3, D1), PIn(C3, R1)}, {PickUp(R1, C3, D3), PIn(C3, R1)}, {PickUp(R1, C3, D2), PIn(C3, R1)}, {PIn(R1, D2), PIn(R1, D3)}, {PIn(C3, D3), PIn(R1, D2)}, {PNotIn(C2, D3), PIn(R1, D2)}, {PickUp(R1, C2, D1), PIn(R1, D2)}, {PickUp(R1, C2, D3), PIn(R1, D2)}, {PickUp(R1, C3, D1), PIn(R1, D2)}, {PickUp(R1, C3, D3), PIn(R1, D2)}, {PickUp(R1, C1, D1), PIn(R1, D2)}, {PickUp(R1, C1, D3), PIn(R1, D2)}, {PutDown(R1, C2, D1), PIn(R1, D2)}, {PutDown(R1, C2, D3), PIn(R1, D2)}, {PutDown(R1, C3, D1), PIn(R1, D2)}, {PutDown(R1, C3, D3), PIn(R1, D2)}, {PutDown(R1, C1, D1), PIn(R1, D2)}, {PutDown(R1, C1, D3), PIn(R1, D2)}, {Move(R1, D1, D3), PIn(R1, D2)}, {Move(R1, D1, D2), PIn(R1, D2)}, {Move(R1, D3, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D3, D2)}, {PIn(C3, D3), PHolding(R1)}, {PHolding(R1), PickUp(R1, C2, D1)}, {PHolding(R1), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D2), PHolding(R1)}, {PHolding(R1), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D3), PHolding(R1)}, {PHolding(R1), PickUp(R1, C3, D2)}, {PHolding(R1), PickUp(R1, C1, D1)}, {PHolding(R1), PickUp(R1, C1, D3)}, {PHolding(R1), PickUp(R1, C1, D2)}, {PIn(C3, D3), PIn(C3, D2)}, {PickUp(R1, C3, D1), PIn(C3, D2)}, {PickUp(R1, C3, D3), PIn(C3, D2)}, {PutDown(R1, C3, D1), PIn(C3, D2)}, {PutDown(R1, C3, D3), PIn(C3, D2)}, {PutDown(R1, C3, D2), PIn(C3, D2)}, {PNotIn(C2, D2), PNotHolding(R1)}, {PNotIn(C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D2), PNotHolding(R1)}, {PutDown(R1, C3, D1), PNotHolding(R1)}, {PutDown(R1, C3, D3), PNotHolding(R1)}, {PutDown(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C1, D2), PNotHolding(R1)}, {PIn(C3, D3), PIn(C2, D2)}, {PNotIn(C3, R1), PIn(C2, D2)}, {PNotIn(C1, D2), PIn(C2, D2)}, {PNotIn(C2, D3), PIn(C2, D2)}, {PIn(C2, D1), PIn(C2, D2)}, {PickUp(R1, C2, D1), PIn(C2, D2)}, {PickUp(R1, C2, D3), PIn(C2, D2)}, {PickUp(R1, C3, D1), PIn(C2, D2)}, {PickUp(R1, C3, D3), PIn(C2, D2)}, {PutDown(R1, C2, D1), PIn(C2, D2)}, {PutDown(R1, C2, D3), PIn(C2, D2)}, {PutDown(R1, C2, D2), PIn(C2, D2)}, {PNotIn(C2, D2), PIn(R1, D3)}, {PickUp(R1, C2, D1), PIn(R1, D3)}, {PickUp(R1, C2, D2), PIn(R1, D3)}, {PickUp(R1, C3, D1), PIn(R1, D3)}, {PickUp(R1, C3, D2), PIn(R1, D3)}, {PickUp(R1, C1, D1), PIn(R1, D3)}, {PickUp(R1, C1, D2), PIn(R1, D3)}, {PutDown(R1, C2, D1), PIn(R1, D3)}, {PutDown(R1, C2, D2), PIn(R1, D3)}, {PutDown(R1, C3, D1), PIn(R1, D3)}, {PutDown(R1, C3, D2), PIn(R1, D3)}, {PutDown(R1, C1, D1), PIn(R1, D3)}, {PutDown(R1, C1, D2), PIn(R1, D3)}, {Move(R1, D1, D3), PIn(R1, D3)}, {Move(R1, D1, D2), PIn(R1, D3)}, {Move(R1, D2, D1), PIn(R1, D3)}, {Move(R1, D2, D3), PIn(R1, D3)}, {PIn(C3, D3), PNotIn(C2, D2)}, {PIn(C3, D3), PNotIn(R1, D3)}, {PIn(C3, D3), PNotIn(C2, D3)}, {PIn(C3, D3), PNotIn(C2, R1)}, {PIn(C3, D3), PickUp(R1, C2, D1)}, {PIn(C3, D3), PickUp(R1, C2, D3)}, {PIn(C3, D3), PickUp(R1, C2, D2)}, {PIn(C3, D3), PickUp(R1, C3, D1)}, {PIn(C3, D3), PickUp(R1, C3, D2)}, {PIn(C3, D3), PickUp(R1, C1, D1)}, {PIn(C3, D3), PickUp(R1, C1, D2)}, {PIn(C3, D3), PutDown(R1, C2, D1)}, {PIn(C3, D3), PutDown(R1, C2, D3)}, {PIn(C3, D3), PutDown(R1, C2, D2)}, {PIn(C3, D3), PutDown(R1, C3, D1)}, {PIn(C3, D3), PutDown(R1, C3, D3)}, {PIn(C3, D3), PutDown(R1, C3, D2)}, {PIn(C3, D3), PutDown(R1, C1, D1)}, {PIn(C3, D3), PutDown(R1, C1, D3)}, {PIn(C3, D3), PutDown(R1, C1, D2)}, {PIn(C3, D3), Move(R1, D1, D3)}, {PIn(C3, D3), Move(R1, D1, D2)}, {PIn(C3, D3), Move(R1, D2, D1)}, {PIn(C3, D3), Move(R1, D2, D3)}, {PickUp(R1, C1, D1), PNotIn(C1, D1)}, {PIn(C1, D1), PIn(C1, D3)}, {PIn(C1, D1), PIn(C1, R1)}, {PIn(C1, D1), PickUp(R1, C1, D3)}, {PIn(C1, D1), PickUp(R1, C1, D2)}, {PIn(C1, D1), PutDown(R1, C1, D1)}, {PIn(C1, D1), PutDown(R1, C1, D3)}, {PIn(C1, D1), PutDown(R1, C1, D2)}, {PNotIn(R1, D3), PNotIn(C2, D2)}, {PNotIn(C2, D2), PNotIn(R1, D2)}, {PNotIn(C3, R1), PNotIn(C2, D2)}, {PNotIn(C1, D2), PNotIn(C2, D2)}, {PNotIn(C2, D3), PNotIn(C2, D2)}, {PNotIn(C2, D2), PIn(C1, D3)}, {PNotIn(C2, D2), PNotIn(C2, R1)}, {PIn(C2, D1), PNotIn(C2, D2)}, {PNotIn(C2, D2), PickUp(R1, C2, D1)}, {PNotIn(C2, D2), PickUp(R1, C2, D3)}, {PNotIn(C2, D2), PickUp(R1, C2, D2)}, {PNotIn(C2, D2), PickUp(R1, C3, D1)}, {PNotIn(C2, D2), PickUp(R1, C3, D3)}, {PNotIn(C2, D2), PickUp(R1, C3, D2)}, {PNotIn(C2, D2), PickUp(R1, C1, D1)}, {PNotIn(C2, D2), PickUp(R1, C1, D3)}, {PNotIn(C2, D2), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D1), PNotIn(C2, D2)}, {PutDown(R1, C2, D3), PNotIn(C2, D2)}, {PutDown(R1, C3, D1), PNotIn(C2, D2)}, {PNotIn(C2, D2), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D1), PNotIn(C2, D2)}, {PutDown(R1, C1, D3), PNotIn(C2, D2)}, {PNotIn(C2, D2), Move(R1, D1, D3)}, {PNotIn(C2, D2), Move(R1, D1, D2)}, {PNotIn(C2, D2), Move(R1, D3, D1)}, {PNotIn(C2, D2), Move(R1, D3, D2)}, {PNotIn(R1, D3), PNotIn(C2, D3)}, {PNotIn(R1, D3), PickUp(R1, C2, D3)}, {PNotIn(R1, D3), PickUp(R1, C3, D3)}, {PNotIn(R1, D3), PickUp(R1, C1, D3)}, {PNotIn(R1, D3), PutDown(R1, C2, D3)}, {PNotIn(R1, D3), PutDown(R1, C3, D3)}, {PNotIn(R1, D3), PutDown(R1, C1, D3)}, {PNotIn(R1, D3), Move(R1, D3, D1)}, {PNotIn(R1, D3), Move(R1, D3, D2)}, {PNotIn(C2, D3), PNotIn(C3, D2)}, {PNotIn(C3, D2), PickUp(R1, C3, D2)}, {PNotIn(C2, D3), PNotIn(R1, D2)}, {PickUp(R1, C2, D2), PNotIn(R1, D2)}, {PickUp(R1, C3, D2), PNotIn(R1, D2)}, {PickUp(R1, C1, D2), PNotIn(R1, D2)}, {PutDown(R1, C2, D2), PNotIn(R1, D2)}, {PutDown(R1, C3, D2), PNotIn(R1, D2)}, {PutDown(R1, C1, D2), PNotIn(R1, D2)}, {PNotIn(R1, D2), Move(R1, D2, D1)}, {PNotIn(R1, D2), Move(R1, D2, D3)}, {PNotIn(C3, R1), PNotIn(C2, D3)}, {PNotIn(C3, R1), PNotIn(C2, R1)}, {PNotIn(C3, R1), PickUp(R1, C2, D3)}, {PNotIn(C3, R1), PickUp(R1, C2, D2)}, {PNotIn(C3, R1), PutDown(R1, C3, D1)}, {PNotIn(C3, R1), PutDown(R1, C3, D3)}, {PNotIn(C3, R1), PutDown(R1, C3, D2)}, {PNotIn(C2, D3), PNotIn(C1, D2)}, {PNotIn(C1, D2), PNotIn(C2, R1)}, {PNotIn(C1, D2), PickUp(R1, C2, D3)}, {PNotIn(C1, D2), PickUp(R1, C2, D2)}, {PNotIn(C1, D2), PickUp(R1, C1, D2)}, {PNotIn(C2, D3), PIn(C1, D3)}, {PNotIn(C2, D3), PNotIn(C2, R1)}, {PIn(C2, D1), PNotIn(C2, D3)}, {PNotIn(C2, D3), PickUp(R1, C2, D1)}, {PNotIn(C2, D3), PickUp(R1, C2, D3)}, {PNotIn(C2, D3), PickUp(R1, C2, D2)}, {PNotIn(C2, D3), PickUp(R1, C3, D1)}, {PNotIn(C2, D3), PickUp(R1, C3, D3)}, {PNotIn(C2, D3), PickUp(R1, C3, D2)}, {PNotIn(C2, D3), PickUp(R1, C1, D1)}, {PNotIn(C2, D3), PickUp(R1, C1, D3)}, {PNotIn(C2, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D1), PNotIn(C2, D3)}, {PNotIn(C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PNotIn(C2, D3)}, {PNotIn(C2, D3), PutDown(R1, C3, D3)}, {PNotIn(C2, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PNotIn(C2, D3)}, {PutDown(R1, C1, D2), PNotIn(C2, D3)}, {PNotIn(C2, D3), Move(R1, D1, D3)}, {PNotIn(C2, D3), Move(R1, D1, D2)}, {PNotIn(C2, D3), Move(R1, D2, D1)}, {PNotIn(C2, D3), Move(R1, D2, D3)}, {PIn(C1, R1), PIn(C1, D3)}, {PickUp(R1, C1, D1), PIn(C1, D3)}, {PickUp(R1, C1, D2), PIn(C1, D3)}, {PutDown(R1, C1, D1), PIn(C1, D3)}, {PutDown(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D2), PIn(C1, D3)}, {PNotIn(R1, D1), PickUp(R1, C2, D1)}, {PNotIn(R1, D1), PickUp(R1, C3, D1)}, {PNotIn(R1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D1), PNotIn(R1, D1)}, {PutDown(R1, C3, D1), PNotIn(R1, D1)}, {PutDown(R1, C1, D1), PNotIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D1, D3)}, {PNotIn(R1, D1), Move(R1, D1, D2)}, {PickUp(R1, C1, D1), PIn(C1, R1)}, {PickUp(R1, C1, D3), PIn(C1, R1)}, {PickUp(R1, C1, D2), PIn(C1, R1)}, {PickUp(R1, C3, D1), PNotIn(C2, R1)}, {PickUp(R1, C3, D3), PNotIn(C2, R1)}, {PutDown(R1, C2, D1), PNotIn(C2, R1)}, {PutDown(R1, C2, D3), PNotIn(C2, R1)}, {PutDown(R1, C2, D2), PNotIn(C2, R1)}, {PIn(C2, D1), PickUp(R1, C2, D3)}, {PIn(C2, D1), PickUp(R1, C2, D2)}, {PIn(C2, D1), PutDown(R1, C2, D1)}, {PIn(C2, D1), PutDown(R1, C2, D3)}, {PIn(C2, D1), PutDown(R1, C2, D2)}, {PickUp(R1, C2, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D2)}, {PickUp(R1, C2, D1), Move(R1, D2, D1)}, {PickUp(R1, C2, D1), Move(R1, D2, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D3), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C2, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C2, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C2, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), Move(R1, D2, D1)}, {PickUp(R1, C2, D3), Move(R1, D2, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C2, D2), PickUp(R1, C3, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C2, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C2, D2), Move(R1, D1, D3)}, {PickUp(R1, C2, D2), Move(R1, D1, D2)}, {PickUp(R1, C2, D2), Move(R1, D3, D1)}, {PickUp(R1, C2, D2), Move(R1, D3, D2)}, {PickUp(R1, C3, D3), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C1, D3), PickUp(R1, C3, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D1), Move(R1, D3, D1)}, {PickUp(R1, C3, D1), Move(R1, D3, D2)}, {PickUp(R1, C3, D1), Move(R1, D2, D1)}, {PickUp(R1, C3, D1), Move(R1, D2, D3)}, {PickUp(R1, C3, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C3, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C3, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C3, D3)}, {PickUp(R1, C3, D3), Move(R1, D2, D1)}, {PickUp(R1, C3, D3), Move(R1, D2, D3)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C3, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), Move(R1, D3, D1)}, {PickUp(R1, C3, D2), Move(R1, D3, D2)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D2)}, {PickUp(R1, C1, D1), Move(R1, D2, D1)}, {PickUp(R1, C1, D1), Move(R1, D2, D3)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), Move(R1, D2, D1)}, {PickUp(R1, C1, D3), Move(R1, D2, D3)}, {PickUp(R1, C1, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), Move(R1, D3, D1)}, {PickUp(R1, C1, D2), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C2, D1), Move(R1, D3, D1)}, {PutDown(R1, C2, D1), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), Move(R1, D2, D1)}, {PutDown(R1, C2, D1), Move(R1, D2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D2)}, {PutDown(R1, C2, D3), Move(R1, D2, D1)}, {PutDown(R1, C2, D3), Move(R1, D2, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D3, D1)}, {PutDown(R1, C2, D2), Move(R1, D3, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C3, D1), Move(R1, D3, D1)}, {PutDown(R1, C3, D1), Move(R1, D3, D2)}, {PutDown(R1, C3, D1), Move(R1, D2, D1)}, {PutDown(R1, C3, D1), Move(R1, D2, D3)}, {PutDown(R1, C3, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D3), Move(R1, D1, D3)}, {PutDown(R1, C3, D3), Move(R1, D1, D2)}, {PutDown(R1, C3, D3), Move(R1, D2, D1)}, {PutDown(R1, C3, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D2), Move(R1, D1, D3)}, {PutDown(R1, C3, D2), Move(R1, D1, D2)}, {PutDown(R1, C3, D2), Move(R1, D3, D1)}, {PutDown(R1, C3, D2), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C1, D1), Move(R1, D3, D1)}, {PutDown(R1, C1, D1), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), Move(R1, D2, D1)}, {PutDown(R1, C1, D1), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D2, D1)}, {PutDown(R1, C1, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D2)}, {PutDown(R1, C1, D2), Move(R1, D3, D1)}, {PutDown(R1, C1, D2), Move(R1, D3, D2)}, {Move(R1, D1, D3), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D3, D2)}, {Move(R1, D3, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), PickUp(R1, C2, D1)}, {Move(R1, D1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D3), Move(R1, D3, D1)}, {PickUp(R1, C2, D3), Move(R1, D3, D2)}, {PickUp(R1, C2, D2), Move(R1, D2, D1)}, {PickUp(R1, C2, D2), Move(R1, D2, D3)}, {Move(R1, D1, D3), PickUp(R1, C3, D1)}, {Move(R1, D1, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D3), Move(R1, D3, D1)}, {PickUp(R1, C3, D3), Move(R1, D3, D2)}, {PickUp(R1, C3, D2), Move(R1, D2, D1)}, {PickUp(R1, C3, D2), Move(R1, D2, D3)}, {Move(R1, D1, D3), PickUp(R1, C1, D1)}, {Move(R1, D1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D2)}, {PickUp(R1, C1, D2), Move(R1, D2, D1)}, {PickUp(R1, C1, D2), Move(R1, D2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D1)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C2, D1), Move(R1, D1, D3)}, {PutDown(R1, C2, D1), Move(R1, D1, D2)}, {PutDown(R1, C2, D3), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D3, D1)}, {PutDown(R1, C2, D3), Move(R1, D3, D2)}, {PutDown(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D2, D1)}, {PutDown(R1, C2, D2), Move(R1, D2, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C3, D1), Move(R1, D1, D3)}, {PutDown(R1, C3, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D3), Move(R1, D3, D1)}, {PutDown(R1, C3, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D2), Move(R1, D2, D1)}, {PutDown(R1, C3, D2), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D3, D1)}, {PutDown(R1, C1, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), Move(R1, D2, D1)}, {PutDown(R1, C1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D1, D3)}, {Move(R1, D3, D1), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D2, D3)}}
-  Next state mutexes: {[{In(C2, D1), NotIn(C2, D1)}, {In(C2, D3), In(C2, R1)}, {In(C2, D2), In(C2, R1)}, {NotIn(C2, R1), In(C2, R1)}, {In(C2, D1), In(C2, R1)}, {NotIn(C1, R1), In(C1, R1)}, {In(R1, D1), In(R1, D2)}, {In(R1, D1), In(R1, D3)}, {NotIn(R1, D1), In(R1, D1)}, {In(R1, D1), NotIn(C3, D3)}, {NotIn(C2, D2), NotIn(C1, D3)}, {NotIn(C1, D3), In(C1, D3)}, {NotIn(C1, D3), NotIn(C3, D1)}, {In(C3, D1), In(C2, D3)}, {In(C2, D3), In(C2, D2)}, {In(C3, D3), In(C2, D3)}, {NotIn(C2, D2), In(C2, D3)}, {NotIn(C2, D3), In(C2, D3)}, {In(C2, D1), In(C2, D3)}, {NotIn(C3, D1), In(C2, D3)}, {In(C2, D3), NotIn(C3, D3)}, {In(C1, D1), In(C1, D2)}, {NotIn(C1, D2), In(C1, D2)}, {In(C1, D3), In(C1, D2)}, {In(C1, D2), In(C1, R1)}, {In(C3, D1), In(C3, R1)}, {In(C3, D1), In(C3, D2)}, {In(C3, D1), In(C3, D3)}, {NotIn(C2, D2), In(C3, D1)}, {In(C3, D1), NotIn(C2, D3)}, {In(C3, D1), NotIn(C3, D1)}, {In(C3, D1), NotIn(C3, D3)}, {In(C3, D2), In(C3, R1)}, {In(C3, D3), In(C3, R1)}, {NotIn(C3, R1), In(C3, R1)}, {In(R1, D3), In(R1, D2)}, {NotIn(R1, D2), In(R1, D2)}, {NotIn(C3, D1), In(R1, D2)}, {NotIn(C3, D3), In(R1, D2)}, {NotHolding(R1), Holding(R1)}, {In(C3, D3), In(C3, D2)}, {NotIn(C3, D2), In(C3, D2)}, {NotIn(C3, D1), In(C3, D2)}, {In(C3, D2), NotIn(C3, D3)}, {NotHolding(R1), NotIn(C3, D1)}, {NotHolding(R1), NotIn(C3, D3)}, {NotIn(C2, D2), In(C2, D2)}, {NotIn(C2, D3), In(C2, D2)}, {In(C2, D1), In(C2, D2)}, {NotIn(C3, D1), In(C2, D2)}, {In(C2, D2), NotIn(C3, D3)}, {NotIn(R1, D3), In(R1, D3)}, {NotIn(C3, D1), In(R1, D3)}, {NotIn(C2, D2), In(C3, D3)}, {NotIn(C2, D3), In(C3, D3)}, {In(C3, D3), NotIn(C3, D1)}, {In(C3, D3), NotIn(C3, D3)}, {In(C1, D1), NotIn(C1, D1)}, {In(C1, D1), In(C1, D3)}, {In(C1, D1), In(C1, R1)}, {NotIn(C2, D2), NotIn(C2, D3)}, {NotIn(C2, R1), NotIn(C2, D2)}, {NotIn(C2, D2), In(C2, D1)}, {NotIn(C2, D2), NotIn(C3, D1)}, {NotIn(C2, D2), NotIn(C3, D3)}, {NotIn(R1, D3), NotIn(C3, D3)}, {NotIn(C2, D3), NotIn(C3, R1)}, {NotIn(C3, R1), NotIn(C3, D1)}, {NotIn(C3, R1), NotIn(C3, D3)}, {NotIn(C1, D2), NotIn(C2, D3)}, {NotIn(C2, R1), NotIn(C2, D3)}, {NotIn(C2, D3), In(C2, D1)}, {NotIn(C2, D3), NotIn(C3, D1)}, {NotIn(C2, D3), NotIn(C3, D3)}, {In(C1, D3), In(C1, R1)}, {NotIn(R1, D1), NotIn(C3, D1)}, {NotIn(C2, R1), NotIn(C3, D1)}, {NotIn(C2, R1), NotIn(C3, D3)}, {NotIn(C3, D1), NotIn(C3, D3)}]}
-
-Level 6:
-<Level>
-  Current State: {Place(D3), NotIn(C2, D1), In(C2, R1), NotIn(C1, R1), In(R1, D1), Place(D2), In(C3, D1), NotIn(C1, D3), Container(C3), In(C2, D3), Container(C1), In(C1, D2), In(C3, R1), In(R1, D2), Holding(R1), In(C3, D2), NotHolding(R1), In(C2, D2), In(R1, D3), In(C3, D3), NotIn(C1, D1), Container(C2), In(C1, D1), NotIn(C3, D3), NotIn(C2, D2), NotIn(R1, D3), NotIn(C3, D2), NotIn(R1, D2), NotIn(C3, R1), Place(D1), NotIn(C1, D2), NotIn(C2, D3), In(C1, D3), NotIn(C3, D1), NotIn(R1, D1), In(C1, R1), Robot(R1), NotIn(C2, R1), In(C2, D1)}
-  Actions: {PPlace(D3), PNotIn(C2, D1), PIn(C2, R1), PNotIn(C1, R1), PIn(R1, D1), PPlace(D2), PIn(C3, D1), PNotIn(C1, D3), PContainer(C3), PIn(C2, D3), PContainer(C1), PIn(C1, D2), PIn(C3, R1), PIn(R1, D2), PHolding(R1), PIn(C3, D2), PNotHolding(R1), PIn(C2, D2), PIn(R1, D3), PIn(C3, D3), PNotIn(C1, D1), PContainer(C2), PIn(C1, D1), PNotIn(C3, D3), PNotIn(C2, D2), PNotIn(R1, D3), PNotIn(C3, D2), PNotIn(R1, D2), PNotIn(C3, R1), PPlace(D1), PNotIn(C1, D2), PNotIn(C2, D3), PIn(C1, D3), PNotIn(C3, D1), PNotIn(R1, D1), PIn(C1, R1), PRobot(R1), PNotIn(C2, R1), PIn(C2, D1), PickUp(R1, C2, D1), PickUp(R1, C2, D3), PickUp(R1, C2, D2), PickUp(R1, C3, D1), PickUp(R1, C3, D3), PickUp(R1, C3, D2), PickUp(R1, C1, D1), PickUp(R1, C1, D3), PickUp(R1, C1, D2), PutDown(R1, C2, D1), PutDown(R1, C2, D3), PutDown(R1, C2, D2), PutDown(R1, C3, D1), PutDown(R1, C3, D3), PutDown(R1, C3, D2), PutDown(R1, C1, D1), PutDown(R1, C1, D3), PutDown(R1, C1, D2), Move(R1, D1, D3), Move(R1, D1, D2), Move(R1, D3, D1), Move(R1, D3, D2), Move(R1, D2, D1), Move(R1, D2, D3)}
-  Mutex: {{PIn(C2, D1), PNotIn(C2, D1)}, {PIn(C2, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D1), PNotIn(C2, D1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D1)}, {PNotIn(C1, R1), PIn(C1, R1)}, {PutDown(R1, C1, D1), PIn(C1, R1)}, {PutDown(R1, C1, D3), PIn(C1, R1)}, {PutDown(R1, C1, D2), PIn(C1, R1)}, {PNotIn(C1, R1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotIn(C1, R1)}, {PutDown(R1, C1, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C1, D2)}, {PNotIn(C1, D3), PIn(C1, D3)}, {PickUp(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D3), PNotIn(C1, D3)}, {PHolding(R1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PHolding(R1)}, {PutDown(R1, C2, D3), PHolding(R1)}, {PHolding(R1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PHolding(R1)}, {PutDown(R1, C3, D3), PHolding(R1)}, {PutDown(R1, C3, D2), PHolding(R1)}, {PutDown(R1, C1, D1), PHolding(R1)}, {PutDown(R1, C1, D3), PHolding(R1)}, {PutDown(R1, C1, D2), PHolding(R1)}, {PickUp(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C2, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C2, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C2, D2)}, {PickUp(R1, C2, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C2, D2)}, {PickUp(R1, C2, D2), PutDown(R1, C3, D3)}, {PickUp(R1, C2, D2), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PickUp(R1, C2, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C2, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C2, D2)}, {PickUp(R1, C3, D1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D1)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C3, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C3, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C3, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C3, D1)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D1)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D1)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D3)}, {PickUp(R1, C3, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C3, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C3, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C3, D3)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D3)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D3)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D3)}, {PickUp(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D2), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D3), PickUp(R1, C3, D2)}, {PutDown(R1, C1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D2), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D1)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D3)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D3)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D2), PNotHolding(R1)}, {PutDown(R1, C2, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C2, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D3), PickUp(R1, C1, D2)}, {PutDown(R1, C3, D2), PickUp(R1, C1, D2)}, {PIn(C1, D1), PNotIn(C1, D1)}, {PIn(C1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C1, D1), PNotIn(C1, D1)}, {PIn(C3, D3), PNotIn(C3, D3)}, {PIn(C3, D3), PickUp(R1, C3, D3)}, {PutDown(R1, C3, D3), PNotIn(C3, D3)}, {PNotIn(C2, D2), PIn(C2, D2)}, {PickUp(R1, C2, D2), PIn(C2, D2)}, {PNotIn(C2, D2), PutDown(R1, C2, D2)}, {PNotIn(R1, D3), PIn(R1, D3)}, {Move(R1, D3, D1), PIn(R1, D3)}, {Move(R1, D3, D2), PIn(R1, D3)}, {PNotIn(R1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D3), Move(R1, D3, D1)}, {Move(R1, D1, D3), Move(R1, D3, D2)}, {PNotIn(R1, D3), Move(R1, D2, D3)}, {Move(R1, D3, D1), Move(R1, D2, D3)}, {Move(R1, D3, D2), Move(R1, D2, D3)}, {PNotIn(C3, D2), PIn(C3, D2)}, {PickUp(R1, C3, D2), PIn(C3, D2)}, {PutDown(R1, C3, D2), PNotIn(C3, D2)}, {PNotIn(R1, D2), PIn(R1, D2)}, {Move(R1, D2, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), PNotIn(R1, D2)}, {Move(R1, D1, D2), Move(R1, D2, D1)}, {Move(R1, D1, D2), Move(R1, D2, D3)}, {PNotIn(R1, D2), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D3, D2)}, {PNotIn(C3, R1), PIn(C3, R1)}, {PutDown(R1, C3, D1), PIn(C3, R1)}, {PutDown(R1, C3, D3), PIn(C3, R1)}, {PutDown(R1, C3, D2), PIn(C3, R1)}, {PNotIn(C3, R1), PickUp(R1, C3, D1)}, {PNotIn(C3, R1), PickUp(R1, C3, D3)}, {PNotIn(C3, R1), PickUp(R1, C3, D2)}, {PNotIn(C1, D2), PIn(C1, D2)}, {PickUp(R1, C1, D2), PIn(C1, D2)}, {PNotIn(C1, D2), PutDown(R1, C1, D2)}, {PNotIn(C2, D3), PIn(C2, D3)}, {PickUp(R1, C2, D3), PIn(C2, D3)}, {PNotIn(C2, D3), PutDown(R1, C2, D3)}, {PIn(C3, D1), PNotIn(C3, D1)}, {PIn(C3, D1), PickUp(R1, C3, D1)}, {PutDown(R1, C3, D1), PNotIn(C3, D1)}, {PNotIn(R1, D1), PIn(R1, D1)}, {Move(R1, D1, D3), PIn(R1, D1)}, {Move(R1, D1, D2), PIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D3, D1)}, {Move(R1, D1, D2), Move(R1, D3, D1)}, {PNotIn(R1, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), Move(R1, D2, D1)}, {PNotIn(C2, R1), PIn(C2, R1)}, {PutDown(R1, C2, D1), PIn(C2, R1)}, {PutDown(R1, C2, D3), PIn(C2, R1)}, {PutDown(R1, C2, D2), PIn(C2, R1)}, {PickUp(R1, C2, D1), PNotIn(C2, R1)}, {PickUp(R1, C2, D3), PNotIn(C2, R1)}, {PickUp(R1, C2, D2), PNotIn(C2, R1)}, {PickUp(R1, C2, D1), PNotIn(C2, D1)}, {PIn(C2, R1), PIn(C2, D3)}, {PIn(C2, R1), PIn(C2, D2)}, {PIn(C2, D1), PIn(C2, R1)}, {PickUp(R1, C2, D1), PIn(C2, R1)}, {PickUp(R1, C2, D3), PIn(C2, R1)}, {PickUp(R1, C2, D2), PIn(C2, R1)}, {PutDown(R1, C1, D1), PNotIn(C1, R1)}, {PutDown(R1, C1, D3), PNotIn(C1, R1)}, {PutDown(R1, C1, D2), PNotIn(C1, R1)}, {PIn(R1, D2), PIn(R1, D1)}, {PIn(R1, D3), PIn(R1, D1)}, {PNotIn(C3, D3), PIn(R1, D1)}, {PickUp(R1, C2, D3), PIn(R1, D1)}, {PickUp(R1, C2, D2), PIn(R1, D1)}, {PickUp(R1, C3, D3), PIn(R1, D1)}, {PickUp(R1, C3, D2), PIn(R1, D1)}, {PickUp(R1, C1, D3), PIn(R1, D1)}, {PickUp(R1, C1, D2), PIn(R1, D1)}, {PutDown(R1, C2, D3), PIn(R1, D1)}, {PutDown(R1, C2, D2), PIn(R1, D1)}, {PutDown(R1, C3, D3), PIn(R1, D1)}, {PutDown(R1, C3, D2), PIn(R1, D1)}, {PutDown(R1, C1, D3), PIn(R1, D1)}, {PutDown(R1, C1, D2), PIn(R1, D1)}, {Move(R1, D3, D1), PIn(R1, D1)}, {Move(R1, D3, D2), PIn(R1, D1)}, {Move(R1, D2, D1), PIn(R1, D1)}, {Move(R1, D2, D3), PIn(R1, D1)}, {PIn(C3, D1), PIn(C2, D3)}, {PIn(C3, D1), PIn(C3, R1)}, {PIn(C3, D1), PIn(C3, D2)}, {PIn(C3, D3), PIn(C3, D1)}, {PIn(C3, D1), PNotIn(C3, D3)}, {PNotIn(C2, D2), PIn(C3, D1)}, {PNotIn(C2, D3), PIn(C3, D1)}, {PIn(C3, D1), PickUp(R1, C2, D3)}, {PIn(C3, D1), PickUp(R1, C3, D3)}, {PIn(C3, D1), PickUp(R1, C3, D2)}, {PutDown(R1, C3, D1), PIn(C3, D1)}, {PIn(C3, D1), PutDown(R1, C3, D3)}, {PIn(C3, D1), PutDown(R1, C3, D2)}, {PNotIn(C2, D2), PNotIn(C1, D3)}, {PNotIn(C1, D3), PNotIn(C3, D1)}, {PNotIn(C1, D3), PickUp(R1, C1, D3)}, {PIn(C2, D2), PIn(C2, D3)}, {PIn(C3, D3), PIn(C2, D3)}, {PNotIn(C3, D3), PIn(C2, D3)}, {PNotIn(C2, D2), PIn(C2, D3)}, {PNotIn(C3, D1), PIn(C2, D3)}, {PIn(C2, D1), PIn(C2, D3)}, {PickUp(R1, C2, D1), PIn(C2, D3)}, {PickUp(R1, C2, D2), PIn(C2, D3)}, {PickUp(R1, C3, D1), PIn(C2, D3)}, {PickUp(R1, C3, D3), PIn(C2, D3)}, {PutDown(R1, C2, D1), PIn(C2, D3)}, {PutDown(R1, C2, D3), PIn(C2, D3)}, {PutDown(R1, C2, D2), PIn(C2, D3)}, {PIn(C1, D1), PIn(C1, D2)}, {PIn(C1, D3), PIn(C1, D2)}, {PIn(C1, R1), PIn(C1, D2)}, {PickUp(R1, C1, D1), PIn(C1, D2)}, {PickUp(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C1, D1), PIn(C1, D2)}, {PutDown(R1, C1, D3), PIn(C1, D2)}, {PutDown(R1, C1, D2), PIn(C1, D2)}, {PIn(C3, R1), PIn(C3, D2)}, {PIn(C3, D3), PIn(C3, R1)}, {PickUp(R1, C3, D1), PIn(C3, R1)}, {PickUp(R1, C3, D3), PIn(C3, R1)}, {PickUp(R1, C3, D2), PIn(C3, R1)}, {PIn(R1, D2), PIn(R1, D3)}, {PIn(R1, D2), PNotIn(C3, D3)}, {PIn(R1, D2), PNotIn(C3, D1)}, {PickUp(R1, C2, D1), PIn(R1, D2)}, {PickUp(R1, C2, D3), PIn(R1, D2)}, {PickUp(R1, C3, D1), PIn(R1, D2)}, {PickUp(R1, C3, D3), PIn(R1, D2)}, {PickUp(R1, C1, D1), PIn(R1, D2)}, {PickUp(R1, C1, D3), PIn(R1, D2)}, {PutDown(R1, C2, D1), PIn(R1, D2)}, {PutDown(R1, C2, D3), PIn(R1, D2)}, {PutDown(R1, C3, D1), PIn(R1, D2)}, {PutDown(R1, C3, D3), PIn(R1, D2)}, {PutDown(R1, C1, D1), PIn(R1, D2)}, {PutDown(R1, C1, D3), PIn(R1, D2)}, {Move(R1, D1, D3), PIn(R1, D2)}, {Move(R1, D1, D2), PIn(R1, D2)}, {Move(R1, D3, D1), PIn(R1, D2)}, {PIn(R1, D2), Move(R1, D3, D2)}, {PHolding(R1), PickUp(R1, C2, D1)}, {PHolding(R1), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D2), PHolding(R1)}, {PHolding(R1), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D3), PHolding(R1)}, {PHolding(R1), PickUp(R1, C3, D2)}, {PHolding(R1), PickUp(R1, C1, D1)}, {PHolding(R1), PickUp(R1, C1, D3)}, {PHolding(R1), PickUp(R1, C1, D2)}, {PIn(C3, D3), PIn(C3, D2)}, {PNotIn(C3, D3), PIn(C3, D2)}, {PNotIn(C3, D1), PIn(C3, D2)}, {PickUp(R1, C3, D1), PIn(C3, D2)}, {PickUp(R1, C3, D3), PIn(C3, D2)}, {PutDown(R1, C3, D1), PIn(C3, D2)}, {PutDown(R1, C3, D3), PIn(C3, D2)}, {PutDown(R1, C3, D2), PIn(C3, D2)}, {PNotHolding(R1), PNotIn(C3, D3)}, {PNotHolding(R1), PNotIn(C3, D1)}, {PutDown(R1, C2, D1), PNotHolding(R1)}, {PutDown(R1, C2, D3), PNotHolding(R1)}, {PutDown(R1, C2, D2), PNotHolding(R1)}, {PutDown(R1, C3, D1), PNotHolding(R1)}, {PutDown(R1, C3, D3), PNotHolding(R1)}, {PutDown(R1, C3, D2), PNotHolding(R1)}, {PutDown(R1, C1, D1), PNotHolding(R1)}, {PutDown(R1, C1, D3), PNotHolding(R1)}, {PutDown(R1, C1, D2), PNotHolding(R1)}, {PNotIn(C3, D3), PIn(C2, D2)}, {PNotIn(C2, D3), PIn(C2, D2)}, {PNotIn(C3, D1), PIn(C2, D2)}, {PIn(C2, D1), PIn(C2, D2)}, {PickUp(R1, C2, D1), PIn(C2, D2)}, {PickUp(R1, C2, D3), PIn(C2, D2)}, {PutDown(R1, C2, D1), PIn(C2, D2)}, {PutDown(R1, C2, D3), PIn(C2, D2)}, {PutDown(R1, C2, D2), PIn(C2, D2)}, {PIn(R1, D3), PNotIn(C3, D1)}, {PickUp(R1, C2, D1), PIn(R1, D3)}, {PickUp(R1, C2, D2), PIn(R1, D3)}, {PickUp(R1, C3, D1), PIn(R1, D3)}, {PickUp(R1, C3, D2), PIn(R1, D3)}, {PickUp(R1, C1, D1), PIn(R1, D3)}, {PickUp(R1, C1, D2), PIn(R1, D3)}, {PutDown(R1, C2, D1), PIn(R1, D3)}, {PutDown(R1, C2, D2), PIn(R1, D3)}, {PutDown(R1, C3, D1), PIn(R1, D3)}, {PutDown(R1, C3, D2), PIn(R1, D3)}, {PutDown(R1, C1, D1), PIn(R1, D3)}, {PutDown(R1, C1, D2), PIn(R1, D3)}, {Move(R1, D1, D3), PIn(R1, D3)}, {Move(R1, D1, D2), PIn(R1, D3)}, {Move(R1, D2, D1), PIn(R1, D3)}, {Move(R1, D2, D3), PIn(R1, D3)}, {PIn(C3, D3), PNotIn(C2, D2)}, {PIn(C3, D3), PNotIn(C2, D3)}, {PIn(C3, D3), PNotIn(C3, D1)}, {PIn(C3, D3), PickUp(R1, C2, D3)}, {PIn(C3, D3), PickUp(R1, C3, D1)}, {PIn(C3, D3), PickUp(R1, C3, D2)}, {PIn(C3, D3), PutDown(R1, C3, D1)}, {PIn(C3, D3), PutDown(R1, C3, D3)}, {PIn(C3, D3), PutDown(R1, C3, D2)}, {PickUp(R1, C1, D1), PNotIn(C1, D1)}, {PIn(C1, D1), PIn(C1, D3)}, {PIn(C1, D1), PIn(C1, R1)}, {PIn(C1, D1), PickUp(R1, C1, D3)}, {PIn(C1, D1), PickUp(R1, C1, D2)}, {PIn(C1, D1), PutDown(R1, C1, D1)}, {PIn(C1, D1), PutDown(R1, C1, D3)}, {PIn(C1, D1), PutDown(R1, C1, D2)}, {PNotIn(C2, D2), PNotIn(C3, D3)}, {PNotIn(R1, D3), PNotIn(C3, D3)}, {PNotIn(C3, R1), PNotIn(C3, D3)}, {PNotIn(C2, D3), PNotIn(C3, D3)}, {PNotIn(C3, D3), PNotIn(C3, D1)}, {PNotIn(C2, R1), PNotIn(C3, D3)}, {PickUp(R1, C2, D1), PNotIn(C3, D3)}, {PickUp(R1, C2, D3), PNotIn(C3, D3)}, {PickUp(R1, C2, D2), PNotIn(C3, D3)}, {PickUp(R1, C3, D1), PNotIn(C3, D3)}, {PickUp(R1, C3, D3), PNotIn(C3, D3)}, {PickUp(R1, C3, D2), PNotIn(C3, D3)}, {PickUp(R1, C1, D1), PNotIn(C3, D3)}, {PickUp(R1, C1, D3), PNotIn(C3, D3)}, {PickUp(R1, C1, D2), PNotIn(C3, D3)}, {PutDown(R1, C2, D1), PNotIn(C3, D3)}, {PutDown(R1, C2, D2), PNotIn(C3, D3)}, {PutDown(R1, C3, D1), PNotIn(C3, D3)}, {PutDown(R1, C3, D2), PNotIn(C3, D3)}, {PutDown(R1, C1, D1), PNotIn(C3, D3)}, {PutDown(R1, C1, D2), PNotIn(C3, D3)}, {Move(R1, D1, D3), PNotIn(C3, D3)}, {Move(R1, D1, D2), PNotIn(C3, D3)}, {Move(R1, D2, D1), PNotIn(C3, D3)}, {Move(R1, D2, D3), PNotIn(C3, D3)}, {PNotIn(C2, D3), PNotIn(C2, D2)}, {PNotIn(C2, D2), PNotIn(C3, D1)}, {PNotIn(C2, D2), PNotIn(C2, R1)}, {PIn(C2, D1), PNotIn(C2, D2)}, {PNotIn(C2, D2), PickUp(R1, C2, D1)}, {PNotIn(C2, D2), PickUp(R1, C2, D3)}, {PNotIn(C2, D2), PickUp(R1, C2, D2)}, {PNotIn(C2, D2), PickUp(R1, C3, D1)}, {PNotIn(C2, D2), PickUp(R1, C3, D3)}, {PNotIn(R1, D3), PickUp(R1, C2, D3)}, {PNotIn(R1, D3), PickUp(R1, C3, D3)}, {PNotIn(R1, D3), PickUp(R1, C1, D3)}, {PNotIn(R1, D3), PutDown(R1, C2, D3)}, {PNotIn(R1, D3), PutDown(R1, C3, D3)}, {PNotIn(R1, D3), PutDown(R1, C1, D3)}, {PNotIn(R1, D3), Move(R1, D3, D1)}, {PNotIn(R1, D3), Move(R1, D3, D2)}, {PNotIn(C3, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C2, D2), PNotIn(R1, D2)}, {PickUp(R1, C3, D2), PNotIn(R1, D2)}, {PickUp(R1, C1, D2), PNotIn(R1, D2)}, {PutDown(R1, C2, D2), PNotIn(R1, D2)}, {PutDown(R1, C3, D2), PNotIn(R1, D2)}, {PutDown(R1, C1, D2), PNotIn(R1, D2)}, {PNotIn(R1, D2), Move(R1, D2, D1)}, {PNotIn(R1, D2), Move(R1, D2, D3)}, {PNotIn(C3, R1), PNotIn(C2, D3)}, {PNotIn(C3, R1), PNotIn(C3, D1)}, {PNotIn(C3, R1), PutDown(R1, C3, D1)}, {PNotIn(C3, R1), PutDown(R1, C3, D3)}, {PNotIn(C3, R1), PutDown(R1, C3, D2)}, {PNotIn(C2, D3), PNotIn(C1, D2)}, {PNotIn(C1, D2), PickUp(R1, C1, D2)}, {PNotIn(C2, D3), PNotIn(C3, D1)}, {PNotIn(C2, D3), PNotIn(C2, R1)}, {PIn(C2, D1), PNotIn(C2, D3)}, {PNotIn(C2, D3), PickUp(R1, C2, D1)}, {PNotIn(C2, D3), PickUp(R1, C2, D3)}, {PNotIn(C2, D3), PickUp(R1, C2, D2)}, {PNotIn(C2, D3), PickUp(R1, C3, D1)}, {PNotIn(C2, D3), PickUp(R1, C3, D3)}, {PIn(C1, R1), PIn(C1, D3)}, {PickUp(R1, C1, D1), PIn(C1, D3)}, {PickUp(R1, C1, D2), PIn(C1, D3)}, {PutDown(R1, C1, D1), PIn(C1, D3)}, {PutDown(R1, C1, D3), PIn(C1, D3)}, {PutDown(R1, C1, D2), PIn(C1, D3)}, {PNotIn(R1, D1), PNotIn(C3, D1)}, {PNotIn(C2, R1), PNotIn(C3, D1)}, {PickUp(R1, C2, D1), PNotIn(C3, D1)}, {PickUp(R1, C2, D3), PNotIn(C3, D1)}, {PickUp(R1, C2, D2), PNotIn(C3, D1)}, {PickUp(R1, C3, D1), PNotIn(C3, D1)}, {PickUp(R1, C3, D3), PNotIn(C3, D1)}, {PickUp(R1, C3, D2), PNotIn(C3, D1)}, {PickUp(R1, C1, D1), PNotIn(C3, D1)}, {PickUp(R1, C1, D3), PNotIn(C3, D1)}, {PickUp(R1, C1, D2), PNotIn(C3, D1)}, {PutDown(R1, C2, D3), PNotIn(C3, D1)}, {PutDown(R1, C2, D2), PNotIn(C3, D1)}, {PutDown(R1, C3, D3), PNotIn(C3, D1)}, {PutDown(R1, C3, D2), PNotIn(C3, D1)}, {PutDown(R1, C1, D3), PNotIn(C3, D1)}, {PutDown(R1, C1, D2), PNotIn(C3, D1)}, {Move(R1, D3, D1), PNotIn(C3, D1)}, {Move(R1, D3, D2), PNotIn(C3, D1)}, {Move(R1, D2, D1), PNotIn(C3, D1)}, {Move(R1, D2, D3), PNotIn(C3, D1)}, {PNotIn(R1, D1), PickUp(R1, C2, D1)}, {PNotIn(R1, D1), PickUp(R1, C3, D1)}, {PNotIn(R1, D1), PickUp(R1, C1, D1)}, {PutDown(R1, C2, D1), PNotIn(R1, D1)}, {PutDown(R1, C3, D1), PNotIn(R1, D1)}, {PutDown(R1, C1, D1), PNotIn(R1, D1)}, {PNotIn(R1, D1), Move(R1, D1, D3)}, {PNotIn(R1, D1), Move(R1, D1, D2)}, {PickUp(R1, C1, D1), PIn(C1, R1)}, {PickUp(R1, C1, D3), PIn(C1, R1)}, {PickUp(R1, C1, D2), PIn(C1, R1)}, {PutDown(R1, C2, D1), PNotIn(C2, R1)}, {PutDown(R1, C2, D3), PNotIn(C2, R1)}, {PutDown(R1, C2, D2), PNotIn(C2, R1)}, {PIn(C2, D1), PickUp(R1, C2, D3)}, {PIn(C2, D1), PickUp(R1, C2, D2)}, {PIn(C2, D1), PutDown(R1, C2, D1)}, {PIn(C2, D1), PutDown(R1, C2, D3)}, {PIn(C2, D1), PutDown(R1, C2, D2)}, {PickUp(R1, C2, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D3), PickUp(R1, C2, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D1)}, {PickUp(R1, C2, D1), Move(R1, D3, D2)}, {PickUp(R1, C2, D1), Move(R1, D2, D1)}, {PickUp(R1, C2, D1), Move(R1, D2, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D3), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C2, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C2, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C2, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C2, D3)}, {PickUp(R1, C2, D3), Move(R1, D2, D1)}, {PickUp(R1, C2, D3), Move(R1, D2, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C2, D2), PickUp(R1, C3, D3)}, {PickUp(R1, C2, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C2, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C2, D2), Move(R1, D1, D3)}, {PickUp(R1, C2, D2), Move(R1, D1, D2)}, {PickUp(R1, C2, D2), Move(R1, D3, D1)}, {PickUp(R1, C2, D2), Move(R1, D3, D2)}, {PickUp(R1, C3, D3), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C1, D3), PickUp(R1, C3, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D1), Move(R1, D3, D1)}, {PickUp(R1, C3, D1), Move(R1, D3, D2)}, {PickUp(R1, C3, D1), Move(R1, D2, D1)}, {PickUp(R1, C3, D1), Move(R1, D2, D3)}, {PickUp(R1, C3, D3), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C3, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C3, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C3, D3)}, {PickUp(R1, C3, D3), Move(R1, D2, D1)}, {PickUp(R1, C3, D3), Move(R1, D2, D3)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C3, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C3, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C3, D2)}, {PickUp(R1, C3, D2), Move(R1, D3, D1)}, {PickUp(R1, C3, D2), Move(R1, D3, D2)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D1)}, {PickUp(R1, C1, D1), Move(R1, D3, D2)}, {PickUp(R1, C1, D1), Move(R1, D2, D1)}, {PickUp(R1, C1, D1), Move(R1, D2, D3)}, {PickUp(R1, C1, D3), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D3), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D3)}, {PickUp(R1, C1, D3), Move(R1, D2, D1)}, {PickUp(R1, C1, D3), Move(R1, D2, D3)}, {PickUp(R1, C1, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PickUp(R1, C1, D2)}, {PickUp(R1, C1, D2), Move(R1, D3, D1)}, {PickUp(R1, C1, D2), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C2, D1), Move(R1, D3, D1)}, {PutDown(R1, C2, D1), Move(R1, D3, D2)}, {PutDown(R1, C2, D1), Move(R1, D2, D1)}, {PutDown(R1, C2, D1), Move(R1, D2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D3)}, {PutDown(R1, C2, D3), Move(R1, D1, D2)}, {PutDown(R1, C2, D3), Move(R1, D2, D1)}, {PutDown(R1, C2, D3), Move(R1, D2, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C3, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D1, D3)}, {Move(R1, D1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D3, D1)}, {PutDown(R1, C2, D2), Move(R1, D3, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C3, D1), Move(R1, D3, D1)}, {PutDown(R1, C3, D1), Move(R1, D3, D2)}, {PutDown(R1, C3, D1), Move(R1, D2, D1)}, {PutDown(R1, C3, D1), Move(R1, D2, D3)}, {PutDown(R1, C3, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D3), Move(R1, D1, D3)}, {PutDown(R1, C3, D3), Move(R1, D1, D2)}, {PutDown(R1, C3, D3), Move(R1, D2, D1)}, {PutDown(R1, C3, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C3, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D2), Move(R1, D1, D3)}, {PutDown(R1, C3, D2), Move(R1, D1, D2)}, {PutDown(R1, C3, D2), Move(R1, D3, D1)}, {PutDown(R1, C3, D2), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D1), PutDown(R1, C1, D2)}, {PutDown(R1, C1, D1), Move(R1, D3, D1)}, {PutDown(R1, C1, D1), Move(R1, D3, D2)}, {PutDown(R1, C1, D1), Move(R1, D2, D1)}, {PutDown(R1, C1, D1), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), PutDown(R1, C1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D3)}, {PutDown(R1, C1, D3), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D2, D1)}, {PutDown(R1, C1, D3), Move(R1, D2, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D3)}, {PutDown(R1, C1, D2), Move(R1, D1, D2)}, {PutDown(R1, C1, D2), Move(R1, D3, D1)}, {PutDown(R1, C1, D2), Move(R1, D3, D2)}, {Move(R1, D1, D3), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D3, D2)}, {Move(R1, D3, D1), Move(R1, D2, D1)}, {Move(R1, D1, D3), PickUp(R1, C2, D1)}, {Move(R1, D1, D2), PickUp(R1, C2, D1)}, {PickUp(R1, C2, D3), Move(R1, D3, D1)}, {PickUp(R1, C2, D3), Move(R1, D3, D2)}, {PickUp(R1, C2, D2), Move(R1, D2, D1)}, {PickUp(R1, C2, D2), Move(R1, D2, D3)}, {Move(R1, D1, D3), PickUp(R1, C3, D1)}, {Move(R1, D1, D2), PickUp(R1, C3, D1)}, {PickUp(R1, C3, D3), Move(R1, D3, D1)}, {PickUp(R1, C3, D3), Move(R1, D3, D2)}, {PickUp(R1, C3, D2), Move(R1, D2, D1)}, {PickUp(R1, C3, D2), Move(R1, D2, D3)}, {Move(R1, D1, D3), PickUp(R1, C1, D1)}, {Move(R1, D1, D2), PickUp(R1, C1, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D1)}, {PickUp(R1, C1, D3), Move(R1, D3, D2)}, {PickUp(R1, C1, D2), Move(R1, D2, D1)}, {PickUp(R1, C1, D2), Move(R1, D2, D3)}, {PutDown(R1, C2, D1), PutDown(R1, C3, D1)}, {PutDown(R1, C2, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C2, D1), Move(R1, D1, D3)}, {PutDown(R1, C2, D1), Move(R1, D1, D2)}, {PutDown(R1, C2, D3), PutDown(R1, C3, D3)}, {PutDown(R1, C1, D3), PutDown(R1, C2, D3)}, {PutDown(R1, C2, D3), Move(R1, D3, D1)}, {PutDown(R1, C2, D3), Move(R1, D3, D2)}, {PutDown(R1, C3, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C1, D2), PutDown(R1, C2, D2)}, {PutDown(R1, C2, D2), Move(R1, D2, D1)}, {PutDown(R1, C2, D2), Move(R1, D2, D3)}, {PutDown(R1, C3, D1), PutDown(R1, C1, D1)}, {PutDown(R1, C3, D1), Move(R1, D1, D3)}, {PutDown(R1, C3, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), PutDown(R1, C3, D3)}, {PutDown(R1, C3, D3), Move(R1, D3, D1)}, {PutDown(R1, C3, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), PutDown(R1, C3, D2)}, {PutDown(R1, C3, D2), Move(R1, D2, D1)}, {PutDown(R1, C3, D2), Move(R1, D2, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D3)}, {PutDown(R1, C1, D1), Move(R1, D1, D2)}, {PutDown(R1, C1, D3), Move(R1, D3, D1)}, {PutDown(R1, C1, D3), Move(R1, D3, D2)}, {PutDown(R1, C1, D2), Move(R1, D2, D1)}, {PutDown(R1, C1, D2), Move(R1, D2, D3)}, {Move(R1, D1, D2), Move(R1, D1, D3)}, {Move(R1, D3, D1), Move(R1, D3, D2)}, {Move(R1, D2, D1), Move(R1, D2, D3)}}
-  Next state mutexes: {[{In(C2, D1), NotIn(C2, D1)}, {In(C2, D3), In(C2, R1)}, {In(C2, D2), In(C2, R1)}, {NotIn(C2, R1), In(C2, R1)}, {In(C2, D1), In(C2, R1)}, {NotIn(C1, R1), In(C1, R1)}, {In(R1, D1), In(R1, D2)}, {In(R1, D1), In(R1, D3)}, {NotIn(R1, D1), In(R1, D1)}, {In(C3, D1), In(C3, R1)}, {In(C3, D1), In(C3, D2)}, {In(C3, D1), In(C3, D3)}, {In(C3, D1), NotIn(C3, D3)}, {In(C3, D1), NotIn(C3, D1)}, {NotIn(C1, D3), In(C1, D3)}, {In(C2, D3), In(C2, D2)}, {NotIn(C2, D3), In(C2, D3)}, {NotIn(C3, D1), In(C2, D3)}, {In(C2, D1), In(C2, D3)}, {In(C1, D1), In(C1, D2)}, {NotIn(C1, D2), In(C1, D2)}, {In(C1, D3), In(C1, D2)}, {In(C1, D2), In(C1, R1)}, {In(C3, D2), In(C3, R1)}, {In(C3, D3), In(C3, R1)}, {NotIn(C3, R1), In(C3, R1)}, {In(R1, D3), In(R1, D2)}, {NotIn(R1, D2), In(R1, D2)}, {NotHolding(R1), Holding(R1)}, {In(C3, D3), In(C3, D2)}, {In(C3, D2), NotIn(C3, D3)}, {NotIn(C3, D2), In(C3, D2)}, {NotIn(C3, D1), In(C3, D2)}, {NotIn(C2, D2), In(C2, D2)}, {In(C2, D1), In(C2, D2)}, {NotIn(R1, D3), In(R1, D3)}, {In(C3, D3), NotIn(C3, D3)}, {In(C3, D3), NotIn(C3, D1)}, {In(C1, D1), NotIn(C1, D1)}, {In(C1, D1), In(C1, D3)}, {In(C1, D1), In(C1, R1)}, {NotIn(C2, D2), NotIn(C3, D3)}, {NotIn(C3, R1), NotIn(C3, D3)}, {NotIn(C2, D3), NotIn(C3, D3)}, {NotIn(C3, D1), NotIn(C3, D3)}, {NotIn(C2, D2), NotIn(C2, D3)}, {NotIn(C2, D2), NotIn(C3, D1)}, {NotIn(C3, R1), NotIn(C3, D1)}, {NotIn(C2, D3), NotIn(C3, D1)}, {In(C1, D3), In(C1, R1)}]}
-
-Level 7:
-<Level>
-  Current State: {Place(D3), NotIn(C2, D1), In(C2, R1), NotIn(C1, R1), In(R1, D1), Place(D2), NotIn(C1, D3), In(C3, D1), Container(C3), In(C2, D3), Container(C1), In(C1, D2), In(C3, R1), In(R1, D2), Holding(R1), In(C3, D2), NotHolding(R1), In(C2, D2), In(R1, D3), In(C3, D3), NotIn(C1, D1), Container(C2), In(C1, D1), NotIn(C3, D3), NotIn(C2, D2), NotIn(R1, D3), NotIn(C3, D2), NotIn(R1, D2), NotIn(C3, R1), Place(D1), NotIn(C1, D2), NotIn(C2, D3), In(C1, D3), NotIn(C3, D1), NotIn(R1, D1), In(C1, R1), Robot(R1), NotIn(C2, R1), In(C2, D1)}
-  Actions: {}
-  Mutex: {{In(C2, D1), NotIn(C2, D1)}, {In(C2, D3), In(C2, R1)}, {In(C2, D2), In(C2, R1)}, {NotIn(C2, R1), In(C2, R1)}, {In(C2, D1), In(C2, R1)}, {NotIn(C1, R1), In(C1, R1)}, {In(R1, D1), In(R1, D2)}, {In(R1, D1), In(R1, D3)}, {NotIn(R1, D1), In(R1, D1)}, {In(C3, D1), In(C3, R1)}, {In(C3, D1), In(C3, D2)}, {In(C3, D1), In(C3, D3)}, {In(C3, D1), NotIn(C3, D3)}, {In(C3, D1), NotIn(C3, D1)}, {NotIn(C1, D3), In(C1, D3)}, {In(C2, D3), In(C2, D2)}, {NotIn(C2, D3), In(C2, D3)}, {NotIn(C3, D1), In(C2, D3)}, {In(C2, D1), In(C2, D3)}, {In(C1, D1), In(C1, D2)}, {NotIn(C1, D2), In(C1, D2)}, {In(C1, D3), In(C1, D2)}, {In(C1, D2), In(C1, R1)}, {In(C3, D2), In(C3, R1)}, {In(C3, D3), In(C3, R1)}, {NotIn(C3, R1), In(C3, R1)}, {In(R1, D3), In(R1, D2)}, {NotIn(R1, D2), In(R1, D2)}, {NotHolding(R1), Holding(R1)}, {In(C3, D3), In(C3, D2)}, {In(C3, D2), NotIn(C3, D3)}, {NotIn(C3, D2), In(C3, D2)}, {NotIn(C3, D1), In(C3, D2)}, {NotIn(C2, D2), In(C2, D2)}, {In(C2, D1), In(C2, D2)}, {NotIn(R1, D3), In(R1, D3)}, {In(C3, D3), NotIn(C3, D3)}, {In(C3, D3), NotIn(C3, D1)}, {In(C1, D1), NotIn(C1, D1)}, {In(C1, D1), In(C1, D3)}, {In(C1, D1), In(C1, R1)}, {NotIn(C2, D2), NotIn(C3, D3)}, {NotIn(C3, R1), NotIn(C3, D3)}, {NotIn(C2, D3), NotIn(C3, D3)}, {NotIn(C3, D1), NotIn(C3, D3)}, {NotIn(C2, D2), NotIn(C2, D3)}, {NotIn(C2, D2), NotIn(C3, D1)}, {NotIn(C3, R1), NotIn(C3, D1)}, {NotIn(C2, D3), NotIn(C3, D1)}, {In(C1, D3), In(C1, R1)}}
-  Next state mutexes: {[]}
