+++ planning.py
@@ -276,420 +276,6 @@ def goal_test(goals, state):
     return all(kb.ask(q) is not False for q in goals)
 
 
-def air_cargo():
-    """
-    [Figure 10.1] AIR-CARGO-PROBLEM
-
-    An air-cargo shipment problem for delivering cargo to different locations,
-    given the starting location and airplanes.
-
-    Example:
-    >>> from planning import *
-    >>> ac = air_cargo()
-    >>> ac.goal_test()
-    False
-    >>> ac.act(expr('Load(C2, P2, JFK)'))
-    >>> ac.act(expr('Load(C1, P1, SFO)'))
-    >>> ac.act(expr('Fly(P1, SFO, JFK)'))
-    >>> ac.act(expr('Fly(P2, JFK, SFO)'))
-    >>> ac.act(expr('Unload(C2, P2, SFO)'))
-    >>> ac.goal_test()
-    False
-    >>> ac.act(expr('Unload(C1, P1, JFK)'))
-    >>> ac.goal_test()
-    True
-    >>>
-    """
-
-    return PlanningProblem(initial='At(C1, SFO) & At(C2, JFK) & At(P1, SFO) & At(P2, JFK)',
-                           goals='At(C1, JFK) & At(C2, SFO)',
-                           actions=[Action('Load(c, p, a)',
-                                           precond='At(c, a) & At(p, a)',
-                                           effect='In(c, p) & ~At(c, a)',
-                                           domain='Cargo(c) & Plane(p) & Airport(a)'),
-                                    Action('Unload(c, p, a)',
-                                           precond='In(c, p) & At(p, a)',
-                                           effect='At(c, a) & ~In(c, p)',
-                                           domain='Cargo(c) & Plane(p) & Airport(a)'),
-                                    Action('Fly(p, f, to)',
-                                           precond='At(p, f)',
-                                           effect='At(p, to) & ~At(p, f)',
-                                           domain='Plane(p) & Airport(f) & Airport(to)')],
-                           domain='Cargo(C1) & Cargo(C2) & Plane(P1) & Plane(P2) & Airport(SFO) & Airport(JFK)')
-
-
-def spare_tire():
-    """
-    [Figure 10.2] SPARE-TIRE-PROBLEM
-
-    A problem involving changing the flat tire of a car
-    with a spare tire from the trunk.
-
-    Example:
-    >>> from planning import *
-    >>> st = spare_tire()
-    >>> st.goal_test()
-    False
-    >>> st.act(expr('Remove(Spare, Trunk)'))
-    >>> st.act(expr('Remove(Flat, Axle)'))
-    >>> st.goal_test()
-    False
-    >>> st.act(expr('PutOn(Spare, Axle)'))
-    >>> st.goal_test()
-    True
-    >>>
-    """
-
-    return PlanningProblem(initial='At(Flat, Axle) & At(Spare, Trunk)',
-                           goals='At(Spare, Axle) & At(Flat, Ground)',
-                           actions=[Action('Remove(obj, loc)',
-                                           precond='At(obj, loc)',
-                                           effect='At(obj, Ground) & ~At(obj, loc)',
-                                           domain='Tire(obj)'),
-                                    Action('PutOn(t, Axle)',
-                                           precond='At(t, Ground) & ~At(Flat, Axle)',
-                                           effect='At(t, Axle) & ~At(t, Ground)',
-                                           domain='Tire(t)'),
-                                    Action('LeaveOvernight',
-                                           precond='',
-                                           effect='~At(Spare, Ground) & ~At(Spare, Axle) & ~At(Spare, Trunk) & \
-                                        ~At(Flat, Ground) & ~At(Flat, Axle) & ~At(Flat, Trunk)')],
-                           domain='Tire(Flat) & Tire(Spare)')
-
-
-def three_block_tower():
-    """
-    [Figure 10.3] THREE-BLOCK-TOWER
-
-    A blocks-world problem of stacking three blocks in a certain configuration,
-    also known as the Sussman Anomaly.
-
-    Example:
-    >>> from planning import *
-    >>> tbt = three_block_tower()
-    >>> tbt.goal_test()
-    False
-    >>> tbt.act(expr('MoveToTable(C, A)'))
-    >>> tbt.act(expr('Move(B, Table, C)'))
-    >>> tbt.goal_test()
-    False
-    >>> tbt.act(expr('Move(A, Table, B)'))
-    >>> tbt.goal_test()
-    True
-    >>>
-    """
-    return PlanningProblem(initial='On(A, Table) & On(B, Table) & On(C, A) & Clear(B) & Clear(C)',
-                           goals='On(A, B) & On(B, C)',
-                           actions=[Action('Move(b, x, y)',
-                                           precond='On(b, x) & Clear(b) & Clear(y)',
-                                           effect='On(b, y) & Clear(x) & ~On(b, x) & ~Clear(y)',
-                                           domain='Block(b) & Block(y)'),
-                                    Action('MoveToTable(b, x)',
-                                           precond='On(b, x) & Clear(b)',
-                                           effect='On(b, Table) & Clear(x) & ~On(b, x)',
-                                           domain='Block(b) & Block(x)')],
-                           domain='Block(A) & Block(B) & Block(C)')
-
-def logisticsPlanCustom(initial_state=None, goal_state=None):
-    if initial_state == None: 
-        initial_state = 'In(C1, R1) & In(C2, D1) & In(C3, D2) & In(R1, D1) & Holding(R1)'
-    if goal_state == None:
-        raise ValueError("Goal must be defined")
-
-    planning_problem = \
-    PlanningProblem(initial = initial_state,
-                    goals = goal_state,
-                    actions=[Action('PickUp(r, c, d)',
-                                    precond='In(r, d) & In (c, d) & ~Holding(r)',
-                                    effect='Holding(r) & ~In(c, d) & In(c, r)', 
-                                    domain='Robot(r) & Place(d) & Container(c)'),
-                             Action('PutDown(r, c, d)', 
-                                    precond='In(r, d) & In(c, r) & Holding(r)',
-                                    effect='~Holding(r) & ~In(c, r) & In(c, d)',
-                                    domain='Robot(r) & Place(d) & Container(c)'),
-                            Action('Move(r, d_start, d_end)',
-                                   precond='In(r,d_start)',
-                                   effect='~In(r, d_start) & In(r, d_end)',
-                                   domain='Robot(r) & Place(d_start) & Place(d_end)')],
-                domain='Container(C1) & Container(C2) & Container(C3) & Place(D1) & Place(D2) & Place(D3) & Robot(R1)')
-    
-    return planning_problem
-
-
-def simple_blocks_world():
-    """
-    SIMPLE-BLOCKS-WORLD
-
-    A simplified definition of the Sussman Anomaly problem.
-
-    Example:
-    >>> from planning import *
-    >>> sbw = simple_blocks_world()
-    >>> sbw.goal_test()
-    False
-    >>> sbw.act(expr('ToTable(A, B)'))
-    >>> sbw.act(expr('FromTable(B, A)'))
-    >>> sbw.goal_test()
-    False
-    >>> sbw.act(expr('FromTable(C, B)'))
-    >>> sbw.goal_test()
-    True
-    >>>
-    """
-
-    return PlanningProblem(initial='On(A, B) & Clear(A) & OnTable(B) & OnTable(C) & Clear(C)',
-                           goals='On(B, A) & On(C, B)',
-                           actions=[Action('ToTable(x, y)',
-                                           precond='On(x, y) & Clear(x)',
-                                           effect='~On(x, y) & Clear(y) & OnTable(x)'),
-                                    Action('FromTable(y, x)',
-                                           precond='OnTable(y) & Clear(y) & Clear(x)',
-                                           effect='~OnTable(y) & ~Clear(x) & On(y, x)')])
-
-
-def have_cake_and_eat_cake_too():
-    """
-    [Figure 10.7] CAKE-PROBLEM
-
-    A problem where we begin with a cake and want to
-    reach the state of having a cake and having eaten a cake.
-    The possible actions include baking a cake and eating a cake.
-
-    Example:
-    >>> from planning import *
-    >>> cp = have_cake_and_eat_cake_too()
-    >>> cp.goal_test()
-    False
-    >>> cp.act(expr('Eat(Cake)'))
-    >>> cp.goal_test()
-    False
-    >>> cp.act(expr('Bake(Cake)'))
-    >>> cp.goal_test()
-    True
-    >>>
-    """
-
-    return PlanningProblem(initial='Have(Cake) & ~Eaten(Cake)',
-                           goals='Have(Cake) & Eaten(Cake)',
-                           actions=[Action('Eat(Cake)',
-                                           precond='Have(Cake)',
-                                           effect='Eaten(Cake) & ~Have(Cake)'),
-                                    Action('Bake(Cake)',
-                                           precond='~Have(Cake)',
-                                           effect='Have(Cake)')])
-
-
-def shopping_problem():
-    """
-    SHOPPING-PROBLEM
-
-    A problem of acquiring some items given their availability at certain stores.
-
-    Example:
-    >>> from planning import *
-    >>> sp = shopping_problem()
-    >>> sp.goal_test()
-    False
-    >>> sp.act(expr('Go(Home, HW)'))
-    >>> sp.act(expr('Buy(Drill, HW)'))
-    >>> sp.act(expr('Go(HW, SM)'))
-    >>> sp.act(expr('Buy(Banana, SM)'))
-    >>> sp.goal_test()
-    False
-    >>> sp.act(expr('Buy(Milk, SM)'))
-    >>> sp.goal_test()
-    True
-    >>>
-    """
-
-    return PlanningProblem(initial='At(Home) & Sells(SM, Milk) & Sells(HW, Drill)',
-                           goals='Have(Milk) & Have(Drill)',
-                           actions=[Action('Buy(x, store)',
-                                           precond='At(store) & Sells(store, x)',
-                                           effect='Have(x)',
-                                           domain='Store(store) & Item(x)'),
-                                    Action('Go(x, y)',
-                                           precond='At(x)',
-                                           effect='At(y) & ~At(x)',
-                                           domain='Rplace(x) & Rplace(y)')],
-                           domain='Rplace(Home) & Rplace(SM) & Rplace(HW) & Store(SM) & Store(HW) & '
-                                  'Item(Milk) & Item(Drill)')
-
-
-def socks_and_shoes():
-    """
-    SOCKS-AND-SHOES-PROBLEM
-
-    A task of wearing socks and shoes on both feet
-
-    Example:
-    >>> from planning import *
-    >>> ss = socks_and_shoes()
-    >>> ss.goal_test()
-    False
-    >>> ss.act(expr('RightSock'))
-    >>> ss.act(expr('RightShoe'))
-    >>> ss.act(expr('LeftSock'))
-    >>> ss.goal_test()
-    False
-    >>> ss.act(expr('LeftShoe'))
-    >>> ss.goal_test()
-    True
-    >>>
-    """
-
-    return PlanningProblem(initial='',
-                           goals='RightShoeOn & LeftShoeOn',
-                           actions=[Action('RightShoe',
-                                           precond='RightSockOn',
-                                           effect='RightShoeOn'),
-                                    Action('RightSock',
-                                           precond='',
-                                           effect='RightSockOn'),
-                                    Action('LeftShoe',
-                                           precond='LeftSockOn',
-                                           effect='LeftShoeOn'),
-                                    Action('LeftSock',
-                                           precond='',
-                                           effect='LeftSockOn')])
-
-
-def double_tennis_problem_simple():
-    return PlanningProblem(
-        initial='At(A, LeftNet) & At(B, RightNet) & Approaching(ball, RightBaseline)',
-        goals='At(A, LeftBaseline) & Returned(ball)',
-        actions=[Action('Hit(actor, ball, loc)',
-                        precond='Approaching(ball, loc) & At(actor, loc)',
-                        effect='Returned(ball)'),
-                 Action('Go(actor, to, loc)',
-                        precond='At(actor, loc)',
-                        effect='At(actor, to) & ~At(actor, loc)')],
-        domain="Loc(LeftBaseline)")
-
-def double_tennis_problem_simple2():
-    return PlanningProblem(
-        initial='At(A, LeftNet) & At(B, LeftNet) & Approaching(ball, RightNet)',
-        goals='At(A, LeftNet) & Returned(ball) & At(B, LeftNet)',
-        actions=[Action('Hit(actor, ball, loc)',
-                        precond='Approaching(ball, loc) & At(actor, loc)',
-                        effect='Returned(ball)'),
-                 Action('Go(actor, to, loc)',
-                        precond='At(actor, loc)',
-                        effect='At(actor, to) & ~At(actor, loc)')])
-
-def double_tennis_problem_simple3():
-    return PlanningProblem(
-        initial='At(A, LeftNet) & Approaching(ball, RightNet)',
-        goals='At(A, LeftNet) & Returned(ball)',
-        actions=[Action('Hit(actor, ball, loc)',
-                        precond='Approaching(ball, loc) & At(actor, loc)',
-                        effect='Returned(ball)'),
-                 Action('Go(actor, to, loc)',
-                        precond='At(actor, loc)',
-                        effect='At(actor, to) & ~At(actor, loc)')])
-
-
-def double_tennis_problem():
-    """
-    [Figure 11.10] DOUBLE-TENNIS-PROBLEM
-
-    A multiagent planning problem involving two partner tennis players
-    trying to return an approaching ball and repositioning around in the court.
-
-    Example:
-    >>> from planning import *
-    >>> dtp = double_tennis_problem()
-    >>> goal_test(dtp.goals, dtp.initial)
-    False
-    >>> dtp.act(expr('Go(A, RightBaseLine, LeftBaseLine)'))
-    >>> dtp.act(expr('Hit(A, Ball, RightBaseLine)'))
-    >>> goal_test(dtp.goals, dtp.initial)
-    False
-    >>> dtp.act(expr('Go(A, LeftNet, RightBaseLine)'))
-    >>> goal_test(dtp.goals, dtp.initial)
-    True
-    >>>
-    """
-
-    # The authors mistakenly used `a` rather than `A`, and also set it to two locations?
-    # This is the original function, but it's faulty. Although, it does technically pass pytests? Just for some reason, not graphPlan...
-    # I believe this fails GraphPlan because of the way it extracts objects, which does not include objects that only exist in goal states?
-    #  - Therefore the GraphPlan fails, but the pytests that manuallys solve pass?
-    #  - To explain why At(a, LeftNet), At(a,RightNet) passes, perhaps since it doesn't see `a` in objects, it takes it as a variable? Where any object can make it true? Which A and B do?
-    """
-    return PlanningProblem(
-        initial='At(A, LeftBaseLine) & At(B, RightNet) & Approaching(Ball, RightBaseLine) & Partner(A, B) & Partner(B, A)',
-        goals='Returned(Ball) & At(a, LeftNet) & At(a, RightNet)',
-        actions=[Action('Hit(actor, Ball, loc)',
-                        precond='Approaching(Ball, loc) & At(actor, loc)',
-                        effect='Returned(Ball)'),
-                 Action('Go(actor, to, loc)',
-                        precond='At(actor, loc)',
-                        effect='At(actor, to) & ~At(actor, loc)')])
-    """
-    
-    """
-    [Figure 11.10] DOUBLE-TENNIS-PROBLEM
-
-    A multiagent planning problem involving two partner tennis players
-    trying to return an approaching ball and repositioning around in the court.
-    """
-
-    # This works, but adds extra unnecessary detail to the problem
-    """
-    return PlanningProblem(
-        initial='At(A, LeftBaseLine) & At(B, RightNet) & Approaching(Ball, RightBaseLine) & Partner(A, B) & Partner(B, A)',
-        goals='Returned(Ball) & At(A, LeftNet) & At(B, RightNet)',
-        actions=[Action('Hit(actor, Ball, loc)',
-                        precond='Approaching(Ball, loc) & At(actor, loc)',
-                        effect='Returned(Ball)'),
-                 Action('Go(actor, to, loc)',
-                        precond='At(actor, loc)',
-                        effect='At(actor, to) & ~At(actor, loc)',
-                        domain='Player(actor) & CourtLoc(to) & CourtLoc(loc)')],
-        domain='Player(A) & Player(B) & Ball(Ball) & CourtLoc(LeftBaseLine) & CourtLoc(RightBaseLine) & CourtLoc(LeftNet) & CourtLoc(RightNet)'
-    )
-    """
-
-    # THIS DOESNT WORK BECAUSE LeftNet isn't recognized as an object
-    """
-    return PlanningProblem(
-        initial='At(A, LeftBaseLine) & At(B, RightNet) & Approaching(Ball, RightBaseLine) & Partner(A, B) & Partner(B, A)',
-        goals='Returned(Ball) & At(A, LeftNet) & At(B, RightNet)',
-        actions=[Action('Hit(actor, Ball, loc)',
-                        precond='Approaching(Ball, loc) & At(actor, loc)',
-                        effect='Returned(Ball)'),
-                 Action('Go(actor, to, loc)',
-                        precond='At(actor, loc)',
-                        effect='At(actor, to) & ~At(actor, loc)')])
-    """
-
-    """
-    return PlanningProblem(
-        initial='At(A, LeftBaseline) & At(B, RightNet) & Approaching(ball, RightBaseline)',
-        goals='At(A, LeftNet) & At(B, RightBaseline) & Returned(ball)',
-        actions=[Action('Hit(actor, ball, loc)',
-                        precond='Approaching(ball, loc) & At(actor, loc) & CourtLoc(loc) & IsBall(ball) & isActor(actor)',
-                        effect='Returned(ball)'),
-                 Action('Go(actor, to, loc)',
-                        precond='At(actor, loc) & CourtLoc(loc) & CourtLoc(to) & IsActor(actor)',
-                        effect='At(actor, to) & ~At(actor, loc)')],
-        domain='CourtLoc(LeftNet) & CourtLoc(RightNet) & CourtLoc(LeftBaseline) & CourtLoc(RightBaseline) & IsBall(ball) & IsActor(A) & IsActor(B)'
-    )
-    """
-
-    return PlanningProblem(
-        initial='At(A, LeftNet) & At(B, RightNet) & Approaching(ball, RightBaseline)',
-        goals='At(A, LeftBaseline) & At(B, LeftNet) & Returned(ball)',   
-        actions=[Action('Hit(actor, ball, loc)',
-                        precond='Approaching(ball, loc) & At(actor, loc)',
-                        effect='Returned(ball)'),
-                 Action('Go(actor, to, loc)',
-                        precond='At(actor, loc)',
-                        effect='At(actor, to) & ~At(actor, loc)')],
-        domain="Loc(LeftBaseline)")
-
-
 class ForwardPlan(search.Problem):
     """
     [Section 10.2.1]
@@ -955,75 +541,12 @@ class Level:
         "This function is only entered after state, actions at this level are computed."
         "Therefore, we're computing it for the current state and current (state+1) action layer"
         
-        #breakpoint()
         self.state_mutexes = self.mutex # save state mutexes
         self.mutex = [] # clear out effects from state mutex prior computation
 
-        """
-        # Technically correct version of Inconsistent effects, but introduces duplicates
-        ic1 = []
-        # Inconsistent effects - one action adds a literal that another deletes
-        pos_nsl, neg_nsl = self.separate(self.next_state_links)
-
-        for negeff in neg_nsl:
-            new_negeff = Expr(negeff.op[3:], *negeff.args)
-            for poseff in pos_nsl:
-                if new_negeff == poseff:
-                    for a in self.next_state_links[poseff]:
-                        for b in self.next_state_links[negeff]:
-                            #if {a, b} not in self.mutex:
-                            self.mutex.append({a, b})
-                            ic1.append({a,b})
-        """
-        
-        """
-        # Version that we have implemented efficiently below
-        ic2 = []
-        for a1, a2 in itertools.combinations(self.next_action_links.keys(), 2):
-            preconds_a1 = self.current_action_links.get(a1, [])
-            preconds_a2 = self.current_action_links.get(a2, [])
-            effects_a1 = self.next_action_links.get(a1, [])
-            effects_a2 = self.next_action_links.get(a2, []) 
-            
-            # Check for interference
-            interference = False 
-            # Check if effect of one action is negated by effect of other action (inconsistent effects?)
-            for e1 in effects_a1:
-                if e1.predicate_negate() in effects_a2:
-                    interference = True
-            for e2 in effects_a2:
-                if e2.predicate_negate() in effects_a1:
-                    interference = True
-
-            if interference:
-                mutex_pair = {a1, a2}
-                #if mutex_pair not in self.mutex:  # <-- avoid duplicates
-                ic2.append(mutex_pair)
-        """
-                    
-                               
-        # Competing needs - preconditions of two actions are mutex at previous proposition layer
-        """
-        pos_csl, neg_csl = self.separate(self.current_state_links)
-        # Why are we looking at syntactic components (negation)???
-        breakpoint()
-        for pos_precond in pos_csl:
-            for neg_precond in neg_csl:
-                new_neg_precond = Expr(neg_precond.op[3:], *neg_precond.args)
-                if new_neg_precond == pos_precond:
-                    for a in self.current_state_links[pos_precond]:
-                        for b in self.current_state_links[neg_precond]:
-                            if {a, b} not in self.mutex:
-                                self.mutex.append({a, b})
-        """
-         
-        # Competing Needs
-        # self.current_state_links = map from current state vars -> actions applicable
-        # self.current_action_links = map from current actions -> starting states
-        # Implement here: Iterate over all valid pairs of actions, and if the states they come from are mutex (use self.state_mutexes), add a mutex pair to self.mutex
         # Competing Needs - two actions are mutex if any of their preconditions are mutex at the previous state level
         for a1, a2 in itertools.combinations(self.current_action_links.keys(), 2):
-            preconds_a1 = self.current_action_links[a1] # states
+            preconds_a1 = self.current_action_links[a1]
             preconds_a2 = self.current_action_links[a2]
             
             if any({p, q} in self.state_mutexes for p in preconds_a1 for q in preconds_a2):
@@ -1031,67 +554,39 @@ class Level:
                 if mutex_pair not in self.mutex:
                     self.mutex.append(mutex_pair)
          
-        # Interference AND Inconsistent Effects
-        # Example - in shopping, ensure Move(Home,HW) and Move(Home,SM) are both mutex in L0
-        # current_action_links = current action -> current state map
+        # Interference AND Inconsistent Effects Mutex Calculation
         for a1, a2 in itertools.combinations(self.next_action_links.keys(), 2):
             preconds_a1 = self.current_action_links.get(a1, [])
             preconds_a2 = self.current_action_links.get(a2, [])
             effects_a1 = self.next_action_links.get(a1, [])
             effects_a2 = self.next_action_links.get(a2, [])
-            
-            #print(a1,a2)
-            #if "PickUp" in a1.op and "PickUp" in a2.op:
-            #    breakpoint()
-            
-            # Check for interference
+
             interference = False
-            # Check if a precondition of one action is negated by effect of another
+            # Interference Check
             for p1 in preconds_a1:
                 if p1.predicate_negate() in effects_a2:
                     interference = True
             for p2 in preconds_a2:
                 if p2.predicate_negate() in effects_a1:
                     interference = True
-            # Check if effect of one action is negated by effect of other action (inconsistent effects)
+
+            # Inconsistent Effects Check
             for e1 in effects_a1:
                 if e1.predicate_negate() in effects_a2:
-                    interference = True # Technically inconsistent effects
+                    interference = True
             for e2 in effects_a2:
                 if e2.predicate_negate() in effects_a1:
-                    interference = True # Technically inconsistent effects
+                    interference = True
 
             if interference:
                 mutex_pair = {a1, a2}
-                if mutex_pair not in self.mutex:  # <-- avoid duplicates
+                if mutex_pair not in self.mutex:
                     self.mutex.append(mutex_pair)
        
-       
-        #print([x for x in self.mutex if x not in savemutextemp])
-        #breakpoint() 
-        
-        # Inconsistent support - two props cannot be true given competing supporting actions
-        """
-        state_mutex = []
-        for pair in self.mutex:
-            next_state_0 = self.next_action_links[list(pair)[0]]
-            if len(pair) == 2:
-                next_state_1 = self.next_action_links[list(pair)[1]]
-            else:
-                next_state_1 = self.next_action_links[list(pair)[0]]
-            if (len(next_state_0) == 1) and (len(next_state_1) == 1):
-                state_mutex.append({next_state_0[0], next_state_1[0]})
-        """
-        
-        #breakpoint()
-    
-    
     def populate_prop_mutexes(self):
         "Compute the next level's proposition mutexes based on our current action mutexes"
+
         # Inconsistent support - two props cannot be true given competing supporting actions
-        
-        # self.next_action_links is a map from {next_state: [actions_leading_there]}
-        # We want to choose two next states, and add a mutex if all actions leading there are mutex
         state_mutex = []
         next_state_pairs = itertools.combinations(self.next_state_links.keys(), 2)
         for next_state_pair in list(next_state_pairs):
@@ -1104,7 +599,6 @@ class Level:
                 continue
             
             # if any two actions that lead to these states is not mutex, do not add a mutex to these states.
-            #if not any([{a1,a2} not in self.mutex and {a2,a1} not in self.mutex for a1 in acts_to_s1 for a2 in acts_to_s2]):
             if all([{a1,a2} in self.mutex or {a2,a1} in self.mutex for a1 in acts_to_s1 for a2 in acts_to_s2]):
                 mutex_pair = {s1, s2}
                 if mutex_pair not in state_mutex:
@@ -1120,7 +614,6 @@ class Level:
                         state_mutex.append(mutex_pair)
         
 
-        #breakpoint()
         self.next_state_mutexes = state_mutex
         return state_mutex
 
@@ -1129,15 +622,12 @@ class Level:
         to_remove = []
 
         # Normalize state mutex set for fast membership checks:
-        # state_mutex_lookup contains frozenset pairs like frozenset({p,q})
         state_mutex_lookup = set()
         for m in self.state_mutexes:
-            # m might already be a set/frozenset or possibly a tuple depending on other code
             state_mutex_lookup.add(frozenset(m))
 
         for action, preconds in list(self.current_action_links.items()):
             invalid = False
-            # print(action, preconds, list(itertools.combinations(preconds, 2)))
 
             for p1, p2 in itertools.combinations(preconds, 2):
                 if frozenset({p1, p2}) in state_mutex_lookup:
@@ -1220,36 +710,11 @@ class Level:
     def perform_actions(self):
         "Performs the necessary actions and returns a new Level"
 
-        # next_state_links.keys() will give the valid next states (the values would be all the actions that could cause it)
         new_kb = FolKB(list(set(self.next_state_links.keys())))
         new_level = Level(new_kb)
         
         return new_level
 
-    """
-    def deduplify(self):
-        "Remove duplicate entries from state, actions, links, and mutex lists."
-
-        # --- Deduplicate state ---
-        self.current_state = list(dict.fromkeys(self.current_state))  # order-preserving unique
-
-        # --- Deduplicate links (values are lists of actions/states) ---
-        for mapping in [self.current_action_links, self.current_state_links,
-                        self.next_action_links, self.next_state_links]:
-            for k, v in mapping.items():
-                mapping[k] = list(dict.fromkeys(v))  # unique but keep order
-
-        # --- Deduplicate mutex pairs ---
-        seen = set()
-        unique_mutex = []
-        for pair in self.mutex:
-            # turn {a,b} into a frozenset so it’s hashable
-            frozen = frozenset(pair)
-            if frozen not in seen:
-                seen.add(frozen)
-                unique_mutex.append(pair)
-        self.mutex = unique_mutex
-    """
 
 class Graph:
     """
@@ -1298,61 +763,6 @@ class Graph:
                 return False
         return True
     
-    """
-    def non_mutex_goals(self, goals, index):
-        "Checks whether the goals are mutually exclusive recursively"
-        
-        #print(f"mutexcheck: {index} == layer {len(self.levels) + index}")
-        num_layers = len(self.levels)
-        
-        # Base case: no goals left or first level reached
-        if num_layers + index == 0 or not goals:
-            return True
-
-        # Base case: level 0, assume persistence actions handle mutexes there
-        if num_layers + index == 0:
-            goal_perm = itertools.combinations(goals, 2)
-            for g in goal_perm:
-                if frozenset(g) in self.levels[0].mutex:
-                    return False
-            return True
-
-        # Check current level mutexes first
-        goal_perm = itertools.combinations(goals, 2)
-        for g in goal_perm:
-            if frozenset(g) in self.levels[index].mutex:
-                return False
-
-        # Recursive check: ensure for each pair of goals there exist non-mutex actions in previous level
-        level = self.levels[index - 1]
-
-        for goal in goals:
-            # if the goal has no supporting actions in previous level, can't satisfy
-            #breakpoint()
-            if goal not in level.next_state_links.keys():
-                return False
-
-        # Generate all combinations of supporting actions
-        supporting_actions_lists = [level.next_state_links[goal] for goal in goals]
-        for action_tuple in itertools.product(*supporting_actions_lists):
-            # Check if any pair of actions is mutex
-            mutex_found = False
-            for a1, a2 in itertools.combinations(action_tuple, 2):
-                if frozenset((a1, a2)) in level.mutex:
-                    mutex_found = True
-                    break
-            if not mutex_found:
-                # At least one combination works, recurse to previous level
-                prev_goals = []
-                for act in action_tuple:
-                    if act in level.current_action_links:
-                        prev_goals += level.current_action_links[act]
-                if self.non_mutex_goals(prev_goals, index - 1):
-                    return True
-
-        # If no combination works, goals are mutex
-        return False
-    """
 
 class GraphPlan:
     """
@@ -1392,10 +802,6 @@ class GraphPlan:
         if check:
             return True
 
-
-
-
-
     def _get_preconditions_for(self, action_set, level):
         #Collects all unique preconditions for a given set of actions in a level.
         all_preconditions = set()
@@ -1409,7 +815,6 @@ class GraphPlan:
     def _find_valid_action_sets(self, goals, level):
         #Finds sets of actions in the given level that are not mutually exclusive
         #and that collectively satisfy all the goals.
-        #This correctly replaces the flawed `itertools.product` logic.
         valid_sets = []
         
         # 1. Map each goal to the list of actions that can achieve it.
@@ -1448,16 +853,18 @@ class GraphPlan:
 
     # Place this within your GraphPlan class, replacing the previous version
     def _extract_solution_recursive(self, goals, level_index):
-        #Recursively searches for a plan backwards from a given level using negative indexing.
+        """
+        Recursively searches for a plan backwards from a given level using negative indexing.
         
-        #Args:
-        #    goals (set): The set of goal propositions to satisfy.
-        #    level_index (int): level index
-        #print(level_index)
-        #breakpoint()
-        # BASE CASE: We've recursed back to the initial proposition layer (Level 0).
-        # The 'goals' at this point are the preconditions for the very first set of actions.
-        # We must check if they exist in the initial state. No further recursion is needed.
+        Args:
+            goals (set): The set of goal propositions to satisfy.
+            level_index (int): level index
+        """
+        
+        #BASE CASE: We've recursed back to the initial proposition layer (Level 0).
+        #The 'goals' at this point are the preconditions for the very first set of actions.
+        #We must check if they exist in the initial state. No further recursion is needed.
+        
         if level_index == 0:
             #breakpoint()
             initial_state = set(self.graph.levels[0].current_state)
@@ -1496,179 +903,6 @@ class GraphPlan:
             self.no_goods.append(nogood_item)
         return None
 
-
-
-    """
-    def extract_solution(self, goals, index, root=False):
-        "Extracts the solution"
-        
-        #print(f"ENTER extract_solution with goals: {goals}")
-        
-        level = self.graph.levels[index]
-        if not self.graph.non_mutex_goals(goals, index):
-            self.no_goods.append((level, goals))
-            #print("  Level failed, goals are all mutex")
-            return
-
-        level = self.graph.levels[index - 1]
-        # print(f"  On level {index-1}")
-        # print(f"  Mutexes: \n    {level.mutex}")
-        
-        #breakpoint()
-
-        # Create all combinations of actions that satisfy each goal
-        # `actions` will be a list of tuples, where each tuple is every action that satisfies one goal literal
-        actions = []
-        for goal in goals:
-            actions.append(level.next_state_links[goal])
-
-        # `all_actions` selects elements from each list and creates a new list of actions that satisfies all goals at next level
-        #all_actions = list(itertools.product(*actions)) # Why only product? Why not all subsets of non-mutex preconditions?
-        #all_actions = powerset_product(actions)
-        print(len(list(itertools.product(*actions))))
-        all_actions = [*list(itertools.product(*actions)), *(powerset_product(actions)[0:5])]
-        print(len(all_actions))
-
-        # Filter out non-mutex actions
-        non_mutex_actions = []
-        for action_tuple in all_actions:
-            
-            #actlst = ["Move(R1, D2, D3)", "PIn(C2, D3)", "PHolding(R1)", "PIn(C3, R1)"]
-            #actlevellst = [str(repr(x)) for x in action_tuple]
-            #print(action_tuple)
-            #if all([x in actlevellst for x in actlst]):
-                #print("hi")
-                #breakpoint()
-                #print("hi")
-            
-            # Get all pairs of actions in our satisfactory action_tuple
-            action_pairs = itertools.combinations(list(set(action_tuple)), 2)
-            non_mutex_actions.append(list(set(action_tuple)))
-            #if not any(["PickUp" in str(x.op) for x in action_tuple]):
-                #breakpoint()
-            # acts = list(set(action_tuple))
-            # print(f"CHECKING tuple {acts} at level {index-1}")
-            for pair in list(action_pairs):
-                if set(pair) in level.mutex:
-                    #print(action_tuple, " failed due to ", pair, " in mutex")
-                    non_mutex_actions.pop(-1)
-                    break # If any actions are mutex, remove the entire tuple from our list
-                
-        if len(non_mutex_actions) == 0:
-            #print("  Level failed: No actions create state and are non-mutex")
-            return
-
-        # At this point, the non_mutex_actions contains a list of lists of valid actions that are all non-mutex and satisfy our goal state
-
-        # Recursion
-        
-        #print(f"Depth of solutions: {len(self.graph.levels)}, current level = {index-1}")
-        for action_list in non_mutex_actions:
-            if [action_list, index] not in self.solution:
-                self.solution.append([action_list, index])
-
-                new_goals = []
-                for act in set(action_list):
-                    if act in level.current_action_links:
-                        new_goals = new_goals + level.current_action_links[act]
-
-                if abs(index) + 1 == len(self.graph.levels):
-                    return
-                elif (level, new_goals) in self.no_goods:
-                    return
-                else:
-                    #print(f"  Goals achieved: {goals}, actions used: {action_list}")
-                    #breakpoint()
-                    self.extract_solution(new_goals, index - 1) # DFS search; goals at this layer, index of this layer
-            
-                
-        # Exit prior to computing solutions if we are not at the root.
-        if not root:
-            return
-        
-        #print(self.solution)
-        #breakpoint()
-
-        if not self.solution:
-            return []
-
-        # 1. Determine the leaf level and the expected length of a valid plan.
-        leaf_level = min(item[1] for item in self.solution)
-        expected_plan_length = abs(leaf_level)
-        #print(f"DEBUG: Target leaf level: {leaf_level}, Expected plan length: {expected_plan_length}\n")
-
-
-        completed_plans = []
-        current_path = []
-        initial_state = set(self.graph.levels[0].current_state)
-        level_zero = self.graph.levels[0]
-        
-        # 2. Iterate through the ordered search log.
-        for i, item in enumerate(self.solution):
-            actions, level = item
-            #print(f"--- Iteration {i+1} ---")
-            #print(f"Processing Item: {item}")
-            #print(f"Path BEFORE changes: {current_path}")
-
-            # 3. Rewind the current path to handle backtracks.
-            rewound = False
-            while current_path and level != current_path[-1][1] - 1:
-                if not rewound:
-                    #print(f"-> Backtrack detected (Lvl {level} doesn't follow Lvl {current_path[-1][1]}). Rewinding path...")
-                    rewound = True
-                popped_item = current_path.pop()
-                #print(f"   ...Popped: {popped_item}")
-            
-            #if rewound:
-            #    print(f"Path AFTER rewind: {current_path}")
-
-            # 4. Append the current item to form the new active path.
-            current_path.append(item)
-            #print(f"Appending item. New Path is now: {current_path}\n")
-            
-            # 5. Check if the active path has a valid structure.
-            is_rooted = current_path[0][1] == -1
-            is_grounded = current_path[-1][1] == leaf_level
-            is_contiguous = len(current_path) == expected_plan_length
-
-            # Only proceed if the path is structurally sound.
-            if is_rooted and is_grounded and is_contiguous:
-                #print(f"-> Path is structurally valid. Performing final initial state check...")
-                
-                # 6. Final Check: Verify first actions' preconditions against the initial state.
-                initial_state_satisfied = True
-                first_actions_in_plan = set(current_path[-1][0])
-
-                for action in first_actions_in_plan:
-                    preconditions = level_zero.current_action_links.get(action,None)
-                    if preconditions is None:
-                        initial_state_satisfied = False
-                        preconditions = []
-                    preconditions = set(preconditions)
-                    #print("Testing preconditions: ", preconditions, ", and initial_state: ", initial_state)
-                    if not preconditions.issubset(initial_state):
-                        initial_state_satisfied = False
-                        #print(f"   ❗️ Path failed validation. Action '{action}' preconditions {preconditions} are not met by initial state.")
-                        break
-                
-                if initial_state_satisfied:
-                    #print(f"   ✅ Success! Plan fully verified and found: {current_path}\n")
-                    completed_plans.append(list(current_path))
-        
-        print(f"--- Loop Finished ---")
-        print(f"Found {len(completed_plans)} complete plan(s) before final formatting.")
-
-        # Format the final solutions
-        solution = []
-        for plan in completed_plans:
-            action_plan = [item[0] for item in plan]
-            action_plan.reverse()
-            solution.append(action_plan)
-
-        return solution
-    """
-
-
     def goal_test(self, kb):
         goal_achieved = all(kb.ask(q) is not False for q in self.graph.planning_problem.goals)
         # print(goal_achieved)
@@ -1680,35 +914,20 @@ class GraphPlan:
         print("Forward Search")
         while True:
             self.graph.expand_graph()
-            #print(self.graph)
-            # print("Number of levels: ", len(self.graph.levels))
-            #breakpoint()
             if (self.goal_test(self.graph.levels[-1].kb) and self.graph.non_mutex_goals(
                     self.graph.planning_problem.goals, -1)):
 
-                print("SOLVED, EXTRACTING SOLUTION")
-                #print(self.graph)
-                #breakpoint()
-                # print(self.graph.non_mutex_goals(self.graph.planning_problem.goals, -1))
-                # self.graph.levels[-1](self.graph.planning_problem.actions, self.graph.objects)
-                # print(self.graph)
-                # print("Last level state:", self.graph.levels[-1].current_state)
-                # print("Last level mutexes:", self.graph.levels[-1].mutex)
-                # print("Next state links:", self.graph.levels[-1].next_state_links)
-                # print("Current action links:", self.graph.levels[-1].current_action_links)
-                # print(f"Extract Solution from {len(self.graph.levels)} levels, {len(self.graph.levels) - 1} actions")
-                #breakpoint()
                 solution = self.extract_solution(self.graph.planning_problem.goals, -1, True)
                 if solution:
-                    print(f"SOLUTION::::!!!!!!!!!!!!!!!!\n{solution}")
                     return [solution]
 
-            # print(len(self.graph.levels))
             if len(self.graph.levels) >= 2 and self.check_leveloff():
-                # breakpoint()
                 return None
 
 class Linearize:
+    """
+    
+    """
 
     def __init__(self, planning_problem):
         self.planning_problem = planning_problem
@@ -1751,9 +970,8 @@ class Linearize:
 
         graphPlan_solution = GraphPlan(self.planning_problem).execute()
         
-        for itr, possible_plan in enumerate(graphPlan_solution):
+        for possible_plan in graphPlan_solution:
             filtered_solution = self.filter(possible_plan)
-            #print(f"Trying filtered plan #{itr}: {filtered_solution}")
             
             ordered_solution = []
             # planning_problem will maintain the current state as we iterate over levels, allowing test application of actions
