+++ planning_envs.py
@@ -1,0 +1,348 @@
+from planning import *
+
+def air_cargo():
+    """
+    [Figure 10.1] AIR-CARGO-PROBLEM
+
+    An air-cargo shipment problem for delivering cargo to different locations,
+    given the starting location and airplanes.
+
+    Example:
+    >>> from planning import *
+    >>> ac = air_cargo()
+    >>> ac.goal_test()
+    False
+    >>> ac.act(expr('Load(C2, P2, JFK)'))
+    >>> ac.act(expr('Load(C1, P1, SFO)'))
+    >>> ac.act(expr('Fly(P1, SFO, JFK)'))
+    >>> ac.act(expr('Fly(P2, JFK, SFO)'))
+    >>> ac.act(expr('Unload(C2, P2, SFO)'))
+    >>> ac.goal_test()
+    False
+    >>> ac.act(expr('Unload(C1, P1, JFK)'))
+    >>> ac.goal_test()
+    True
+    >>>
+    """
+
+    return PlanningProblem(initial='At(C1, SFO) & At(C2, JFK) & At(P1, SFO) & At(P2, JFK)',
+                           goals='At(C1, JFK) & At(C2, SFO)',
+                           actions=[Action('Load(c, p, a)',
+                                           precond='At(c, a) & At(p, a)',
+                                           effect='In(c, p) & ~At(c, a)',
+                                           domain='Cargo(c) & Plane(p) & Airport(a)'),
+                                    Action('Unload(c, p, a)',
+                                           precond='In(c, p) & At(p, a)',
+                                           effect='At(c, a) & ~In(c, p)',
+                                           domain='Cargo(c) & Plane(p) & Airport(a)'),
+                                    Action('Fly(p, f, to)',
+                                           precond='At(p, f)',
+                                           effect='At(p, to) & ~At(p, f)',
+                                           domain='Plane(p) & Airport(f) & Airport(to)')],
+                           domain='Cargo(C1) & Cargo(C2) & Plane(P1) & Plane(P2) & Airport(SFO) & Airport(JFK)')
+
+
+def spare_tire():
+    """
+    [Figure 10.2] SPARE-TIRE-PROBLEM
+
+    A problem involving changing the flat tire of a car
+    with a spare tire from the trunk.
+
+    Example:
+    >>> from planning import *
+    >>> st = spare_tire()
+    >>> st.goal_test()
+    False
+    >>> st.act(expr('Remove(Spare, Trunk)'))
+    >>> st.act(expr('Remove(Flat, Axle)'))
+    >>> st.goal_test()
+    False
+    >>> st.act(expr('PutOn(Spare, Axle)'))
+    >>> st.goal_test()
+    True
+    >>>
+    """
+
+    return PlanningProblem(initial='At(Flat, Axle) & At(Spare, Trunk)',
+                           goals='At(Spare, Axle) & At(Flat, Ground)',
+                           actions=[Action('Remove(obj, loc)',
+                                           precond='At(obj, loc)',
+                                           effect='At(obj, Ground) & ~At(obj, loc)',
+                                           domain='Tire(obj)'),
+                                    Action('PutOn(t, Axle)',
+                                           precond='At(t, Ground) & ~At(Flat, Axle)',
+                                           effect='At(t, Axle) & ~At(t, Ground)',
+                                           domain='Tire(t)'),
+                                    Action('LeaveOvernight',
+                                           precond='',
+                                           effect='~At(Spare, Ground) & ~At(Spare, Axle) & ~At(Spare, Trunk) & \
+                                        ~At(Flat, Ground) & ~At(Flat, Axle) & ~At(Flat, Trunk)')],
+                           domain='Tire(Flat) & Tire(Spare)')
+
+
+def three_block_tower():
+    """
+    [Figure 10.3] THREE-BLOCK-TOWER
+
+    A blocks-world problem of stacking three blocks in a certain configuration,
+    also known as the Sussman Anomaly.
+
+    Example:
+    >>> from planning import *
+    >>> tbt = three_block_tower()
+    >>> tbt.goal_test()
+    False
+    >>> tbt.act(expr('MoveToTable(C, A)'))
+    >>> tbt.act(expr('Move(B, Table, C)'))
+    >>> tbt.goal_test()
+    False
+    >>> tbt.act(expr('Move(A, Table, B)'))
+    >>> tbt.goal_test()
+    True
+    >>>
+    """
+    return PlanningProblem(initial='On(A, Table) & On(B, Table) & On(C, A) & Clear(B) & Clear(C)',
+                           goals='On(A, B) & On(B, C)',
+                           actions=[Action('Move(b, x, y)',
+                                           precond='On(b, x) & Clear(b) & Clear(y)',
+                                           effect='On(b, y) & Clear(x) & ~On(b, x) & ~Clear(y)',
+                                           domain='Block(b) & Block(y)'),
+                                    Action('MoveToTable(b, x)',
+                                           precond='On(b, x) & Clear(b)',
+                                           effect='On(b, Table) & Clear(x) & ~On(b, x)',
+                                           domain='Block(b) & Block(x)')],
+                           domain='Block(A) & Block(B) & Block(C)')
+
+def logisticsPlanCustom(initial_state=None, goal_state=None):
+    if initial_state == None: 
+        initial_state = 'In(C1, R1) & In(C2, D1) & In(C3, D2) & In(R1, D1) & Holding(R1)'
+    if goal_state == None:
+        raise ValueError("Goal must be defined")
+
+    planning_problem = \
+    PlanningProblem(initial = initial_state,
+                    goals = goal_state,
+                    actions=[Action('PickUp(r, c, d)',
+                                    precond='In(r, d) & In (c, d) & ~Holding(r)',
+                                    effect='Holding(r) & ~In(c, d) & In(c, r)', 
+                                    domain='Robot(r) & Place(d) & Container(c)'),
+                             Action('PutDown(r, c, d)', 
+                                    precond='In(r, d) & In(c, r) & Holding(r)',
+                                    effect='~Holding(r) & ~In(c, r) & In(c, d)',
+                                    domain='Robot(r) & Place(d) & Container(c)'),
+                            Action('Move(r, d_start, d_end)',
+                                   precond='In(r,d_start)',
+                                   effect='~In(r, d_start) & In(r, d_end)',
+                                   domain='Robot(r) & Place(d_start) & Place(d_end)')],
+                domain='Container(C1) & Container(C2) & Container(C3) & Place(D1) & Place(D2) & Place(D3) & Robot(R1)')
+    
+    return planning_problem
+
+
+def simple_blocks_world():
+    """
+    SIMPLE-BLOCKS-WORLD
+
+    A simplified definition of the Sussman Anomaly problem.
+
+    Example:
+    >>> from planning import *
+    >>> sbw = simple_blocks_world()
+    >>> sbw.goal_test()
+    False
+    >>> sbw.act(expr('ToTable(A, B)'))
+    >>> sbw.act(expr('FromTable(B, A)'))
+    >>> sbw.goal_test()
+    False
+    >>> sbw.act(expr('FromTable(C, B)'))
+    >>> sbw.goal_test()
+    True
+    >>>
+    """
+
+    return PlanningProblem(initial='On(A, B) & Clear(A) & OnTable(B) & OnTable(C) & Clear(C)',
+                           goals='On(B, A) & On(C, B)',
+                           actions=[Action('ToTable(x, y)',
+                                           precond='On(x, y) & Clear(x)',
+                                           effect='~On(x, y) & Clear(y) & OnTable(x)'),
+                                    Action('FromTable(y, x)',
+                                           precond='OnTable(y) & Clear(y) & Clear(x)',
+                                           effect='~OnTable(y) & ~Clear(x) & On(y, x)')])
+
+
+def have_cake_and_eat_cake_too():
+    """
+    [Figure 10.7] CAKE-PROBLEM
+
+    A problem where we begin with a cake and want to
+    reach the state of having a cake and having eaten a cake.
+    The possible actions include baking a cake and eating a cake.
+
+    Example:
+    >>> from planning import *
+    >>> cp = have_cake_and_eat_cake_too()
+    >>> cp.goal_test()
+    False
+    >>> cp.act(expr('Eat(Cake)'))
+    >>> cp.goal_test()
+    False
+    >>> cp.act(expr('Bake(Cake)'))
+    >>> cp.goal_test()
+    True
+    >>>
+    """
+
+    return PlanningProblem(initial='Have(Cake) & ~Eaten(Cake)',
+                           goals='Have(Cake) & Eaten(Cake)',
+                           actions=[Action('Eat(Cake)',
+                                           precond='Have(Cake)',
+                                           effect='Eaten(Cake) & ~Have(Cake)'),
+                                    Action('Bake(Cake)',
+                                           precond='~Have(Cake)',
+                                           effect='Have(Cake)')])
+
+
+def shopping_problem():
+    """
+    SHOPPING-PROBLEM
+
+    A problem of acquiring some items given their availability at certain stores.
+
+    Example:
+    >>> from planning import *
+    >>> sp = shopping_problem()
+    >>> sp.goal_test()
+    False
+    >>> sp.act(expr('Go(Home, HW)'))
+    >>> sp.act(expr('Buy(Drill, HW)'))
+    >>> sp.act(expr('Go(HW, SM)'))
+    >>> sp.act(expr('Buy(Banana, SM)'))
+    >>> sp.goal_test()
+    False
+    >>> sp.act(expr('Buy(Milk, SM)'))
+    >>> sp.goal_test()
+    True
+    >>>
+    """
+
+    return PlanningProblem(initial='At(Home) & Sells(SM, Milk) & Sells(HW, Drill)',
+                           goals='Have(Milk) & Have(Drill)',
+                           actions=[Action('Buy(x, store)',
+                                           precond='At(store) & Sells(store, x)',
+                                           effect='Have(x)',
+                                           domain='Store(store) & Item(x)'),
+                                    Action('Go(x, y)',
+                                           precond='At(x)',
+                                           effect='At(y) & ~At(x)',
+                                           domain='Rplace(x) & Rplace(y)')],
+                           domain='Rplace(Home) & Rplace(SM) & Rplace(HW) & Store(SM) & Store(HW) & '
+                                  'Item(Milk) & Item(Drill)')
+
+
+def socks_and_shoes():
+    """
+    SOCKS-AND-SHOES-PROBLEM
+
+    A task of wearing socks and shoes on both feet
+
+    Example:
+    >>> from planning import *
+    >>> ss = socks_and_shoes()
+    >>> ss.goal_test()
+    False
+    >>> ss.act(expr('RightSock'))
+    >>> ss.act(expr('RightShoe'))
+    >>> ss.act(expr('LeftSock'))
+    >>> ss.goal_test()
+    False
+    >>> ss.act(expr('LeftShoe'))
+    >>> ss.goal_test()
+    True
+    >>>
+    """
+
+    return PlanningProblem(initial='',
+                           goals='RightShoeOn & LeftShoeOn',
+                           actions=[Action('RightShoe',
+                                           precond='RightSockOn',
+                                           effect='RightShoeOn'),
+                                    Action('RightSock',
+                                           precond='',
+                                           effect='RightSockOn'),
+                                    Action('LeftShoe',
+                                           precond='LeftSockOn',
+                                           effect='LeftShoeOn'),
+                                    Action('LeftSock',
+                                           precond='',
+                                           effect='LeftSockOn')])
+
+
+def double_tennis_problem_simple():
+    return PlanningProblem(
+        initial='At(A, LeftNet) & At(B, RightNet) & Approaching(ball, RightBaseline)',
+        goals='At(A, LeftBaseline) & Returned(ball)',
+        actions=[Action('Hit(actor, ball, loc)',
+                        precond='Approaching(ball, loc) & At(actor, loc)',
+                        effect='Returned(ball)'),
+                 Action('Go(actor, to, loc)',
+                        precond='At(actor, loc)',
+                        effect='At(actor, to) & ~At(actor, loc)')],
+        domain="Loc(LeftBaseline)")
+
+def double_tennis_problem_simple2():
+    return PlanningProblem(
+        initial='At(A, LeftNet) & At(B, LeftNet) & Approaching(ball, RightNet)',
+        goals='At(A, LeftNet) & Returned(ball) & At(B, LeftNet)',
+        actions=[Action('Hit(actor, ball, loc)',
+                        precond='Approaching(ball, loc) & At(actor, loc)',
+                        effect='Returned(ball)'),
+                 Action('Go(actor, to, loc)',
+                        precond='At(actor, loc)',
+                        effect='At(actor, to) & ~At(actor, loc)')])
+
+def double_tennis_problem_simple3():
+    return PlanningProblem(
+        initial='At(A, LeftNet) & Approaching(ball, RightNet)',
+        goals='At(A, LeftNet) & Returned(ball)',
+        actions=[Action('Hit(actor, ball, loc)',
+                        precond='Approaching(ball, loc) & At(actor, loc)',
+                        effect='Returned(ball)'),
+                 Action('Go(actor, to, loc)',
+                        precond='At(actor, loc)',
+                        effect='At(actor, to) & ~At(actor, loc)')])
+
+
+def double_tennis_problem():
+    """
+    [Figure 11.10] DOUBLE-TENNIS-PROBLEM
+
+    A multiagent planning problem involving two partner tennis players
+    trying to return an approaching ball and repositioning around in the court.
+
+    Example:
+    >>> from planning import *
+    >>> dtp = double_tennis_problem()
+    >>> goal_test(dtp.goals, dtp.initial)
+    False
+    >>> dtp.act(expr('Go(A, RightBaseLine, LeftBaseLine)'))
+    >>> dtp.act(expr('Hit(A, Ball, RightBaseLine)'))
+    >>> goal_test(dtp.goals, dtp.initial)
+    False
+    >>> dtp.act(expr('Go(A, LeftNet, RightBaseLine)'))
+    >>> goal_test(dtp.goals, dtp.initial)
+    True
+    >>>
+    """
+
+    return PlanningProblem(
+        initial='At(A, LeftNet) & At(B, RightNet) & Approaching(ball, RightBaseline)',
+        goals='At(A, LeftBaseline) & At(B, LeftNet) & Returned(ball)',   
+        actions=[Action('Hit(actor, ball, loc)',
+                        precond='Approaching(ball, loc) & At(actor, loc)',
+                        effect='Returned(ball)'),
+                 Action('Go(actor, to, loc)',
+                        precond='At(actor, loc)',
+                        effect='At(actor, to) & ~At(actor, loc)')],
+        domain="Loc(LeftBaseline)")
+
